Binomial random walk model

This is a state-space model defined by a Binomial measurement error and a latent random walk.

default_prior_BinRW()

fit_BinRW(train, test = NULL, max_score, prior = default_prior_BinRW(), ...)

sample_prior_BinRW(
  N_patient = 1,
  t_max = c(2),
  max_score,
  prior = default_prior_BinRW(),
  ...
)

Arguments

train
Training dataframe (details of the format in prepare_data_lgtd())

test
Testing dataframe (details of the format in prepare_data_lgtd())

max_score
Maximum value that the score can take

prior
A named list with elements sigma, mu_logit_y0 and sigma_logit_y0 specifying priors for the corresponding parameters. Each element of the list should be a vector of length 2, containing values for x1 and x2, x2 > 0, such as:

sigma ~ normal(x1, x2).

mu_logit_y0 ~ normal(x1, x2)

sigma_logit_y0 ~ normal(x1, x2)

NB: For sigma and sigma_logit_y0, usually x1=0 to define a half-normal distribution since the parameters are constrained to be positive.

...
Arguments to be pass to rstan::sampling()

N_patient
Number of patients

t_max
Vector of size N_patient indicating the time-series length

Value

default_prior_BinRW: named list defining the default priors

fit_BinRW: Stanfit object

sample_prior_BinRW: Stanfit object

Details

default_prior_BinRW returns the default prior

fit_BinRW fits the Binomial random walk model

sample_prior_BinRW samples the prior predictive distribution

The default prior for sigma translates to an odd ratio increment of at most 5 (~ 2 * upper bound of prior).

The default priors for mu_logit_y0 and sigma_logit_y0 translates to an approximately uniform prior on y0.

For more details see the vignette.

Model's parameters

Population parameters:

sigma: Standard deviation of the random walk (in logit scale)

mu_logit_y0: Population mean of the initial condition (in logit scale)

sigma_logit_y0: Population standard deviation of the initial condition (logit scale)

Patient-dependent parameters:

logit_y0: Logit of the initial condition

Observation-dependent (patient- and time-dependent) parameters:

y_lat: Latent score (probability)

logit_lat: logit of y_lat

See list_parameters(model = "BinRW").