Performance of (population) historical forecast

Performance of (population) historical forecast

Usage

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd        RPS
#> 1      59 -5.303305 0.19393183
#> 2      46 -5.303305 0.12602139
#> 3      52 -5.303305 0.15527512
#> 4      55 -5.303305 0.17124527
#> 5      60 -5.303305 0.19985223
#> 6      46 -5.303305 0.12602139
#> 7      57 -5.303305 0.18238954
#> 8      53 -5.303305 0.16049900
#> 9      51 -5.303305 0.15015074
#> 10     51 -5.303305 0.15015074
#> 11     46 -5.303305 0.12602139
#> 12     55 -5.303305 0.17124527
#> 13     57 -5.303305 0.18238954
#> 14     53 -5.303305 0.16049900
#> 15     46 -5.303305 0.12602139
#> 16     48 -5.303305 0.13537462
#> 17     48 -5.303305 0.13537462
#> 18     51 -5.303305 0.15015074
#> 19     53 -5.303305 0.16049900
#> 20     45 -5.303305 0.12149402
#> 21     58 -5.303305 0.18811094
#> 22     56 -5.303305 0.17676765
#> 23     49 -5.303305 0.14020049
#> 24     48 -5.303305 0.13537462
#> 25     54 -5.303305 0.16582238
#> 26     39 -5.303305 0.09641940
#> 27     46 -5.303305 0.12602139
#> 28     57 -5.303305 0.18238954
#> 29     50 -5.303305 0.14512586
#> 30     55 -5.303305 0.17124527
#> 31     54 -5.303305 0.16582238
#> 32     47 -5.303305 0.13064825
#> 33     50 -5.303305 0.14512586
#> 34     55 -5.303305 0.17124527
#> 35     52 -5.303305 0.15527512
#> 36     57 -5.303305 0.18238954
#> 37     51 -5.303305 0.15015074
#> 38     52 -5.303305 0.15527512
#> 39     53 -5.303305 0.16049900
#> 40     51 -5.303305 0.15015074
#> 41     49 -5.303305 0.14020049
#> 42     57 -5.303305 0.18238954
#> 43     48 -5.303305 0.13537462
#> 44     55 -5.303305 0.17124527
#> 45     37 -5.303305 0.08885721
#> 46     43 -5.303305 0.11273780
#> 47     41 -5.303305 0.10437959
#> 48     50 -5.303305 0.14512586
#> 49     44 -5.303305 0.11706616
#> 50     49 -5.303305 0.14020049
#> 51     48 -5.303305 0.13537462
#> 52     49 -5.303305 0.14020049
#> 53     49 -5.303305 0.14020049
#> 54     51 -5.303305 0.15015074
#> 55     50 -5.303305 0.14512586
#> 56     55 -5.303305 0.17124527
#> 57     47 -5.303305 0.13064825
#> 58     49 -5.303305 0.14020049
#> 59     44 -5.303305 0.11706616
#> 60     64 -5.303305 0.22452885
#> 61     43 -5.303305 0.11273780
#> 62     52 -5.303305 0.15527512
#> 63     42 -5.303305 0.10850895
#> 64     54 -5.303305 0.16582238
#> 65     49 -5.303305 0.14020049
#> 66     45 -5.303305 0.12149402
#> 67     50 -5.303305 0.14512586
#> 68     55 -5.303305 0.17124527
#> 69     51 -5.303305 0.15015074
#> 70     53 -5.303305 0.16049900
#> 71     50 -5.303305 0.14512586
#> 72     46 -5.303305 0.12602139
#> 73     47 -5.303305 0.13064825
#> 74     51 -5.303305 0.15015074
#> 75     46 -5.303305 0.12602139
#> 76     44 -5.303305 0.11706616
#> 77     53 -5.303305 0.16049900
#> 78     43 -5.303305 0.11273780
#> 79     50 -5.303305 0.14512586
#> 80     43 -5.303305 0.11273780
#> 81     46 -5.303305 0.12602139
#> 82     58 -5.303305 0.18811094
#> 83     48 -5.303305 0.13537462
#> 84     56 -5.303305 0.17676765
#> 85     49 -5.303305 0.14020049
#> 86     44 -5.303305 0.11706616
#> 87     45 -5.303305 0.12149402
#> 88     53 -5.303305 0.16049900
#> 89     50 -5.303305 0.14512586
#> 90     50 -5.303305 0.14512586
#> 91     45 -5.303305 0.12149402
#> 92     46 -5.303305 0.12602139
#> 93     52 -5.303305 0.15527512
#> 94     47 -5.303305 0.13064825
#> 95     41 -5.303305 0.10437959
#> 96     46 -5.303305 0.12602139
#> 97     56 -5.303305 0.17676765
#> 98     53 -5.303305 0.16049900
#> 99     55 -5.303305 0.17124527
#> 100    51 -5.303305 0.15015074