Performance of a uniform forecast

Performance of a uniform forecast

Usage

add_uniform_pred(
  test,
  max_score,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the samples 0:max_score are returned.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Examples

max_score <- 100
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_uniform_pred(test, max_score)
#>     Score       lpd        RPS
#> 1      39 -4.615121 0.09613861
#> 2      50 -4.615121 0.08415842
#> 3      42 -4.615121 0.09049505
#> 4      46 -4.615121 0.08574257
#> 5      58 -4.615121 0.09049505
#> 6      51 -4.615121 0.08425743
#> 7      55 -4.615121 0.08663366
#> 8      48 -4.615121 0.08455446
#> 9      47 -4.615121 0.08504950
#> 10     51 -4.615121 0.08425743
#> 11     42 -4.615121 0.09049505
#> 12     47 -4.615121 0.08504950
#> 13     57 -4.615121 0.08900990
#> 14     46 -4.615121 0.08574257
#> 15     52 -4.615121 0.08455446
#> 16     54 -4.615121 0.08574257
#> 17     44 -4.615121 0.08772277
#> 18     39 -4.615121 0.09613861
#> 19     58 -4.615121 0.09049505
#> 20     51 -4.615121 0.08425743
#> 21     41 -4.615121 0.09217822
#> 22     58 -4.615121 0.09049505
#> 23     46 -4.615121 0.08574257
#> 24     51 -4.615121 0.08425743
#> 25     43 -4.615121 0.08900990
#> 26     54 -4.615121 0.08574257
#> 27     48 -4.615121 0.08455446
#> 28     51 -4.615121 0.08425743
#> 29     50 -4.615121 0.08415842
#> 30     56 -4.615121 0.08772277
#> 31     56 -4.615121 0.08772277
#> 32     44 -4.615121 0.08772277
#> 33     45 -4.615121 0.08663366
#> 34     47 -4.615121 0.08504950
#> 35     46 -4.615121 0.08574257
#> 36     43 -4.615121 0.08900990
#> 37     40 -4.615121 0.09405941
#> 38     48 -4.615121 0.08455446
#> 39     45 -4.615121 0.08663366
#> 40     43 -4.615121 0.08900990
#> 41     54 -4.615121 0.08574257
#> 42     47 -4.615121 0.08504950
#> 43     54 -4.615121 0.08574257
#> 44     60 -4.615121 0.09405941
#> 45     47 -4.615121 0.08504950
#> 46     52 -4.615121 0.08455446
#> 47     45 -4.615121 0.08663366
#> 48     52 -4.615121 0.08455446
#> 49     49 -4.615121 0.08425743
#> 50     56 -4.615121 0.08772277
#> 51     51 -4.615121 0.08425743
#> 52     48 -4.615121 0.08455446
#> 53     51 -4.615121 0.08425743
#> 54     42 -4.615121 0.09049505
#> 55     56 -4.615121 0.08772277
#> 56     50 -4.615121 0.08415842
#> 57     49 -4.615121 0.08425743
#> 58     52 -4.615121 0.08455446
#> 59     55 -4.615121 0.08663366
#> 60     49 -4.615121 0.08425743
#> 61     35 -4.615121 0.10643564
#> 62     61 -4.615121 0.09613861
#> 63     51 -4.615121 0.08425743
#> 64     56 -4.615121 0.08772277
#> 65     48 -4.615121 0.08455446
#> 66     47 -4.615121 0.08504950
#> 67     54 -4.615121 0.08574257
#> 68     54 -4.615121 0.08574257
#> 69     55 -4.615121 0.08663366
#> 70     53 -4.615121 0.08504950
#> 71     51 -4.615121 0.08425743
#> 72     48 -4.615121 0.08455446
#> 73     48 -4.615121 0.08455446
#> 74     47 -4.615121 0.08504950
#> 75     53 -4.615121 0.08504950
#> 76     51 -4.615121 0.08425743
#> 77     47 -4.615121 0.08504950
#> 78     50 -4.615121 0.08415842
#> 79     52 -4.615121 0.08455446
#> 80     58 -4.615121 0.09049505
#> 81     44 -4.615121 0.08772277
#> 82     53 -4.615121 0.08504950
#> 83     62 -4.615121 0.09841584
#> 84     50 -4.615121 0.08415842
#> 85     53 -4.615121 0.08504950
#> 86     47 -4.615121 0.08504950
#> 87     51 -4.615121 0.08425743
#> 88     51 -4.615121 0.08425743
#> 89     44 -4.615121 0.08772277
#> 90     51 -4.615121 0.08425743
#> 91     59 -4.615121 0.09217822
#> 92     57 -4.615121 0.08900990
#> 93     55 -4.615121 0.08663366
#> 94     49 -4.615121 0.08425743
#> 95     57 -4.615121 0.08900990
#> 96     46 -4.615121 0.08574257
#> 97     51 -4.615121 0.08425743
#> 98     54 -4.615121 0.08574257
#> 99     46 -4.615121 0.08574257
#> 100    56 -4.615121 0.08772277