Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1] -0.07771195 -0.76391518  0.34005430  0.97678676 -1.61963089 -0.23206527
#>   [7] -0.58364885 -1.11116548  1.63662505  0.78911104 -1.41244830  1.43778756
#>  [13]  0.74133008  1.45455193 -0.28020072 -0.39781330 -0.20440178 -1.13104447
#>  [19]  0.13528393 -0.12412952 -1.09383193 -0.72041591  0.04755686  0.47798473
#>  [25] -1.49784032  0.99670743 -0.57681987  0.89686081  1.26505140 -1.38600645
#>  [31] -0.99667359  0.85011856 -0.23563359  0.65827455  1.18363475  0.86268154
#>  [37]  0.15005188 -1.27869932 -0.45104675 -0.80577908  0.60710189 -0.66343512
#>  [43]  0.51060110  1.23448633  0.76010288  0.67739122  1.76920697  0.70385886
#>  [49] -0.45281858  0.57180586  0.30518079  0.06970389  0.20629782 -0.31473681
#>  [55] -0.70521443  0.99377906  0.75050689 -0.77859343 -0.18354953  1.59832245
#>  [61]  0.08227113 -0.89442495  0.03975008  0.88842088  1.48482367 -1.20702633
#>  [67] -1.28886896  1.27380971  1.23221044  1.04201713 -0.10628245 -0.52551681
#>  [73] -0.37189353  0.24421332  1.37983733  0.03869373  1.86342850  0.26748177
#>  [79] -0.18019881 -1.34525010 -1.62544629  0.14409098  0.23118327  0.37089994
#>  [85]  0.17873516  0.97496853 -1.17823176 -1.14629576  0.59605462  1.16816417
#>  [91] -0.20017839  0.09677364  0.15007400 -2.03564352  0.63460791  0.45303274
#>  [97]  1.57894069 -0.35918211 -0.27299454 -1.51835358
#> 
#> [[2]]
#>   [1]  1.465486717  1.804455107 -0.128425532 -0.291428806 -0.176627071
#>   [6] -2.157603450  1.407760643  0.056775976 -1.194103760  0.393168652
#>  [11] -0.487393668  0.474498822  0.603879901 -0.244634607 -0.098974223
#>  [16]  2.213089679 -0.871651227  1.815473330  0.229119923  0.247101733
#>  [21] -0.480965716 -0.862425884  0.498867203  0.374205381 -2.526458086
#>  [26]  0.784618882  0.313929120 -0.649795568 -0.010562492  0.385727678
#>  [31] -0.056673174 -1.238343229  0.267631635 -1.607683896 -0.938116837
#>  [36]  1.488308169  0.537787265 -0.761571229 -0.244678891  0.681758174
#>  [41]  0.407604989  0.642815918  0.877654023  0.467037898  1.715596446
#>  [46]  0.156670116 -0.927662449  0.570366501  1.151399752  0.467892799
#>  [51]  1.424486029  1.502088048 -0.245311596  0.547977519 -1.994747638
#>  [56] -0.831417077 -0.592399538  1.922223991 -0.357177667  0.683861642
#>  [61] -0.916699533 -0.138472786  0.951024759 -1.766055458  1.632562772
#>  [66] -1.014201290  0.071734545  0.438450143  0.004515792  0.330621051
#>  [71] -0.284517704  0.281407357 -0.016972266  0.037255614  2.963705643
#>  [76]  1.565375765 -0.289642286  0.470614066 -1.165379303  0.377658961
#>  [81]  2.589986635  0.851055564  0.183903923  0.141547005 -2.669518244
#>  [86]  1.344243325 -1.374275627  1.428768507  0.076969042  1.190870406
#>  [91]  1.429738982 -0.015299603 -1.840425467  0.632137081 -1.088619956
#>  [96] -1.176446237  0.427342178  2.097764351  1.102628927  0.850436907
#> 
#> [[3]]
#>   [1]  0.26416988  0.17120641 -0.07113646 -0.69454870 -2.55321944 -0.38787286
#>   [7]  1.18848870  0.47680615 -0.75873979 -0.58244884  1.92790250  0.15496231
#>  [13] -1.16080940  1.31961807 -0.89466077 -1.92658979 -1.31188200  1.01706257
#>  [19]  0.06735534  0.44979269  0.13057877  0.42042934 -1.40214142  0.28890649
#>  [25]  0.16424138 -1.61865988  0.89368436 -0.19779756  0.25017955  0.26803806
#>  [31] -0.83106616 -0.39045714 -1.27992034 -0.85974997  1.14530257 -1.10775172
#>  [37] -0.74674655 -1.21416232 -1.21007735 -0.14017590 -1.83295854  1.46832688
#>  [43]  0.13678665  1.75512393  1.24878065  1.47722139  2.19252275  1.25694755
#>  [49] -0.90166349 -1.21561917 -0.93514711 -1.23476993  1.24424633 -1.87203296
#>  [55] -0.19295043  0.74015190 -0.39728789 -0.28432418 -0.05914413 -0.63928485
#>  [61]  0.02212267  1.07348970  0.06966478  3.01068280  0.81100578  0.74872888
#>  [67]  0.18907878  1.59686805 -0.72542208 -0.45272910 -0.59570952  0.13979960
#>  [73]  1.98227031 -0.03890318 -0.15787552  0.49102366 -0.52697635  0.98533165
#>  [79]  1.08187326 -1.32707850 -0.33687029 -2.90389427  1.36892306  0.24297992
#>  [85] -0.61087320 -1.02113542 -0.44539718 -0.03799768  1.71821177  0.18292028
#>  [91]  0.11916838  1.65445396  0.93782674 -1.06755348 -1.09123033 -0.62367807
#>  [97]  0.95505546  1.15602207  1.04908079 -0.42772702
#> 
#> [[4]]
#>   [1] -0.64286662  0.19437392  0.72198235 -0.45006400  1.00976665 -0.62999232
#>   [7] -0.51708630  0.60509291 -1.01482989 -0.89330499 -0.27681346 -1.77244124
#>  [13]  0.19174436  0.44537465  1.25281200 -0.37673766 -0.40500921 -0.72293251
#>  [19]  0.02712111  1.80411606 -0.27909835  1.49426426 -0.07201569  0.41297117
#>  [25]  0.08296232  0.95424952  0.54825043  0.88317886 -0.55446598  0.55555106
#>  [31]  0.84499351  1.32793673 -0.52371146 -0.22638034 -0.18558445  0.50731087
#>  [37]  0.11552558  0.96435298 -0.72644154 -0.24846146  0.22344038 -1.40647198
#>  [43]  0.25781455  0.71190793 -0.12453429  0.63361421 -0.09114892 -1.25051631
#>  [49]  1.84380639 -0.08792330 -0.26054770 -0.19231419 -0.13850549 -0.28986764
#>  [55] -0.96712578 -0.43654597  1.71103433  0.57870700  0.28752669 -0.66627861
#>  [61] -1.18910451  0.74195961  1.54444733 -0.25226664 -1.85247879  0.82310514
#>  [67]  0.62947535 -0.02848251  1.59916489  1.35709268 -0.56404494 -1.94722330
#>  [73]  2.20870830 -0.63212107  0.44200316 -0.04318006  1.14995177 -0.33665133
#>  [79] -0.15334417 -1.05638195  1.10305371 -1.53988884 -0.11758489 -1.14854718
#>  [85] -0.33787472  1.18277492 -2.99265066  1.60521165  0.51818446  0.50914258
#>  [91] -0.23170723 -1.13961833 -1.22240236  0.09898116 -0.27868243  0.63040707
#>  [97]  0.75189413  0.42096348  1.69176119 -0.24612837
#> 
#> [[5]]
#>   [1]  0.62942719 -0.92153347  0.52272293 -1.89695543  0.29075374  0.23416416
#>   [7] -0.25252726  1.05092004 -0.63899418  1.12117926  1.11819730 -0.95878190
#>  [13] -0.70526558  1.68787766 -0.01546906  0.84920739 -2.08536674 -0.47968609
#>  [19] -1.19983594 -2.16128401 -1.64243232  2.11519194  0.97409594  1.68033596
#>  [25]  0.90130096  1.41428768 -1.68385488 -0.38386697 -0.14283355 -0.84260898
#>  [31]  0.26081864 -1.74975772 -0.41628984 -0.36526779  0.93659529 -0.94983094
#>  [37] -0.38748645  0.06922384 -0.91909117 -0.58397687 -0.48957878  0.43199267
#>  [43]  0.90802065 -0.63982401 -0.38515944  1.51235940  0.80737065 -0.88343294
#>  [49] -3.48377493  2.82641464 -0.22831038 -1.09270559 -0.56965539  0.20813797
#>  [55]  0.53739758  0.67256945  1.46107389  0.27490012 -1.24408001 -1.64067531
#>  [61]  1.52977588 -1.10895794 -0.25507906 -0.95815568 -0.33175594  0.40613081
#>  [67] -0.54752823 -1.06002900  0.10726144 -0.62758667  0.98378643  1.80658602
#>  [73] -1.31480889  0.01559522  1.53573415 -0.23530112  0.35985572  0.93375174
#>  [79] -0.75069753  0.50826909 -0.63059904 -0.71529593  0.71984392 -0.54930960
#>  [85]  1.17159002 -1.19981879  1.51112462  1.12725388 -1.20791370  0.18754400
#>  [91]  0.20547090  0.56240175  1.75178746  1.63691314 -1.32445633  0.76114690
#>  [97] -0.16919396 -0.15602703 -0.00221299 -0.58157271
#> 
#> [[6]]
#>   [1] -0.155193019  0.311085741  0.011646685 -0.147835079  0.168926187
#>   [6] -0.278977480 -1.496436674  1.326191646 -0.655123436 -0.748721061
#>  [11]  0.983987759  0.411845767 -1.286800989 -3.087209224  0.659914395
#>  [16]  0.129997178  0.144003762 -0.978360370  0.100579463  0.298773979
#>  [21]  0.936129523  0.417433017  0.254318942 -1.222047859  0.420283026
#>  [26] -2.037247160  0.291286293  0.847824402 -0.463807395  0.918789137
#>  [31]  0.549601451  0.248927855  0.957050202 -0.455508275  1.239391766
#>  [36] -0.105174427  0.089992384 -0.630469784  0.327952459  0.957316718
#>  [41]  1.105459785 -1.989168682 -1.789855561 -1.492910800 -0.708499855
#>  [46]  1.102667821 -0.183358312 -1.984847109 -1.579745306 -2.031250850
#>  [51]  0.326528832  0.037931610 -1.204304117 -0.925787327 -0.284949003
#>  [56]  0.411031625 -1.162908656  0.340094873 -0.556249210 -0.567540634
#>  [61] -1.754486261  0.393314420  1.336920532  0.901382214 -0.069676231
#>  [66]  0.981465134  0.467746693  0.363329409  0.362648590 -0.076566152
#>  [71]  0.179495104 -0.440047179  1.845910620 -0.244546858  0.799276224
#>  [76]  0.065812655  1.370071721  0.008900061  0.364798386 -1.380011785
#>  [81] -0.195474390 -0.645412560  0.405443220  0.155097535  0.991609261
#>  [86]  0.143563147 -0.447262943 -0.600271292 -0.746377206  1.135008072
#>  [91] -1.212050723 -0.079120014  0.049989669 -1.914686731 -1.266684608
#>  [96] -0.379449335  1.527345892  0.107087638  0.023313145  0.423903317
#> 
#> [[7]]
#>   [1]  0.124858041 -0.110886848  1.075863614  0.870780513 -0.204237309
#>   [6]  1.125638178  2.585871324 -0.107897032  1.046308148  1.489901405
#>  [11] -1.898865083  2.355806948 -2.695454532 -1.357033967 -1.073552316
#>  [16]  0.009417637  0.458431763 -0.178530726 -1.359119570  0.687124794
#>  [21]  1.370455642  0.518415452  0.850553954  2.035799506  0.157506712
#>  [26] -1.641791211  1.039771702 -0.460280388 -0.558243706 -1.559313223
#>  [31] -0.983221779 -0.461118885 -1.722370714  1.043358233 -0.440264959
#>  [36]  1.197312837  1.087698083  0.093864772  0.166584385  0.353673851
#>  [41] -0.315133466  0.878714226  1.098494075 -0.217185053  0.065340332
#>  [46] -1.513940106 -1.990547865  0.850860477 -1.355448258 -0.596019541
#>  [51] -0.249854293 -0.958910758  1.625394506 -0.382500253  0.481796210
#>  [56]  0.556518857 -0.227139677  0.867979312 -0.676877024  0.612688274
#>  [61] -1.130167650  1.050255972  0.859599576  0.723543022 -0.831008836
#>  [66]  0.131863418  1.135322151 -0.316340092  1.134910911 -0.746699974
#>  [71] -1.275094290 -0.567743392 -0.945481991  0.636833081  0.696787082
#>  [76] -0.963679285  1.091988905  0.591733692 -1.117388738 -0.105323509
#>  [81]  0.309086678 -0.092661870 -0.658306772 -0.103741166 -0.215739564
#>  [86]  1.472463243 -0.490135831 -0.519050243  0.227823981 -0.095625659
#>  [91] -0.313001560 -0.579468102 -2.049157778 -0.269657125  0.682201736
#>  [96]  1.748546061 -0.008891330 -0.554824260  1.276640757 -0.401546949
#> 
#> [[8]]
#>   [1] -0.049985110 -1.537438714 -0.638904187 -1.688630848 -1.301652661
#>   [6]  0.106538883 -0.289386810 -0.829283673  0.421866302 -0.640922358
#>  [11] -1.385140338 -0.138538545 -0.298465574 -1.532173495  2.344399545
#>  [16] -0.035951999  1.479813910  1.140888007  2.723266661  0.107037518
#>  [21]  1.005283730 -1.180462045  0.680563509  0.207114479 -2.505983001
#>  [26] -0.149702326  0.358709360  1.090556007  0.192961482  0.481719649
#>  [31]  0.471385767  0.436634831 -0.552095607  0.768020662 -2.983636701
#>  [36]  0.757292422 -0.520452848  1.569804665  0.247060895  0.389986438
#>  [41] -1.459270538 -0.580634500  1.935218901 -1.099958427 -0.258820653
#>  [46]  0.902045388 -1.747858450 -2.136912801  0.158100000  0.321681632
#>  [51] -1.807792593 -1.432194013  0.556319227  1.035531417  1.752230887
#>  [56]  0.207436092  0.900347815 -0.719268121 -0.281192527  0.789933540
#>  [61]  0.476682417  0.308289379 -0.362788161 -0.628512826  0.575531648
#>  [66]  0.217722350 -1.317204415  0.249847431 -0.179047097 -1.794195829
#>  [71] -0.661622545  0.476273111  3.063490495  0.706323947  0.198094370
#>  [76] -2.382971079 -0.519084715 -0.875957641  0.506001362 -0.475795214
#>  [81] -1.087779439  0.739068235 -0.612480225  0.940191905 -0.325508030
#>  [86]  0.539754927  1.618829123  0.051606538 -0.812341540 -1.449317418
#>  [91] -1.349483296  0.668938761  0.002733768  1.396144658 -0.435649891
#>  [96] -0.133409552 -0.408536782  0.282149008 -2.112795833 -0.504107351
#> 
#> [[9]]
#>   [1] -0.57427456 -0.68254023  1.35029431 -0.90938944 -1.55562346  0.66444021
#>   [7] -1.06951326 -0.60434941 -0.16594150  1.32748135  0.06638126 -1.36938331
#>  [13]  1.26947536 -0.23083877  1.42853643 -0.81356062  1.21535188 -0.21558691
#>  [19] -1.20529677  2.64852445  0.25483099  0.99911365  0.60314904  0.23872141
#>  [25] -0.44930781 -0.51188267 -0.06779377 -0.72199922 -0.63576492  0.13721887
#>  [31] -0.44162493  0.04294371 -1.53196645 -2.43280511  0.12102748 -0.59402621
#>  [37]  1.03840259 -0.25681161  0.15415670  0.31824838  1.50005795  1.00747809
#>  [43]  1.48029291 -0.62240969 -1.82636440 -0.64036838 -0.47162462  1.76696756
#>  [49] -0.58345182 -0.08206377  2.05867576 -0.01112287 -0.41470246 -0.22608091
#>  [55]  0.13930422  1.11316896 -0.64174809  0.44117156  0.04959896 -2.00446369
#>  [61]  0.57733081  0.54598741 -0.01926320 -0.16378906 -0.03639418  0.17257776
#>  [67] -1.60806192  0.76195592  0.47735051 -2.65703366 -0.58065109 -1.22142648
#>  [73] -0.63750989 -0.54374609 -0.51010901  0.95946579  2.27812115 -0.69069536
#>  [79]  0.36025298 -0.03683946  1.02876647 -0.78507988 -1.21740410  1.36461961
#>  [85] -0.76772266 -1.18637769  0.57949318 -0.97167920  0.27029927  0.22525594
#>  [91] -0.02542103 -0.11368237  1.06482032 -0.92460744  0.93326615 -1.73372638
#>  [97]  3.53961829  0.65098084 -0.03284057 -0.56964996
#> 
#> [[10]]
#>   [1] -1.129649208  1.817003448 -1.515520418 -1.482618636  1.033971976
#>   [6]  1.146123401 -0.771824359 -0.520822367  0.308664601  0.912043018
#>  [11]  1.106561742 -1.676273462 -0.377725131  0.542709092  0.772184856
#>  [16]  1.887570504 -0.805809925 -1.456624609  1.200877947 -0.174928850
#>  [21]  1.459196658  0.417844865 -0.392537492  0.684233844 -0.805795177
#>  [26] -0.161521813 -1.349797698 -0.509517701 -0.275166917  1.162925447
#>  [31] -0.181425104  0.565889956  1.164233531  1.793665109  0.349080607
#>  [36] -0.207297302 -0.163073442 -1.303572354  0.196046866 -1.061084029
#>  [41] -0.579738490  0.231841345 -0.101044155 -0.062112468  2.096408069
#>  [46] -1.374100769 -1.119169577  1.733456668 -2.029495073 -1.606049672
#>  [51]  1.634472515  0.233204799 -0.009301704 -1.636130189 -1.021147110
#>  [56] -0.574264847  0.341813787  0.602321748 -0.058284936 -0.178774576
#>  [61] -0.514487683 -0.679647813  0.148994146 -1.736388364  0.111911635
#>  [66] -0.065044894  1.600534059 -1.534275876 -1.585720644  0.366134122
#>  [71]  0.574642309  0.742691333  0.475064005 -0.289898725  0.342612604
#>  [76] -1.928049429  1.014252185 -0.594450006 -0.531533582  0.084686118
#>  [81] -0.081586088  1.259259337  0.838780729  0.397052541  0.257894034
#>  [86]  0.492450986 -3.090782827  1.640825867 -1.986748545 -0.558696147
#>  [91] -0.957816449  0.490241566 -0.160603153 -0.995439433 -0.590935637
#>  [96]  2.281940559  0.494522733 -0.910702760 -0.946559592  1.261897172
#>