Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd        RPS
#> 1      54 -5.303305 0.16197520
#> 2      45 -5.303305 0.11764684
#> 3      49 -5.303305 0.13635331
#> 4      50 -5.303305 0.14127868
#> 5      49 -5.303305 0.13635331
#> 6      53 -5.303305 0.15665182
#> 7      47 -5.303305 0.12680107
#> 8      58 -5.303305 0.18426376
#> 9      56 -5.303305 0.17292047
#> 10     59 -5.303305 0.19008465
#> 11     53 -5.303305 0.15665182
#> 12     55 -5.303305 0.16739808
#> 13     46 -5.303305 0.12217420
#> 14     42 -5.303305 0.10466177
#> 15     52 -5.303305 0.15142793
#> 16     52 -5.303305 0.15142793
#> 17     55 -5.303305 0.16739808
#> 18     57 -5.303305 0.17854236
#> 19     55 -5.303305 0.16739808
#> 20     58 -5.303305 0.18426376
#> 21     54 -5.303305 0.16197520
#> 22     50 -5.303305 0.14127868
#> 23     44 -5.303305 0.11321898
#> 24     52 -5.303305 0.15142793
#> 25     49 -5.303305 0.13635331
#> 26     54 -5.303305 0.16197520
#> 27     52 -5.303305 0.15142793
#> 28     49 -5.303305 0.13635331
#> 29     51 -5.303305 0.14630356
#> 30     54 -5.303305 0.16197520
#> 31     57 -5.303305 0.17854236
#> 32     57 -5.303305 0.17854236
#> 33     45 -5.303305 0.11764684
#> 34     46 -5.303305 0.12217420
#> 35     49 -5.303305 0.13635331
#> 36     48 -5.303305 0.13152744
#> 37     48 -5.303305 0.13152744
#> 38     47 -5.303305 0.12680107
#> 39     55 -5.303305 0.16739808
#> 40     55 -5.303305 0.16739808
#> 41     49 -5.303305 0.13635331
#> 42     54 -5.303305 0.16197520
#> 43     50 -5.303305 0.14127868
#> 44     51 -5.303305 0.14630356
#> 45     52 -5.303305 0.15142793
#> 46     48 -5.303305 0.13152744
#> 47     43 -5.303305 0.10889062
#> 48     42 -5.303305 0.10466177
#> 49     54 -5.303305 0.16197520
#> 50     57 -5.303305 0.17854236
#> 51     50 -5.303305 0.14127868
#> 52     56 -5.303305 0.17292047
#> 53     53 -5.303305 0.15665182
#> 54     49 -5.303305 0.13635331
#> 55     43 -5.303305 0.10889062
#> 56     50 -5.303305 0.14127868
#> 57     45 -5.303305 0.11764684
#> 58     51 -5.303305 0.14630356
#> 59     47 -5.303305 0.12680107
#> 60     60 -5.303305 0.19600505
#> 61     61 -5.303305 0.20202495
#> 62     40 -5.303305 0.09650256
#> 63     53 -5.303305 0.15665182
#> 64     48 -5.303305 0.13152744
#> 65     50 -5.303305 0.14127868
#> 66     58 -5.303305 0.18426376
#> 67     46 -5.303305 0.12217420
#> 68     45 -5.303305 0.11764684
#> 69     55 -5.303305 0.16739808
#> 70     43 -5.303305 0.10889062
#> 71     46 -5.303305 0.12217420
#> 72     49 -5.303305 0.13635331
#> 73     48 -5.303305 0.13152744
#> 74     50 -5.303305 0.14127868
#> 75     48 -5.303305 0.13152744
#> 76     50 -5.303305 0.14127868
#> 77     60 -5.303305 0.19600505
#> 78     54 -5.303305 0.16197520
#> 79     53 -5.303305 0.15665182
#> 80     47 -5.303305 0.12680107
#> 81     56 -5.303305 0.17292047
#> 82     51 -5.303305 0.14630356
#> 83     48 -5.303305 0.13152744
#> 84     55 -5.303305 0.16739808
#> 85     46 -5.303305 0.12217420
#> 86     61 -5.303305 0.20202495
#> 87     42 -5.303305 0.10466177
#> 88     41 -5.303305 0.10053241
#> 89     44 -5.303305 0.11321898
#> 90     47 -5.303305 0.12680107
#> 91     43 -5.303305 0.10889062
#> 92     51 -5.303305 0.14630356
#> 93     46 -5.303305 0.12217420
#> 94     44 -5.303305 0.11321898
#> 95     47 -5.303305 0.12680107
#> 96     48 -5.303305 0.13152744
#> 97     48 -5.303305 0.13152744
#> 98     53 -5.303305 0.15665182
#> 99     49 -5.303305 0.13635331
#> 100    47 -5.303305 0.12680107