Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd       RPS
#> 1      51 -5.303305 0.1498569
#> 2      44 -5.303305 0.1167724
#> 3      61 -5.303305 0.2055783
#> 4      47 -5.303305 0.1303544
#> 5      45 -5.303305 0.1212002
#> 6      50 -5.303305 0.1448321
#> 7      51 -5.303305 0.1498569
#> 8      59 -5.303305 0.1936380
#> 9      51 -5.303305 0.1498569
#> 10     53 -5.303305 0.1602052
#> 11     63 -5.303305 0.2179166
#> 12     46 -5.303305 0.1257276
#> 13     51 -5.303305 0.1498569
#> 14     53 -5.303305 0.1602052
#> 15     43 -5.303305 0.1124440
#> 16     43 -5.303305 0.1124440
#> 17     52 -5.303305 0.1549813
#> 18     41 -5.303305 0.1040858
#> 19     42 -5.303305 0.1082151
#> 20     50 -5.303305 0.1448321
#> 21     61 -5.303305 0.2055783
#> 22     40 -5.303305 0.1000559
#> 23     46 -5.303305 0.1257276
#> 24     46 -5.303305 0.1257276
#> 25     51 -5.303305 0.1498569
#> 26     48 -5.303305 0.1350808
#> 27     53 -5.303305 0.1602052
#> 28     55 -5.303305 0.1709515
#> 29     50 -5.303305 0.1448321
#> 30     54 -5.303305 0.1655286
#> 31     45 -5.303305 0.1212002
#> 32     45 -5.303305 0.1212002
#> 33     48 -5.303305 0.1350808
#> 34     51 -5.303305 0.1498569
#> 35     49 -5.303305 0.1399067
#> 36     48 -5.303305 0.1350808
#> 37     54 -5.303305 0.1655286
#> 38     41 -5.303305 0.1040858
#> 39     41 -5.303305 0.1040858
#> 40     62 -5.303305 0.2116977
#> 41     42 -5.303305 0.1082151
#> 42     52 -5.303305 0.1549813
#> 43     43 -5.303305 0.1124440
#> 44     44 -5.303305 0.1167724
#> 45     45 -5.303305 0.1212002
#> 46     57 -5.303305 0.1820957
#> 47     54 -5.303305 0.1655286
#> 48     60 -5.303305 0.1995584
#> 49     53 -5.303305 0.1602052
#> 50     53 -5.303305 0.1602052
#> 51     51 -5.303305 0.1498569
#> 52     50 -5.303305 0.1448321
#> 53     55 -5.303305 0.1709515
#> 54     41 -5.303305 0.1040858
#> 55     57 -5.303305 0.1820957
#> 56     55 -5.303305 0.1709515
#> 57     37 -5.303305 0.0885634
#> 58     51 -5.303305 0.1498569
#> 59     50 -5.303305 0.1448321
#> 60     54 -5.303305 0.1655286
#> 61     52 -5.303305 0.1549813
#> 62     50 -5.303305 0.1448321
#> 63     42 -5.303305 0.1082151
#> 64     57 -5.303305 0.1820957
#> 65     48 -5.303305 0.1350808
#> 66     44 -5.303305 0.1167724
#> 67     41 -5.303305 0.1040858
#> 68     48 -5.303305 0.1350808
#> 69     44 -5.303305 0.1167724
#> 70     40 -5.303305 0.1000559
#> 71     46 -5.303305 0.1257276
#> 72     49 -5.303305 0.1399067
#> 73     53 -5.303305 0.1602052
#> 74     53 -5.303305 0.1602052
#> 75     49 -5.303305 0.1399067
#> 76     45 -5.303305 0.1212002
#> 77     49 -5.303305 0.1399067
#> 78     51 -5.303305 0.1498569
#> 79     58 -5.303305 0.1878171
#> 80     51 -5.303305 0.1498569
#> 81     47 -5.303305 0.1303544
#> 82     52 -5.303305 0.1549813
#> 83     49 -5.303305 0.1399067
#> 84     48 -5.303305 0.1350808
#> 85     52 -5.303305 0.1549813
#> 86     52 -5.303305 0.1549813
#> 87     52 -5.303305 0.1549813
#> 88     51 -5.303305 0.1498569
#> 89     49 -5.303305 0.1399067
#> 90     43 -5.303305 0.1124440
#> 91     47 -5.303305 0.1303544
#> 92     44 -5.303305 0.1167724
#> 93     43 -5.303305 0.1124440
#> 94     48 -5.303305 0.1350808
#> 95     58 -5.303305 0.1878171
#> 96     53 -5.303305 0.1602052
#> 97     52 -5.303305 0.1549813
#> 98     55 -5.303305 0.1709515
#> 99     41 -5.303305 0.1040858
#> 100    50 -5.303305 0.1448321