Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd        RPS
#> 1      57 -5.303305 0.17861439
#> 2      50 -5.303305 0.14135071
#> 3      48 -5.303305 0.13159947
#> 4      43 -5.303305 0.10896265
#> 5      51 -5.303305 0.14637558
#> 6      55 -5.303305 0.16747011
#> 7      47 -5.303305 0.12687310
#> 8      51 -5.303305 0.14637558
#> 9      50 -5.303305 0.14135071
#> 10     51 -5.303305 0.14637558
#> 11     55 -5.303305 0.16747011
#> 12     55 -5.303305 0.16747011
#> 13     51 -5.303305 0.14637558
#> 14     42 -5.303305 0.10473379
#> 15     49 -5.303305 0.13642534
#> 16     54 -5.303305 0.16204723
#> 17     56 -5.303305 0.17299250
#> 18     53 -5.303305 0.15672384
#> 19     51 -5.303305 0.14637558
#> 20     53 -5.303305 0.15672384
#> 21     52 -5.303305 0.15149996
#> 22     49 -5.303305 0.13642534
#> 23     56 -5.303305 0.17299250
#> 24     45 -5.303305 0.11771887
#> 25     55 -5.303305 0.16747011
#> 26     48 -5.303305 0.13159947
#> 27     38 -5.303305 0.08881340
#> 28     50 -5.303305 0.14135071
#> 29     58 -5.303305 0.18433578
#> 30     56 -5.303305 0.17299250
#> 31     53 -5.303305 0.15672384
#> 32     45 -5.303305 0.11771887
#> 33     53 -5.303305 0.15672384
#> 34     48 -5.303305 0.13159947
#> 35     47 -5.303305 0.12687310
#> 36     46 -5.303305 0.12224623
#> 37     50 -5.303305 0.14135071
#> 38     51 -5.303305 0.14637558
#> 39     43 -5.303305 0.10896265
#> 40     44 -5.303305 0.11329101
#> 41     60 -5.303305 0.19607708
#> 42     42 -5.303305 0.10473379
#> 43     61 -5.303305 0.20209698
#> 44     40 -5.303305 0.09657459
#> 45     47 -5.303305 0.12687310
#> 46     51 -5.303305 0.14637558
#> 47     52 -5.303305 0.15149996
#> 48     47 -5.303305 0.12687310
#> 49     51 -5.303305 0.14637558
#> 50     44 -5.303305 0.11329101
#> 51     49 -5.303305 0.13642534
#> 52     45 -5.303305 0.11771887
#> 53     46 -5.303305 0.12224623
#> 54     49 -5.303305 0.13642534
#> 55     40 -5.303305 0.09657459
#> 56     44 -5.303305 0.11329101
#> 57     51 -5.303305 0.14637558
#> 58     43 -5.303305 0.10896265
#> 59     49 -5.303305 0.13642534
#> 60     47 -5.303305 0.12687310
#> 61     57 -5.303305 0.17861439
#> 62     46 -5.303305 0.12224623
#> 63     57 -5.303305 0.17861439
#> 64     41 -5.303305 0.10060444
#> 65     56 -5.303305 0.17299250
#> 66     58 -5.303305 0.18433578
#> 67     53 -5.303305 0.15672384
#> 68     54 -5.303305 0.16204723
#> 69     56 -5.303305 0.17299250
#> 70     39 -5.303305 0.09264424
#> 71     50 -5.303305 0.14135071
#> 72     55 -5.303305 0.16747011
#> 73     45 -5.303305 0.11771887
#> 74     50 -5.303305 0.14135071
#> 75     48 -5.303305 0.13159947
#> 76     46 -5.303305 0.12224623
#> 77     48 -5.303305 0.13159947
#> 78     45 -5.303305 0.11771887
#> 79     45 -5.303305 0.11771887
#> 80     47 -5.303305 0.12687310
#> 81     53 -5.303305 0.15672384
#> 82     44 -5.303305 0.11329101
#> 83     53 -5.303305 0.15672384
#> 84     50 -5.303305 0.14135071
#> 85     58 -5.303305 0.18433578
#> 86     43 -5.303305 0.10896265
#> 87     57 -5.303305 0.17861439
#> 88     54 -5.303305 0.16204723
#> 89     54 -5.303305 0.16204723
#> 90     49 -5.303305 0.13642534
#> 91     48 -5.303305 0.13159947
#> 92     43 -5.303305 0.10896265
#> 93     52 -5.303305 0.15149996
#> 94     46 -5.303305 0.12224623
#> 95     47 -5.303305 0.12687310
#> 96     48 -5.303305 0.13159947
#> 97     55 -5.303305 0.16747011
#> 98     45 -5.303305 0.11771887
#> 99     55 -5.303305 0.16747011
#> 100    51 -5.303305 0.14637558