Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd        RPS
#> 1      57 -5.303305 0.18034380
#> 2      58 -5.303305 0.18606520
#> 3      54 -5.303305 0.16377664
#> 4      46 -5.303305 0.12397564
#> 5      42 -5.303305 0.10646321
#> 6      59 -5.303305 0.19188609
#> 7      45 -5.303305 0.11944828
#> 8      51 -5.303305 0.14810500
#> 9      48 -5.303305 0.13332888
#> 10     51 -5.303305 0.14810500
#> 11     42 -5.303305 0.10646321
#> 12     44 -5.303305 0.11502042
#> 13     52 -5.303305 0.15322938
#> 14     47 -5.303305 0.12860251
#> 15     51 -5.303305 0.14810500
#> 16     51 -5.303305 0.14810500
#> 17     40 -5.303305 0.09830400
#> 18     51 -5.303305 0.14810500
#> 19     43 -5.303305 0.11069206
#> 20     56 -5.303305 0.17472191
#> 21     47 -5.303305 0.12860251
#> 22     60 -5.303305 0.19780649
#> 23     53 -5.303305 0.15845326
#> 24     59 -5.303305 0.19188609
#> 25     54 -5.303305 0.16377664
#> 26     45 -5.303305 0.11944828
#> 27     48 -5.303305 0.13332888
#> 28     45 -5.303305 0.11944828
#> 29     53 -5.303305 0.15845326
#> 30     48 -5.303305 0.13332888
#> 31     55 -5.303305 0.16919952
#> 32     45 -5.303305 0.11944828
#> 33     47 -5.303305 0.12860251
#> 34     48 -5.303305 0.13332888
#> 35     48 -5.303305 0.13332888
#> 36     45 -5.303305 0.11944828
#> 37     45 -5.303305 0.11944828
#> 38     47 -5.303305 0.12860251
#> 39     51 -5.303305 0.14810500
#> 40     61 -5.303305 0.20382639
#> 41     45 -5.303305 0.11944828
#> 42     51 -5.303305 0.14810500
#> 43     48 -5.303305 0.13332888
#> 44     46 -5.303305 0.12397564
#> 45     48 -5.303305 0.13332888
#> 46     56 -5.303305 0.17472191
#> 47     51 -5.303305 0.14810500
#> 48     56 -5.303305 0.17472191
#> 49     38 -5.303305 0.09054281
#> 50     49 -5.303305 0.13815475
#> 51     49 -5.303305 0.13815475
#> 52     42 -5.303305 0.10646321
#> 53     51 -5.303305 0.14810500
#> 54     45 -5.303305 0.11944828
#> 55     53 -5.303305 0.15845326
#> 56     48 -5.303305 0.13332888
#> 57     54 -5.303305 0.16377664
#> 58     44 -5.303305 0.11502042
#> 59     57 -5.303305 0.18034380
#> 60     44 -5.303305 0.11502042
#> 61     50 -5.303305 0.14308012
#> 62     57 -5.303305 0.18034380
#> 63     54 -5.303305 0.16377664
#> 64     45 -5.303305 0.11944828
#> 65     55 -5.303305 0.16919952
#> 66     53 -5.303305 0.15845326
#> 67     43 -5.303305 0.11069206
#> 68     43 -5.303305 0.11069206
#> 69     54 -5.303305 0.16377664
#> 70     57 -5.303305 0.18034380
#> 71     50 -5.303305 0.14308012
#> 72     53 -5.303305 0.15845326
#> 73     51 -5.303305 0.14810500
#> 74     44 -5.303305 0.11502042
#> 75     51 -5.303305 0.14810500
#> 76     44 -5.303305 0.11502042
#> 77     49 -5.303305 0.13815475
#> 78     59 -5.303305 0.19188609
#> 79     62 -5.303305 0.20994579
#> 80     46 -5.303305 0.12397564
#> 81     41 -5.303305 0.10233385
#> 82     42 -5.303305 0.10646321
#> 83     55 -5.303305 0.16919952
#> 84     40 -5.303305 0.09830400
#> 85     42 -5.303305 0.10646321
#> 86     60 -5.303305 0.19780649
#> 87     52 -5.303305 0.15322938
#> 88     48 -5.303305 0.13332888
#> 89     53 -5.303305 0.15845326
#> 90     51 -5.303305 0.14810500
#> 91     54 -5.303305 0.16377664
#> 92     54 -5.303305 0.16377664
#> 93     45 -5.303305 0.11944828
#> 94     47 -5.303305 0.12860251
#> 95     53 -5.303305 0.15845326
#> 96     65 -5.303305 0.22890102
#> 97     45 -5.303305 0.11944828
#> 98     49 -5.303305 0.13815475
#> 99     46 -5.303305 0.12397564
#> 100    50 -5.303305 0.14308012