Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd        RPS
#> 1      42 -5.303305 0.10446722
#> 2      55 -5.303305 0.16720353
#> 3      50 -5.303305 0.14108413
#> 4      53 -5.303305 0.15645727
#> 5      50 -5.303305 0.14108413
#> 6      52 -5.303305 0.15123339
#> 7      45 -5.303305 0.11745229
#> 8      48 -5.303305 0.13133289
#> 9      51 -5.303305 0.14610901
#> 10     53 -5.303305 0.15645727
#> 11     49 -5.303305 0.13615876
#> 12     49 -5.303305 0.13615876
#> 13     48 -5.303305 0.13133289
#> 14     48 -5.303305 0.13133289
#> 15     42 -5.303305 0.10446722
#> 16     50 -5.303305 0.14108413
#> 17     49 -5.303305 0.13615876
#> 18     49 -5.303305 0.13615876
#> 19     49 -5.303305 0.13615876
#> 20     44 -5.303305 0.11302443
#> 21     53 -5.303305 0.15645727
#> 22     53 -5.303305 0.15645727
#> 23     54 -5.303305 0.16178065
#> 24     57 -5.303305 0.17834781
#> 25     44 -5.303305 0.11302443
#> 26     54 -5.303305 0.16178065
#> 27     50 -5.303305 0.14108413
#> 28     51 -5.303305 0.14610901
#> 29     47 -5.303305 0.12660652
#> 30     49 -5.303305 0.13615876
#> 31     57 -5.303305 0.17834781
#> 32     55 -5.303305 0.16720353
#> 33     52 -5.303305 0.15123339
#> 34     54 -5.303305 0.16178065
#> 35     56 -5.303305 0.17272592
#> 36     49 -5.303305 0.13615876
#> 37     43 -5.303305 0.10869607
#> 38     50 -5.303305 0.14108413
#> 39     40 -5.303305 0.09630801
#> 40     46 -5.303305 0.12197965
#> 41     60 -5.303305 0.19581050
#> 42     46 -5.303305 0.12197965
#> 43     51 -5.303305 0.14610901
#> 44     50 -5.303305 0.14108413
#> 45     39 -5.303305 0.09237766
#> 46     48 -5.303305 0.13133289
#> 47     49 -5.303305 0.13615876
#> 48     53 -5.303305 0.15645727
#> 49     50 -5.303305 0.14108413
#> 50     45 -5.303305 0.11745229
#> 51     51 -5.303305 0.14610901
#> 52     52 -5.303305 0.15123339
#> 53     55 -5.303305 0.16720353
#> 54     63 -5.303305 0.21416871
#> 55     52 -5.303305 0.15123339
#> 56     48 -5.303305 0.13133289
#> 57     48 -5.303305 0.13133289
#> 58     50 -5.303305 0.14108413
#> 59     47 -5.303305 0.12660652
#> 60     50 -5.303305 0.14108413
#> 61     44 -5.303305 0.11302443
#> 62     49 -5.303305 0.13615876
#> 63     52 -5.303305 0.15123339
#> 64     50 -5.303305 0.14108413
#> 65     47 -5.303305 0.12660652
#> 66     49 -5.303305 0.13615876
#> 67     50 -5.303305 0.14108413
#> 68     52 -5.303305 0.15123339
#> 69     50 -5.303305 0.14108413
#> 70     47 -5.303305 0.12660652
#> 71     50 -5.303305 0.14108413
#> 72     58 -5.303305 0.18406921
#> 73     55 -5.303305 0.16720353
#> 74     57 -5.303305 0.17834781
#> 75     46 -5.303305 0.12197965
#> 76     58 -5.303305 0.18406921
#> 77     49 -5.303305 0.13615876
#> 78     37 -5.303305 0.08481547
#> 79     54 -5.303305 0.16178065
#> 80     52 -5.303305 0.15123339
#> 81     51 -5.303305 0.14610901
#> 82     46 -5.303305 0.12197965
#> 83     49 -5.303305 0.13615876
#> 84     51 -5.303305 0.14610901
#> 85     48 -5.303305 0.13133289
#> 86     45 -5.303305 0.11745229
#> 87     41 -5.303305 0.10033786
#> 88     47 -5.303305 0.12660652
#> 89     42 -5.303305 0.10446722
#> 90     55 -5.303305 0.16720353
#> 91     33 -5.303305 0.07088513
#> 92     46 -5.303305 0.12197965
#> 93     47 -5.303305 0.12660652
#> 94     46 -5.303305 0.12197965
#> 95     51 -5.303305 0.14610901
#> 96     45 -5.303305 0.11745229
#> 97     60 -5.303305 0.19581050
#> 98     50 -5.303305 0.14108413
#> 99     45 -5.303305 0.11745229
#> 100    57 -5.303305 0.17834781