Performance of (population) historical forecast

Performance of (population) historical forecast

Usage

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

train

Training dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

add_uniform

Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)
#>     Score       lpd       RPS
#> 1      47 -5.303305 0.1283500
#> 2      56 -5.303305 0.1744694
#> 3      51 -5.303305 0.1478525
#> 4      52 -5.303305 0.1529769
#> 5      61 -5.303305 0.2035739
#> 6      50 -5.303305 0.1428277
#> 7      55 -5.303305 0.1689471
#> 8      45 -5.303305 0.1191958
#> 9      46 -5.303305 0.1237232
#> 10     54 -5.303305 0.1635242
#> 11     57 -5.303305 0.1800913
#> 12     45 -5.303305 0.1191958
#> 13     49 -5.303305 0.1379023
#> 14     62 -5.303305 0.2096933
#> 15     55 -5.303305 0.1689471
#> 16     49 -5.303305 0.1379023
#> 17     52 -5.303305 0.1529769
#> 18     49 -5.303305 0.1379023
#> 19     53 -5.303305 0.1582008
#> 20     59 -5.303305 0.1916336
#> 21     50 -5.303305 0.1428277
#> 22     50 -5.303305 0.1428277
#> 23     55 -5.303305 0.1689471
#> 24     56 -5.303305 0.1744694
#> 25     49 -5.303305 0.1379023
#> 26     51 -5.303305 0.1478525
#> 27     43 -5.303305 0.1104396
#> 28     48 -5.303305 0.1330764
#> 29     63 -5.303305 0.2159122
#> 30     48 -5.303305 0.1330764
#> 31     53 -5.303305 0.1582008
#> 32     53 -5.303305 0.1582008
#> 33     53 -5.303305 0.1582008
#> 34     55 -5.303305 0.1689471
#> 35     50 -5.303305 0.1428277
#> 36     46 -5.303305 0.1237232
#> 37     55 -5.303305 0.1689471
#> 38     46 -5.303305 0.1237232
#> 39     45 -5.303305 0.1191958
#> 40     54 -5.303305 0.1635242
#> 41     57 -5.303305 0.1800913
#> 42     53 -5.303305 0.1582008
#> 43     47 -5.303305 0.1283500
#> 44     54 -5.303305 0.1635242
#> 45     46 -5.303305 0.1237232
#> 46     41 -5.303305 0.1020814
#> 47     44 -5.303305 0.1147680
#> 48     49 -5.303305 0.1379023
#> 49     51 -5.303305 0.1478525
#> 50     51 -5.303305 0.1478525
#> 51     53 -5.303305 0.1582008
#> 52     56 -5.303305 0.1744694
#> 53     49 -5.303305 0.1379023
#> 54     57 -5.303305 0.1800913
#> 55     50 -5.303305 0.1428277
#> 56     53 -5.303305 0.1582008
#> 57     46 -5.303305 0.1237232
#> 58     46 -5.303305 0.1237232
#> 59     53 -5.303305 0.1582008
#> 60     55 -5.303305 0.1689471
#> 61     52 -5.303305 0.1529769
#> 62     50 -5.303305 0.1428277
#> 63     54 -5.303305 0.1635242
#> 64     53 -5.303305 0.1582008
#> 65     50 -5.303305 0.1428277
#> 66     45 -5.303305 0.1191958
#> 67     59 -5.303305 0.1916336
#> 68     50 -5.303305 0.1428277
#> 69     54 -5.303305 0.1635242
#> 70     46 -5.303305 0.1237232
#> 71     58 -5.303305 0.1858127
#> 72     53 -5.303305 0.1582008
#> 73     46 -5.303305 0.1237232
#> 74     50 -5.303305 0.1428277
#> 75     49 -5.303305 0.1379023
#> 76     57 -5.303305 0.1800913
#> 77     46 -5.303305 0.1237232
#> 78     49 -5.303305 0.1379023
#> 79     52 -5.303305 0.1529769
#> 80     52 -5.303305 0.1529769
#> 81     45 -5.303305 0.1191958
#> 82     46 -5.303305 0.1237232
#> 83     55 -5.303305 0.1689471
#> 84     53 -5.303305 0.1582008
#> 85     42 -5.303305 0.1062107
#> 86     50 -5.303305 0.1428277
#> 87     50 -5.303305 0.1428277
#> 88     48 -5.303305 0.1330764
#> 89     46 -5.303305 0.1237232
#> 90     49 -5.303305 0.1379023
#> 91     49 -5.303305 0.1379023
#> 92     47 -5.303305 0.1283500
#> 93     54 -5.303305 0.1635242
#> 94     58 -5.303305 0.1858127
#> 95     47 -5.303305 0.1283500
#> 96     58 -5.303305 0.1858127
#> 97     48 -5.303305 0.1330764
#> 98     52 -5.303305 0.1529769
#> 99     45 -5.303305 0.1191958
#> 100    46 -5.303305 0.1237232