Performance of (population) historical forecast

Performance of (population) historical forecast

add_historical_pred(
  test,
  train,
  max_score,
  discrete = TRUE,
  add_uniform = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test
Testing dataframe. The only requirements is that it contains a column "Score".

train
Training dataframe. The only requirements is that it contains a column "Score".

max_score
Maximum value that the score can take

discrete
Whether to estimate a discrete or continuous historical forecast

add_uniform
Whether to include samples from uniform distribution when computing a discrete historical forecast. This ensures that all states are visited.

include_samples
Whether to return samples from the historical forecast in the output

n_samples
If include_samples=TRUE, how many samples to return. When NULL, the function return the training set.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Details

The continuous historical forecast is calculated by considering the training set as samples from the predictive distribution.

Examples

max_score <- 100
train <- data.frame(Score = rbinom(1e2, max_score, 0.2))
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_historical_pred(test, train, max_score)

#>     Score       lpd        RPS
#> 1      56 -5.303305 0.17492042
#> 2      52 -5.303305 0.15342789
#> 3      43 -5.303305 0.11089057
#> 4      50 -5.303305 0.14327863
#> 5      40 -5.303305 0.09850251
#> 6      47 -5.303305 0.12880102
#> 7      37 -5.303305 0.08700998
#> 8      51 -5.303305 0.14830351
#> 9      51 -5.303305 0.14830351
#> 10     47 -5.303305 0.12880102
#> 11     41 -5.303305 0.10253236
#> 12     51 -5.303305 0.14830351
#> 13     43 -5.303305 0.11089057
#> 14     43 -5.303305 0.11089057
#> 15     48 -5.303305 0.13352739
#> 16     47 -5.303305 0.12880102
#> 17     52 -5.303305 0.15342789
#> 18     53 -5.303305 0.15865177
#> 19     42 -5.303305 0.10666172
#> 20     46 -5.303305 0.12417415
#> 21     52 -5.303305 0.15342789
#> 22     49 -5.303305 0.13835326
#> 23     37 -5.303305 0.08700998
#> 24     42 -5.303305 0.10666172
#> 25     52 -5.303305 0.15342789
#> 26     50 -5.303305 0.14327863
#> 27     50 -5.303305 0.14327863
#> 28     34 -5.303305 0.07641296
#> 29     58 -5.303305 0.18626371
#> 30     43 -5.303305 0.11089057
#> 31     51 -5.303305 0.14830351
#> 32     62 -5.303305 0.21014430
#> 33     49 -5.303305 0.13835326
#> 34     57 -5.303305 0.18054231
#> 35     49 -5.303305 0.13835326
#> 36     49 -5.303305 0.13835326
#> 37     54 -5.303305 0.16397515
#> 38     45 -5.303305 0.11964679
#> 39     52 -5.303305 0.15342789
#> 40     43 -5.303305 0.11089057
#> 41     48 -5.303305 0.13352739
#> 42     48 -5.303305 0.13352739
#> 43     47 -5.303305 0.12880102
#> 44     57 -5.303305 0.18054231
#> 45     46 -5.303305 0.12417415
#> 46     48 -5.303305 0.13352739
#> 47     46 -5.303305 0.12417415
#> 48     48 -5.303305 0.13352739
#> 49     40 -5.303305 0.09850251
#> 50     45 -5.303305 0.11964679
#> 51     44 -5.303305 0.11521893
#> 52     49 -5.303305 0.13835326
#> 53     55 -5.303305 0.16939803
#> 54     47 -5.303305 0.12880102
#> 55     44 -5.303305 0.11521893
#> 56     49 -5.303305 0.13835326
#> 57     47 -5.303305 0.12880102
#> 58     45 -5.303305 0.11964679
#> 59     52 -5.303305 0.15342789
#> 60     57 -5.303305 0.18054231
#> 61     54 -5.303305 0.16397515
#> 62     54 -5.303305 0.16397515
#> 63     52 -5.303305 0.15342789
#> 64     50 -5.303305 0.14327863
#> 65     63 -5.303305 0.21636321
#> 66     50 -5.303305 0.14327863
#> 67     55 -5.303305 0.16939803
#> 68     58 -5.303305 0.18626371
#> 69     55 -5.303305 0.16939803
#> 70     55 -5.303305 0.16939803
#> 71     56 -5.303305 0.17492042
#> 72     44 -5.303305 0.11521893
#> 73     56 -5.303305 0.17492042
#> 74     44 -5.303305 0.11521893
#> 75     54 -5.303305 0.16397515
#> 76     47 -5.303305 0.12880102
#> 77     51 -5.303305 0.14830351
#> 78     48 -5.303305 0.13352739
#> 79     55 -5.303305 0.16939803
#> 80     55 -5.303305 0.16939803
#> 81     45 -5.303305 0.11964679
#> 82     57 -5.303305 0.18054231
#> 83     48 -5.303305 0.13352739
#> 84     44 -5.303305 0.11521893
#> 85     50 -5.303305 0.14327863
#> 86     59 -5.303305 0.19208460
#> 87     53 -5.303305 0.15865177
#> 88     40 -5.303305 0.09850251
#> 89     39 -5.303305 0.09457216
#> 90     41 -5.303305 0.10253236
#> 91     57 -5.303305 0.18054231
#> 92     46 -5.303305 0.12417415
#> 93     48 -5.303305 0.13352739
#> 94     46 -5.303305 0.12417415
#> 95     39 -5.303305 0.09457216
#> 96     41 -5.303305 0.10253236
#> 97     53 -5.303305 0.15865177
#> 98     50 -5.303305 0.14327863
#> 99     42 -5.303305 0.10666172
#> 100    56 -5.303305 0.17492042