Performance of a uniform forecast

Performance of a uniform forecast

add_uniform_pred(
  test,
  max_score,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the samples 0:max_score are returned.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Examples

max_score <- 100
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_uniform_pred(test, max_score)
#>     Score       lpd        RPS
#> 1      52 -4.615121 0.08455446
#> 2      55 -4.615121 0.08663366
#> 3      53 -4.615121 0.08504950
#> 4      52 -4.615121 0.08455446
#> 5      50 -4.615121 0.08415842
#> 6      54 -4.615121 0.08574257
#> 7      54 -4.615121 0.08574257
#> 8      46 -4.615121 0.08574257
#> 9      45 -4.615121 0.08663366
#> 10     47 -4.615121 0.08504950
#> 11     44 -4.615121 0.08772277
#> 12     57 -4.615121 0.08900990
#> 13     52 -4.615121 0.08455446
#> 14     49 -4.615121 0.08425743
#> 15     43 -4.615121 0.08900990
#> 16     47 -4.615121 0.08504950
#> 17     53 -4.615121 0.08504950
#> 18     48 -4.615121 0.08455446
#> 19     51 -4.615121 0.08425743
#> 20     41 -4.615121 0.09217822
#> 21     50 -4.615121 0.08415842
#> 22     53 -4.615121 0.08504950
#> 23     57 -4.615121 0.08900990
#> 24     48 -4.615121 0.08455446
#> 25     44 -4.615121 0.08772277
#> 26     59 -4.615121 0.09217822
#> 27     45 -4.615121 0.08663366
#> 28     53 -4.615121 0.08504950
#> 29     51 -4.615121 0.08425743
#> 30     55 -4.615121 0.08663366
#> 31     43 -4.615121 0.08900990
#> 32     49 -4.615121 0.08425743
#> 33     52 -4.615121 0.08455446
#> 34     57 -4.615121 0.08900990
#> 35     39 -4.615121 0.09613861
#> 36     68 -4.615121 0.11623762
#> 37     50 -4.615121 0.08415842
#> 38     52 -4.615121 0.08455446
#> 39     55 -4.615121 0.08663366
#> 40     49 -4.615121 0.08425743
#> 41     40 -4.615121 0.09405941
#> 42     54 -4.615121 0.08574257
#> 43     52 -4.615121 0.08455446
#> 44     54 -4.615121 0.08574257
#> 45     60 -4.615121 0.09405941
#> 46     52 -4.615121 0.08455446
#> 47     51 -4.615121 0.08425743
#> 48     47 -4.615121 0.08504950
#> 49     50 -4.615121 0.08415842
#> 50     49 -4.615121 0.08425743
#> 51     47 -4.615121 0.08504950
#> 52     44 -4.615121 0.08772277
#> 53     48 -4.615121 0.08455446
#> 54     43 -4.615121 0.08900990
#> 55     52 -4.615121 0.08455446
#> 56     52 -4.615121 0.08455446
#> 57     58 -4.615121 0.09049505
#> 58     49 -4.615121 0.08425743
#> 59     42 -4.615121 0.09049505
#> 60     54 -4.615121 0.08574257
#> 61     46 -4.615121 0.08574257
#> 62     52 -4.615121 0.08455446
#> 63     54 -4.615121 0.08574257
#> 64     40 -4.615121 0.09405941
#> 65     38 -4.615121 0.09841584
#> 66     47 -4.615121 0.08504950
#> 67     47 -4.615121 0.08504950
#> 68     50 -4.615121 0.08415842
#> 69     58 -4.615121 0.09049505
#> 70     47 -4.615121 0.08504950
#> 71     48 -4.615121 0.08455446
#> 72     41 -4.615121 0.09217822
#> 73     48 -4.615121 0.08455446
#> 74     57 -4.615121 0.08900990
#> 75     50 -4.615121 0.08415842
#> 76     53 -4.615121 0.08504950
#> 77     44 -4.615121 0.08772277
#> 78     49 -4.615121 0.08425743
#> 79     64 -4.615121 0.10356436
#> 80     48 -4.615121 0.08455446
#> 81     53 -4.615121 0.08504950
#> 82     49 -4.615121 0.08425743
#> 83     49 -4.615121 0.08425743
#> 84     50 -4.615121 0.08415842
#> 85     59 -4.615121 0.09217822
#> 86     41 -4.615121 0.09217822
#> 87     50 -4.615121 0.08415842
#> 88     51 -4.615121 0.08425743
#> 89     47 -4.615121 0.08504950
#> 90     46 -4.615121 0.08574257
#> 91     44 -4.615121 0.08772277
#> 92     58 -4.615121 0.09049505
#> 93     47 -4.615121 0.08504950
#> 94     52 -4.615121 0.08455446
#> 95     58 -4.615121 0.09049505
#> 96     56 -4.615121 0.08772277
#> 97     53 -4.615121 0.08504950
#> 98     55 -4.615121 0.08663366
#> 99     50 -4.615121 0.08415842
#> 100    49 -4.615121 0.08425743