Performance of a uniform forecast

Performance of a uniform forecast

add_uniform_pred(
  test,
  max_score,
  discrete = TRUE,
  include_samples = FALSE,
  n_samples = NULL
)

Arguments

test

Testing dataframe. The only requirements is that it contains a column "Score".

max_score

Maximum value that the score can take

discrete

Whether to estimate a discrete or continuous historical forecast

include_samples

Whether to return samples from the historical forecast in the output

n_samples

If include_samples=TRUE, how many samples to return. When NULL, the samples 0:max_score are returned.

Value

Dataframe test appended by the columns "lpd", "RPS" (or CRPS if discrete=FALSE) and optionally "Samples"

Examples

max_score <- 100
test <- data.frame(Score = rbinom(1e2, max_score, 0.5))
add_uniform_pred(test, max_score)
#>     Score       lpd        RPS
#> 1      45 -4.615121 0.08663366
#> 2      58 -4.615121 0.09049505
#> 3      56 -4.615121 0.08772277
#> 4      43 -4.615121 0.08900990
#> 5      45 -4.615121 0.08663366
#> 6      37 -4.615121 0.10089109
#> 7      47 -4.615121 0.08504950
#> 8      45 -4.615121 0.08663366
#> 9      54 -4.615121 0.08574257
#> 10     56 -4.615121 0.08772277
#> 11     47 -4.615121 0.08504950
#> 12     50 -4.615121 0.08415842
#> 13     47 -4.615121 0.08504950
#> 14     48 -4.615121 0.08455446
#> 15     59 -4.615121 0.09217822
#> 16     51 -4.615121 0.08425743
#> 17     40 -4.615121 0.09405941
#> 18     50 -4.615121 0.08415842
#> 19     42 -4.615121 0.09049505
#> 20     53 -4.615121 0.08504950
#> 21     48 -4.615121 0.08455446
#> 22     47 -4.615121 0.08504950
#> 23     45 -4.615121 0.08663366
#> 24     53 -4.615121 0.08504950
#> 25     52 -4.615121 0.08455446
#> 26     47 -4.615121 0.08504950
#> 27     47 -4.615121 0.08504950
#> 28     50 -4.615121 0.08415842
#> 29     53 -4.615121 0.08504950
#> 30     50 -4.615121 0.08415842
#> 31     43 -4.615121 0.08900990
#> 32     49 -4.615121 0.08425743
#> 33     42 -4.615121 0.09049505
#> 34     54 -4.615121 0.08574257
#> 35     59 -4.615121 0.09217822
#> 36     53 -4.615121 0.08504950
#> 37     53 -4.615121 0.08504950
#> 38     52 -4.615121 0.08455446
#> 39     54 -4.615121 0.08574257
#> 40     44 -4.615121 0.08772277
#> 41     46 -4.615121 0.08574257
#> 42     44 -4.615121 0.08772277
#> 43     54 -4.615121 0.08574257
#> 44     51 -4.615121 0.08425743
#> 45     43 -4.615121 0.08900990
#> 46     58 -4.615121 0.09049505
#> 47     51 -4.615121 0.08425743
#> 48     51 -4.615121 0.08425743
#> 49     52 -4.615121 0.08455446
#> 50     53 -4.615121 0.08504950
#> 51     52 -4.615121 0.08455446
#> 52     56 -4.615121 0.08772277
#> 53     46 -4.615121 0.08574257
#> 54     49 -4.615121 0.08425743
#> 55     57 -4.615121 0.08900990
#> 56     43 -4.615121 0.08900990
#> 57     41 -4.615121 0.09217822
#> 58     38 -4.615121 0.09841584
#> 59     52 -4.615121 0.08455446
#> 60     56 -4.615121 0.08772277
#> 61     48 -4.615121 0.08455446
#> 62     45 -4.615121 0.08663366
#> 63     46 -4.615121 0.08574257
#> 64     45 -4.615121 0.08663366
#> 65     41 -4.615121 0.09217822
#> 66     46 -4.615121 0.08574257
#> 67     53 -4.615121 0.08504950
#> 68     56 -4.615121 0.08772277
#> 69     45 -4.615121 0.08663366
#> 70     46 -4.615121 0.08574257
#> 71     60 -4.615121 0.09405941
#> 72     54 -4.615121 0.08574257
#> 73     49 -4.615121 0.08425743
#> 74     56 -4.615121 0.08772277
#> 75     46 -4.615121 0.08574257
#> 76     58 -4.615121 0.09049505
#> 77     45 -4.615121 0.08663366
#> 78     46 -4.615121 0.08574257
#> 79     46 -4.615121 0.08574257
#> 80     43 -4.615121 0.08900990
#> 81     49 -4.615121 0.08425743
#> 82     51 -4.615121 0.08425743
#> 83     40 -4.615121 0.09405941
#> 84     41 -4.615121 0.09217822
#> 85     48 -4.615121 0.08455446
#> 86     37 -4.615121 0.10089109
#> 87     51 -4.615121 0.08425743
#> 88     67 -4.615121 0.11277228
#> 89     56 -4.615121 0.08772277
#> 90     43 -4.615121 0.08900990
#> 91     55 -4.615121 0.08663366
#> 92     44 -4.615121 0.08772277
#> 93     50 -4.615121 0.08415842
#> 94     48 -4.615121 0.08455446
#> 95     58 -4.615121 0.09049505
#> 96     53 -4.615121 0.08504950
#> 97     50 -4.615121 0.08415842
#> 98     49 -4.615121 0.08425743
#> 99     47 -4.615121 0.08504950
#> 100    41 -4.615121 0.09217822