Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "y_pred", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object. Default to "y_pred" (variable containing predictions).

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  0.41514383 -1.34024929 -2.59869654 -0.97628333  0.28236679  0.23524224
#>   [7]  0.15063578 -0.38619422  0.19856182  1.40432580 -0.78461748  0.46206649
#>  [13] -2.09194579 -0.61317261  1.34472937 -0.02456666  0.92395382 -0.04865559
#>  [19] -0.79333116  0.22068164  2.01505621 -1.27343968 -0.03169485 -2.02681618
#>  [25]  0.86275750 -0.92620852 -0.30986632  0.66563816  0.61329064  0.23450677
#>  [31] -0.31000403  0.66753478  2.37083408  0.62088270  1.07737029  0.26033141
#>  [37] -1.57278540 -1.41657016 -0.12335578 -0.19865750  2.19465422 -1.65927090
#>  [43] -0.14227785  1.24649431  1.24914377  1.19901388 -1.97003801  0.88263948
#>  [49] -0.25571989 -0.30651079  1.44687897 -0.96193761  0.21311399  1.52855203
#>  [55] -0.60589259 -0.24133368  0.96444718 -0.68336321 -0.80629982  0.08577795
#>  [61] -0.28682667 -0.65634903 -0.20042861 -0.94765347 -0.38152911  0.86627168
#>  [67] -0.36332312  1.26316799  0.37165491 -1.96853217 -0.76837335 -0.48915706
#>  [73]  1.48228790 -0.37028382 -0.36125648  0.43072979  0.61269443  1.08173045
#>  [79]  0.51056989  0.32528908  0.78083099 -1.31752932  1.97921839  0.30835299
#>  [85]  0.42854799  0.76502195  1.77353632 -1.59842039 -0.09564733 -0.39334173
#>  [91]  0.96530265 -0.38464179  0.39720843  1.25334220  1.10403784 -0.70802445
#>  [97]  0.52948822 -0.08179137 -0.45365088  0.47868383
#> 
#> [[2]]
#>   [1] -2.3831051559  0.1287591882  0.8740927099  1.7895559286  1.7175729251
#>   [6]  1.3100410731 -0.5927351427 -0.9299461170  0.2531705768  0.5308611654
#>  [11]  0.2059611331 -0.5343626634  0.1507109911 -0.7029577090  0.8628449493
#>  [16]  1.0882708145  0.1174042636 -1.4599633602 -0.0941582051 -0.8927011648
#>  [21] -1.5416126039  0.2187197002 -1.1750208032  0.1428971671  0.4652102346
#>  [26]  0.2284533349  1.5128945445 -0.2334472945  1.3215461154  0.2342592092
#>  [31]  0.0341516826  0.5793838758 -0.6152840153 -0.5635264487 -1.2320297736
#>  [36]  0.5953161889 -0.1565595012  0.6827582933  0.9961368101  1.7192710670
#>  [41]  0.5774346932 -0.1189886656 -0.1523708410  0.0762204648 -0.4582471285
#>  [46]  1.5954511268  2.2106812059 -1.1927067201 -0.7934059354 -0.0141538607
#>  [51]  1.9193248522  0.6935760951 -0.0673451877 -0.7020318067  0.0756368357
#>  [56]  1.1031246494 -0.8070903156 -0.9329960990  0.4835397014 -0.2833028420
#>  [61]  0.1145927127  0.7757413799  0.5899440261  1.3596243876 -0.1659909464
#>  [66]  1.5296599214 -1.1454896486 -1.3552364742  1.2762484934  0.1953863142
#>  [71]  0.0739846228 -1.0873939700 -0.0007610976  0.8893745784 -0.7507067853
#>  [76] -0.2803016623  0.7502042971  1.1535160785 -0.7548944719 -0.7919817557
#>  [81]  0.4998908199  0.4663239797 -0.1898714252 -1.3042936753  0.2882158659
#>  [86]  1.2791282361  0.7965651694 -0.5195783151 -1.2891088161  1.1749660128
#>  [91]  1.1816561359 -0.9177394740  1.2236424386  1.0502270644 -1.9740191677
#>  [96] -0.7785417633  0.7882820855  0.4442546374  0.6325730169  0.0393028261
#> 
#> [[3]]
#>   [1]  0.52850565  1.97224058 -0.37341918  1.16668740  0.17339024 -0.71651911
#>   [7] -0.35911539  0.91527368 -0.77981043 -0.46944458  1.12128851 -0.14456121
#>  [13]  0.53509066  0.32854309  0.85417520  0.16522820 -1.61014950  0.66001803
#>  [19]  0.31779114  0.24310557 -1.55731352  0.19567740  0.48586747 -0.26282909
#>  [25]  0.90544515 -0.13730068  0.02094155 -0.33771260 -0.54588367 -2.60530987
#>  [31] -0.94242946  0.77497483 -1.52327445 -1.05899234 -1.93085933 -0.75183000
#>  [37]  0.81482910  0.18248093  2.28824872 -0.88234580  0.21828716  0.68086889
#>  [43]  0.15864814  0.59447750  0.64707485  1.47730543  2.22321065  0.03820436
#>  [49] -1.14973463 -1.19931221  1.65813026 -1.35891716 -1.03015919  1.35751090
#>  [55]  1.28169696  0.13722338  0.27178472 -0.88453724 -0.55697469 -0.26756785
#>  [61]  1.37531459 -1.28577378  1.51665446 -1.49169693 -0.83172511 -0.08895754
#>  [67]  0.37663128 -0.68855874  0.12809425  0.84133969  0.34143697  0.09943852
#>  [73]  0.98751306 -1.51765287  0.93971529 -0.67822906  0.11593605  0.48568510
#>  [79]  0.26223873 -1.11090018  0.67368740 -0.72225455 -0.20856380 -1.39582756
#>  [85]  0.64332909  0.02656880 -1.79101810 -0.75511817 -0.66127132  0.25834275
#>  [91]  0.66913765 -0.22827177 -0.26650255  1.01965474 -1.45066904 -0.74246123
#>  [97]  1.38749159 -0.42991508  0.83664337  0.46507986
#> 
#> [[4]]
#>   [1] -0.892846953 -0.397009608 -0.543521623  0.688574709  0.542721283
#>   [6] -0.539402317 -0.151403940  0.080572752  0.153087613  0.556410050
#>  [11]  1.308077344  0.494482031 -0.885231892 -0.253943555  0.101297626
#>  [16] -0.908933888  0.008428643  1.456524645  1.279012888 -0.366527699
#>  [21]  0.964101036  0.229504257 -0.365216761  0.683294507 -0.639691673
#>  [26]  1.212487959 -1.044446305  0.436883079 -0.830305951  0.572028020
#>  [31] -0.972521599 -0.089815639 -2.299129245 -0.604593843  1.129165377
#>  [36] -0.001060085 -0.523556730  0.802538322 -0.775397473  0.540146606
#>  [41] -1.279499293 -0.218365663  0.042113784 -2.699645836  0.139001354
#>  [46]  0.877934669  0.052969813  1.471054628  0.126346413  1.085414225
#>  [51]  1.115106048 -2.195799624 -0.964733336  0.358619828  1.034714212
#>  [56]  0.613483495  1.193710031  1.159517193 -0.133518594 -0.951980909
#>  [61] -0.114569964 -0.097971943 -0.828834138 -0.498549269  0.013189019
#>  [66]  0.082415051 -0.176500389  0.709957112  1.250768761  1.278206658
#>  [71] -0.869992388  1.438210470  1.110882700 -0.157845410  1.198899126
#>  [76] -0.595396921 -1.133589746  0.506927896  0.779815617  0.226165684
#>  [81] -0.458163559 -1.540378208  0.534708101 -0.692079665  0.209027612
#>  [86] -0.414229936  0.746951025 -0.745855962  2.389223994 -1.608737438
#>  [91] -0.035569179  0.918656257 -2.283909205  0.641332218 -0.210945408
#>  [96]  0.060718882 -1.024032034 -0.259782173 -0.204017361  0.342324381
#> 
#> [[5]]
#>   [1] -0.51947242  0.87886237 -1.29688459  0.93739238  0.08885764 -1.68279804
#>   [7]  0.37657747  0.40501982 -1.01784975  0.28001047 -1.39846368 -0.52916484
#>  [13]  0.06071811  0.77322846  0.02919030 -2.50469225  1.18521853  0.24943974
#>  [19] -0.66953394 -1.31992626  1.40543370  2.03308452 -0.84725759 -0.44016444
#>  [25] -2.22558029  0.85076048  0.63829875  0.26397153 -0.21769055 -1.16016210
#>  [31] -0.49650127 -3.09531640 -0.41697848 -0.26129262 -0.11409048  0.85058934
#>  [37] -1.44474579 -0.96458719 -0.73999746  0.24245809 -1.44870176  1.31193225
#>  [43] -0.52734366  0.51370629  0.23127996  0.48306784 -0.49533796  1.57872020
#>  [49]  0.33950726 -0.71709273  0.59441488 -0.85165912  1.23884372 -0.44540020
#>  [55] -0.53112252 -0.14810215 -1.31742415  1.01008752  0.94274877 -0.49448518
#>  [61]  1.73957562 -0.51678249  1.69404940 -0.54117971  0.61719804  0.80011207
#>  [67]  0.98415702 -0.57887141 -0.44105205 -1.90867376  1.46376983 -1.13111011
#>  [73]  0.41179754  0.78867886 -1.07938648 -0.98737637 -0.74664547  1.07771782
#>  [79]  0.55813424  0.47183261 -0.33361150  2.09398510  0.46815510  2.30183682
#>  [85]  0.49335309  0.56387657  0.07325628  1.63332562  1.80957990 -0.33717966
#>  [91] -1.65842213 -1.23028527 -0.22713004 -0.59145027 -0.45640049 -0.09573545
#>  [97] -0.12098812 -1.09930143  0.37958014 -0.66108406
#> 
#> [[6]]
#>   [1]  0.26552023 -0.64540685 -1.00917230 -0.15586066  0.02886736 -1.47884750
#>   [7]  1.06697488  0.94766282 -0.18533819 -2.06056333  1.24216509 -2.33285105
#>  [13]  0.71732222 -1.10464816  1.50498613  0.37744063 -1.42071417  0.50920163
#>  [19]  0.85873811  0.99832426  1.54770137 -0.69238677  0.59281437 -0.29873648
#>  [25] -0.40027536 -0.45793860 -1.08514918 -0.74334087 -0.16777695  0.21368326
#>  [31] -0.58230884 -0.80971656 -0.89314067  1.64016999  0.19193454 -0.44026630
#>  [37]  1.32195429 -0.54805214 -0.64064099 -1.67717121  0.56291386 -0.30936298
#>  [43]  0.57257433 -0.31764050 -1.61004468 -1.67907983  0.88608381 -0.17002462
#>  [49] -1.15055095 -2.91895329 -0.46817987 -1.08445051  1.04442678  0.46453272
#>  [55]  0.66589037 -0.50284720  0.38381661  0.52416981 -0.18836447 -0.94044699
#>  [61] -0.45027088 -1.59343709 -0.59265420 -0.18472707  1.20321448 -1.03683855
#>  [67]  0.95575519  1.78941516 -0.46518679 -1.48600521 -0.87563126 -0.33661902
#>  [73] -1.13160498 -1.21568145 -0.25036811 -0.58311346 -1.85540674  0.19303995
#>  [79] -0.22924471  1.54148476 -0.03424732  1.14125433 -0.15706047  0.30714072
#>  [85]  1.28062175 -0.14684103 -1.26666879  1.25901753  1.75083089 -1.17098916
#>  [91]  1.29481387  0.25133195 -0.37592231 -1.80084158 -0.01579668 -0.74655870
#>  [97]  0.48670087  1.14723281 -1.50182040 -0.59884497
#> 
#> [[7]]
#>   [1] -0.13617641  0.51758041  1.40371375 -1.30101562 -0.36629099 -1.21268641
#>   [7]  1.18958880  0.17018575 -1.48093550  1.60452699 -0.49677425  0.39774042
#>  [13]  2.25303981  0.19637827 -0.84901919  1.05604950  1.01409576 -0.02989132
#>  [19]  0.05235161  1.61781436  0.35684490 -0.84883438  1.25787246  0.07440877
#>  [25] -0.88153884 -0.57441843  0.09492219  0.76571125  0.73566119  1.29404827
#>  [31]  0.38649559 -1.99550008  1.11860929 -0.76723639  1.04421832  0.09395924
#>  [37]  0.44755738  0.66153411 -1.88088275 -1.66603333  0.46548942  0.32430281
#>  [43] -1.50407341  1.97380481 -0.69009514 -0.39236167  1.06366018  0.11907472
#>  [49]  0.47828119  1.69367522 -0.48435524  1.12624234  0.21261592 -0.22340873
#>  [55]  0.31586765 -0.66282551 -0.66391878  0.71162200 -0.60943722  0.54862911
#>  [61]  3.16637994  1.38165771 -0.63849931  1.67866811 -0.66670201  0.97823326
#>  [67] -0.74576604 -0.15550780 -0.21254793  0.19039183 -1.29149704 -0.27903484
#>  [73]  0.93072069 -0.40474227  1.34109898  1.77039693 -0.37702150 -1.40087853
#>  [79] -0.49731695 -0.67077510 -2.00156755  1.75014176 -1.86811146 -0.10117619
#>  [85]  0.62969903  0.55674589 -0.60313981 -0.77212716  0.34500362 -0.62077347
#>  [91]  0.76081415 -1.34090047  0.39649279 -0.60250162  0.91801213 -0.58483522
#>  [97]  0.36879521 -0.53838969 -2.11890772  0.39994655
#> 
#> [[8]]
#>   [1]  0.978980864 -1.189140427 -0.742430588 -1.945559976 -1.446260143
#>   [6]  0.051199672 -1.386814888 -1.220064282  1.164904076 -1.218883724
#>  [11]  0.028772758 -0.540015062  0.233761098  0.170664412  1.672346395
#>  [16] -0.313217332 -1.171799482  0.396969597  0.433590605 -0.470764955
#>  [21] -0.176137777 -0.568761394 -1.072380318  0.806541750 -1.368966943
#>  [26]  0.387247278  0.172662550  1.002017543  1.772415662 -0.854054784
#>  [31] -0.738864089 -0.345441596  0.918729223  0.418421102  1.243390264
#>  [36]  1.922869608  0.266883044  1.706477371 -2.103114263 -1.053030226
#>  [41]  1.711242771 -0.506987153 -0.095818096  0.331986158 -1.865074014
#>  [46] -0.866038066 -0.330517254  0.979055035 -0.093392333 -0.863597629
#>  [51] -0.334265884  1.831033469 -1.570379787  0.525941125 -0.118469757
#>  [56]  0.239201543 -0.305564566  1.537913658  1.459931236  0.494137133
#>  [61] -0.665922004 -0.274004776 -1.718374977 -0.373691388  0.993332674
#>  [66]  0.149057342  0.843937519  0.120308010  2.226764014  0.140068077
#>  [71] -0.910110971  0.405100537 -0.631855230 -0.559426158  1.359558452
#>  [76] -0.124197704  1.551036852 -0.125820159 -1.175921363 -0.967876779
#>  [81] -2.141429852 -0.427549913 -1.603802069 -0.264501841 -0.153014717
#>  [86] -0.373726683 -0.397535198  0.901523171 -0.975390506 -1.236012948
#>  [91]  2.039843578  1.823506103 -0.355522387 -0.087607756 -1.730321105
#>  [96] -0.859179745  1.596609026 -0.007032296 -0.338583172 -0.486699303
#> 
#> [[9]]
#>   [1] -1.942054716  0.710464820 -0.094403667 -1.238617055 -1.480702411
#>   [6] -2.352604102  0.827662810 -2.625078866  0.408537075  0.232361614
#>  [11]  0.180359651 -0.956373832  1.125555966  0.247194004 -1.918410473
#>  [16]  3.003107827 -2.826620849 -0.220268334 -0.162688926  0.207319435
#>  [21]  0.205207754  0.799617165 -1.622632379 -0.951515271 -1.623418557
#>  [26]  0.132541874  1.116524516 -0.146823858 -0.407183975 -0.152202303
#>  [31] -1.624897275  2.344822782 -0.968378609 -0.369477826  0.397672994
#>  [36]  0.673962727 -0.945077492 -0.023166056 -0.056982158 -0.815842276
#>  [41] -1.136404954  0.566755781  0.756233115 -1.296670689 -0.332150303
#>  [46]  0.293870145 -0.900746328  1.158519970 -0.449641123 -0.052249276
#>  [51]  0.784770028 -0.470864720 -0.724008274 -0.951025910 -1.065789793
#>  [56]  1.834493388 -0.085704085 -2.104229729  0.008435734  0.162193559
#>  [61]  0.843955718 -1.760727186 -1.549470887 -0.227564666  0.148638020
#>  [66]  0.262264812 -0.451519129 -0.758784489 -0.517313196 -0.300330521
#>  [71] -0.137574590  1.569966281 -1.523522379  0.492276384 -0.652361719
#>  [76] -1.025850298  0.365834004 -0.068356920  1.968844681 -1.443807594
#>  [81] -0.409074059  1.062088512  1.055252141 -1.559952905  0.283359995
#>  [86] -0.231669422 -1.169304574  0.530748773  0.010546112 -0.102817152
#>  [91]  0.616491297 -2.116705248 -0.037320256 -0.206979661  1.275672446
#>  [96] -0.539119794  0.574725853 -0.522952404 -0.235995472 -0.780003814
#> 
#> [[10]]
#>   [1]  1.07785860  0.71343109  1.37374694  1.14991953  0.31028008 -0.27186802
#>   [7] -1.29504000 -0.65928717  1.01289533 -1.04072804 -0.59145025 -0.28065925
#>  [13]  0.10611673  0.36409472  1.63064514 -0.35976874 -1.01894800  1.14607142
#>  [19] -0.41838390 -1.78089824 -0.21436974 -0.85501963 -2.25448868  0.52959249
#>  [25] -0.24245126 -0.23581550 -0.94872193 -0.04296334 -1.04628566 -0.73182029
#>  [31] -3.12212162 -1.22189629 -0.02460784 -0.11292382 -0.89928982  0.45794257
#>  [37] -0.32563398  0.50373777 -1.25099522 -0.71407050 -0.35000908  0.49091232
#>  [43]  0.16870480 -0.48086011  0.11981535  0.57961960  0.01987974 -2.08178311
#>  [49] -0.62685161 -1.31464318 -0.19957157 -1.16661614  0.79441190 -0.19056094
#>  [55] -1.36768734  1.55699210 -2.02435009  0.52349086  0.14368938 -0.94963575
#>  [61] -0.20558221  1.12982725  0.40737348  1.47865990  0.51575386  0.56361041
#>  [67] -1.92119761 -2.82556481 -0.33947080 -0.34296115  1.10436143 -0.34565695
#>  [73]  0.20777304 -0.88270220  1.41562156  0.63777438  0.08632632 -0.19836825
#>  [79] -1.28839826 -0.32292366 -0.34340658 -0.68166786  1.51952606  0.52053299
#>  [85] -1.40660289  1.06383367 -0.76258094 -1.56158469  0.40800687  0.65536249
#>  [91] -0.93564467 -1.29062251 -0.11108104 -0.18558719 -0.38901679  0.96556388
#>  [97] -0.71188712 -0.72463243 -0.03648583 -0.67352426
#>