Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

Usage

samples_to_list(object, par_name = "y_pred", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object. Default to "y_pred" (variable containing predictions).

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  1.192816352 -1.444344362  0.255458729 -0.079880854  0.783372145
#>   [6]  1.107952079 -0.452654070 -1.401415632 -0.509990723  1.562905383
#>  [11] -0.347334211  0.683714301  0.825479660  0.311709224  1.763198672
#>  [16]  0.261834809  0.138689204 -0.345026689  0.737805137  0.330902565
#>  [21]  0.106067409 -0.878674610  1.463621162 -1.121818156  1.205909866
#>  [26]  0.590022411 -0.535741379  1.187797648  0.551681421 -2.138907220
#>  [31] -0.792653848 -0.695657374  1.094752616 -0.100362614  1.374677921
#>  [36]  0.359881635 -0.911636110  0.372154103  0.824123998 -0.277860876
#>  [41]  0.290705252 -1.462234408 -0.936234046 -0.005393725  1.146549255
#>  [46] -0.300266206 -1.672400721  0.342697385  0.004421065 -1.329155221
#>  [51] -0.755702197 -0.401535183 -0.142446497 -0.614792116 -1.294828736
#>  [56] -1.256173491  0.480847546 -0.254530960 -0.698209465 -1.475120921
#>  [61] -0.818271446  2.083201459  0.267543327  0.718185357  0.709885113
#>  [66] -2.386306238  1.404265692 -0.209610722 -0.990249372  1.961966620
#>  [71]  0.594045910 -0.281695649 -0.568785963 -0.065095506 -0.960744079
#>  [76]  1.169699561  0.018659447  1.856655616 -1.113184454 -0.252767978
#>  [81] -1.396072244 -0.619151548  0.263314005 -1.404014797  0.823751531
#>  [86] -0.085016007  0.649380818  1.011451413  1.629933411  0.855285024
#>  [91] -0.301875728 -0.395757012  0.252985427 -0.084844855  0.645852682
#>  [96] -1.389283606  1.483731455 -0.064638732 -0.805267654 -1.934885466
#> 
#> [[2]]
#>   [1]  0.08484413  0.51687622  1.41375161  0.39713983  0.37010424  1.73923589
#>   [7] -0.68153988 -1.11248045 -0.50197427 -0.45545620 -0.95503454 -0.34784849
#>  [13]  0.22522393 -0.59512055  1.36536633  1.74617818 -0.17434665  1.00339677
#>  [19] -0.48560012 -0.52588290 -0.13652250  1.08675927 -0.34185392  1.77385232
#>  [25] -0.17168849 -0.87486044 -0.89005250 -1.11555464 -0.40113332  0.23996381
#>  [31]  0.65854431  0.85166108 -0.62901011 -0.79454969 -0.40371944  1.30141076
#>  [37]  1.20973243  2.28758844  0.70336470 -0.60386586  1.02624779  0.50741909
#>  [43] -0.87608115 -0.71209045 -0.60493192  0.33180791 -0.38419582  1.86682384
#>  [49]  0.05620488 -2.37549193 -0.05972797  0.42387511  1.41780153  1.51697497
#>  [55]  0.95264669 -0.50883594 -0.93151956 -0.89403525  0.70194970 -0.29554239
#>  [61]  0.35112089  1.03655147  0.28401640 -0.49001972  0.51862789  1.84030336
#>  [67]  1.01635979  0.51043708 -1.98510890  2.17144117 -0.62395461  0.93994245
#>  [73] -0.42718539  1.61985553 -0.28783710 -1.09983951  0.30297204 -0.74738746
#>  [79] -0.62417622  0.40981039  0.26364199 -0.32743474 -0.26446903  0.33028409
#>  [85] -0.27123434 -0.55850463 -0.48729364  0.91727594  0.46176051  2.40768979
#>  [91]  0.68675647 -1.25265217  0.12767029 -0.88804576  0.93349971  0.93675224
#>  [97]  0.49051353 -0.07953660  0.74045414  0.24437098
#> 
#> [[3]]
#>   [1] -0.06103008 -1.55866809 -0.18457723 -0.88789202  0.62763447  0.78208595
#>   [7] -1.70301997  1.30401652 -0.94560273 -0.29895192 -0.11270688 -0.50888259
#>  [13] -0.64262955  1.05189685 -0.98010827  0.57785397  0.27472747  0.28393422
#>  [19] -0.79814933 -0.10147950  0.73936886 -1.94578575 -0.44878154  0.19015021
#>  [25] -1.48560480 -1.53219433 -0.55113202 -0.16504810 -0.79828398  0.92665042
#>  [31] -2.11625115 -0.97677426  0.03431268  0.45629474 -0.09319871  0.06120260
#>  [37] -1.22382708 -0.17700586  0.15860971  2.64891611  0.15454608 -0.12561159
#>  [43]  0.69867088 -1.39446424 -0.39846920  0.46239462 -1.17945094  1.39089009
#>  [49] -0.60361621  0.70969501 -0.82249442  0.24869571  0.83113548  0.96840035
#>  [55]  0.25300953 -0.27684488 -1.54461443  0.79494642 -1.40220043 -1.46796683
#>  [61]  0.42130248  0.77903711 -1.34585675  1.31926771  0.36004228 -1.92500383
#>  [67] -0.33658898  1.01111520  0.37400749  0.66263736  0.53660945 -0.07241280
#>  [73]  0.62119347  1.71447523 -0.59025979 -0.70632172 -0.48962004 -0.79409250
#>  [79]  0.41336026  0.44155262 -0.64997739  1.00788314 -0.10213169  0.32117748
#>  [85] -0.39904470 -1.42034114 -0.90655113  0.82785448 -1.31127320  1.06918203
#>  [91]  0.12117447  1.57529572  1.58688605 -1.46125005  0.97518429  1.11455494
#>  [97]  2.61778081 -1.76995515 -3.23740021  0.30700413
#> 
#> [[4]]
#>   [1] -0.174265752  0.657495410 -0.008831523  0.421837595  0.650005772
#>   [6]  0.390177072 -0.768557294  0.856942396 -0.353348987  0.406633202
#>  [11]  0.182891462 -1.876715379 -0.510741518 -0.158000116  1.215151659
#>  [16]  0.042218893  0.397407043  1.290503157  0.866671362  0.616154225
#>  [21] -1.482557543  0.136187368  1.294988672 -1.290650016 -0.977093403
#>  [26]  1.580500626 -0.198718784  0.178222356  0.942686623 -0.723284390
#>  [31] -0.396482359  0.315830272 -1.304693990 -1.809377201 -0.918340185
#>  [36]  0.328061194  0.197851881 -1.143879686 -0.064507457 -0.711701969
#>  [41] -1.948634720  0.305012422 -1.510894658 -0.488629527  1.941073936
#>  [46]  0.088127170 -0.471546085  0.356480834  2.437968092  0.842094303
#>  [51]  0.838628315 -0.177798768 -0.337828787  1.883786434 -0.115525568
#>  [56] -1.072619499  0.173673376  0.601400133  0.096800691 -0.345193170
#>  [61]  0.368975842 -1.539204524 -2.565885119  0.447413479 -0.454570556
#>  [66] -0.060526601 -0.589556414  0.250647396  0.130262580  0.209611195
#>  [71] -1.263310840 -0.663480994 -0.444005565 -0.589947306 -0.509446904
#>  [76] -0.319052393  0.940056781  0.545919126  0.887302535 -0.576802660
#>  [81] -0.652275997  0.586327429 -1.590766030  0.603602495  0.765801044
#>  [86]  0.798276526 -0.092398312 -0.932317441  0.302354993  1.589085197
#>  [91]  0.913859114 -0.476254330 -1.533084543 -1.742228969  0.189036168
#>  [96]  1.209282466 -0.203970419 -1.557476513  0.371621987  0.459539957
#> 
#> [[5]]
#>   [1] -0.86277594  0.67292393 -0.59329735 -0.13489906 -0.04633516  0.35199693
#>   [7]  1.62280594 -0.79464507 -0.23049986 -1.44351177 -0.52425207  1.26628943
#>  [13] -0.71617199  0.14548212  0.96053098 -0.67424691  1.12467803  2.47293735
#>  [19]  1.74020335 -1.53075669  1.41702094 -0.39685031  0.48999401  0.71862163
#>  [25] -1.78788581  1.20675202 -1.57149222  1.10356699  1.72887445 -0.11600011
#>  [31] -0.07319157  0.30214692  0.14651204  1.50528851  0.07467346  0.40495924
#>  [37]  0.13197059  0.79605215  0.35968998  0.71575058  0.11974905 -0.72601810
#>  [43] -0.27974141 -0.57414146  0.52048222 -1.23750182 -0.74541211  1.58557275
#>  [49] -0.35586285 -1.51305956  1.51557414  0.84428005  0.33901072  0.49973128
#>  [55]  0.87800578 -0.82969225  0.43109539 -1.25098050  0.68328345 -0.79931099
#>  [61]  0.40124381 -0.02780726 -1.30972238 -0.70708721 -1.46737704 -0.08421625
#>  [67]  0.05817145 -1.41343575 -0.22945194  0.66127757 -0.97509056 -0.69556901
#>  [73] -1.28258947 -1.30746432 -1.50847617  0.19102599  1.61651866 -0.19085495
#>  [79] -0.03855950  0.56202508  0.64327098  0.30449537 -0.39741918  0.37146595
#>  [85]  0.96195119  0.71031599 -1.14549838 -0.98185265 -0.92117508 -0.06650559
#>  [91] -1.31892585 -0.58514071 -0.26759452  0.23861358 -0.31664116 -1.28955774
#>  [97]  0.90584001  0.02976631 -0.34927184  0.58408005
#> 
#> [[6]]
#>   [1] -0.27064906 -0.59792633  1.11260749 -0.27365534  1.20093166  0.36723092
#>   [7] -0.74724124 -1.41487085 -0.93108864  1.43418487  1.44503438 -1.02784713
#>  [13]  0.27790238 -0.39761916  0.25374031 -1.26047622 -0.11465603  1.39502319
#>  [19] -1.47615111 -0.08441358  0.54650824 -0.82054972  0.52879615 -0.22545194
#>  [25]  0.15845059 -1.75429575 -0.38745576  0.75426858  0.63401004  0.35567270
#>  [31]  0.86643619  0.27126674  0.63315895  0.78528378 -1.39431105 -0.12315526
#>  [37]  1.65625632 -0.58826145  0.72493595 -1.04506180  1.65335447 -1.09639133
#>  [43] -0.51611585 -0.26687012  0.63262447 -0.17403882 -0.39223928 -0.33492180
#>  [49]  0.26835997  0.52749529  0.68708263  0.83221930  0.45659361  0.24648807
#>  [55]  0.32907091  0.29811925 -0.49550869 -0.13439721  0.00860995  0.81829294
#>  [61] -0.89852773 -0.72530289  0.43793313 -0.83728747 -0.80291812 -0.06528758
#>  [67] -1.32739903 -0.69440341 -0.64270406  0.41587004  0.94736761  0.10276019
#>  [73] -0.54905999 -0.92219076  0.20959217  0.11715161  1.34632425 -1.24548202
#>  [79]  0.20879001 -0.54308209  1.39308285  1.05017220 -0.10679694 -0.58353832
#>  [85] -0.31568020  0.47630777 -1.10421098  0.24695358 -0.48491616 -0.61965897
#>  [91] -0.18863221 -0.90766227  0.34789709  0.43510151 -1.31491747 -0.39270947
#>  [97]  0.87004058 -1.09073773 -0.59826094  0.88819035
#> 
#> [[7]]
#>   [1] -1.378913534 -1.651805843 -0.560608081 -2.174049106 -1.061595905
#>   [6]  1.040918777  1.065953354  0.694540276  0.139099757 -0.201092837
#>  [11]  0.033904279  0.363965989  0.124707178 -0.565872772  0.141839251
#>  [16]  0.353873280  0.719574788  1.168043576 -0.863860579  1.091431232
#>  [21]  0.443296812 -2.562523538  2.284566149  0.067855920  0.373212837
#>  [26]  0.121768686  0.582370503  1.082792223  0.720162148  0.086594651
#>  [31]  1.393854437 -0.295764492 -0.511164709  0.371753698  3.077088151
#>  [36] -0.841773477 -0.757351584  0.225515311  0.821011433  0.443006006
#>  [41] -0.796659635 -0.545376677  0.543297629 -1.661003664  0.523849432
#>  [46]  1.330772482  1.611648232  0.072022299 -0.710756752 -0.871306717
#>  [51] -0.307498250 -1.738088974  1.040427979  2.256791945  1.741951478
#>  [56]  0.111568078  0.778078319  1.293373040 -0.291159884  0.531275075
#>  [61]  0.055874846  2.098537411  0.777969786  0.017406662  0.909998561
#>  [66]  0.054319704  1.710472073  0.661729231 -0.610819108 -0.102528388
#>  [71] -0.769892028 -1.250920393  0.569002047 -0.611318615  0.397500945
#>  [76]  0.896814349  0.005403033 -0.850009853  1.241061065  0.364704370
#>  [81] -0.775399745 -0.395000655 -3.262717401 -0.915951797  1.264955073
#>  [86]  0.096894386 -1.509469689  0.063721662 -2.142291259  0.349268423
#>  [91]  1.634436182  2.003644354  0.220709202 -0.765586295 -1.118195566
#>  [96]  1.020963508  0.880848356 -1.647363168  0.240867062 -2.122455318
#> 
#> [[8]]
#>   [1]  0.11313826  0.25818136 -0.57919735 -0.04287906  0.75483207 -0.16765081
#>   [7] -0.75245662 -0.59244911 -1.32180897 -0.45743264  2.90270571  0.61997324
#>  [13] -3.51803913 -0.04368998  0.25764244 -0.22623901  0.40793869  0.07640337
#>  [19]  1.38634507  0.93422264  1.30137733  0.32791961 -1.16004958  0.75403149
#>  [25]  1.09331454  1.22328185 -0.41623610 -0.66327214  0.99175861 -0.18927626
#>  [31] -0.36397755  0.17534740  1.20970475  0.32591187 -0.10945910  0.03397584
#>  [37] -1.21728965  0.51167018  1.17094139  0.34454991  0.54983876 -0.95612763
#>  [43]  0.30716056  0.09151208 -0.93636236  1.90577393  0.33369602 -0.44658884
#>  [49]  0.33714160  1.00830968  2.43336661  0.68526645  0.28961148 -0.13567761
#>  [55] -0.15330647 -1.01651397 -0.06133107 -0.40519219 -0.68370077  0.73367431
#>  [61] -1.68757778  0.33437046 -0.54043706  1.02046305 -1.00433366 -0.11780586
#>  [67]  0.49529099  2.58014889  1.80346755 -1.23211530 -0.06505926 -1.13000150
#>  [73]  0.82501143  0.59267509 -0.66034799  0.05987437 -1.05545080  0.98828970
#>  [79] -1.35048647 -1.12441979  0.22991078 -1.02374175 -1.08769667 -1.34363433
#>  [85] -0.35955168 -1.25745660 -1.15764579 -0.19059760 -0.50379130 -0.46928925
#>  [91]  1.00168370  1.27346352  0.61625003 -0.45335837  0.52291020  0.35340073
#>  [97]  0.22512154  1.64184381 -0.85181344 -1.80005350
#> 
#> [[9]]
#>   [1] -0.343881985 -2.134494444 -0.134482408 -0.266116528  0.321186231
#>   [6]  0.531502473 -0.014906647  1.030454468 -1.753157700 -1.183287727
#>  [11] -1.305423747 -0.520529285 -1.034937208  0.582139467  2.483638919
#>  [16] -0.965248566  0.035295043  0.523856760  0.106144678  0.372260544
#>  [21]  0.841765649  1.860808703 -1.010373717 -0.177632029  2.024361762
#>  [26]  0.634418718  1.158487264  1.290535988 -0.144495121 -0.251715293
#>  [31]  1.248173043  0.143074636 -1.385329844  0.844608897  0.152772403
#>  [36]  0.064325634  0.861263761 -0.273301258  0.560629369  1.290768491
#>  [41]  1.524215763 -0.620363005  0.505385695 -0.782027287 -1.123941254
#>  [46]  0.465798155  0.521638593  0.510090302  0.439940673 -0.111903188
#>  [51] -1.398188537 -0.411521208 -0.142930663  0.794905960 -0.574940440
#>  [56] -0.293892923 -0.290658114  0.231612869 -0.067205397  1.244004813
#>  [61] -0.293926347 -0.073971454  0.799525641 -0.770533606  0.245472769
#>  [66] -1.566192019 -0.822696777 -0.644508315 -0.298573710 -1.037365350
#>  [71]  0.124802637  0.811443555  2.256749672 -0.613394724 -0.071723617
#>  [76] -0.813328645  0.037673482  1.185847732  0.334429189  0.250543996
#>  [81]  0.296057310 -0.509491535  0.201138185  0.006443739  0.355031041
#>  [86] -1.391452022  1.274653131 -1.386808735 -0.603088708  0.338214355
#>  [91] -0.086534307  0.820669672 -0.078786266  0.026965436 -0.570338527
#>  [96] -0.268144920  0.742066313 -0.256792876 -0.920429958 -2.045600298
#> 
#> [[10]]
#>   [1] -1.165839542 -0.838151113  0.714856201 -0.512552266 -1.101897286
#>   [6]  0.531503905 -1.885063653 -0.988782348 -0.202846034 -1.937313888
#>  [11]  0.890362697 -0.112423112 -0.224072409 -1.283825031 -1.342499017
#>  [16] -1.070571511 -1.552867101 -0.493495109  0.238320886  0.027418373
#>  [21]  0.228853710 -0.069811729  0.961724978  0.420007325 -0.384825946
#>  [26] -2.155742917 -1.309645316  0.425808345 -1.527475420 -0.075326634
#>  [31] -1.724481025  0.312383488 -0.557059748 -1.112549886  0.673401399
#>  [36] -0.001044417 -1.149541691  0.964228573  0.405855477 -0.509258781
#>  [41] -0.917171363  0.383263335  1.032982066  0.130521400  0.889833963
#>  [46] -0.672486250  0.074985296  2.301279647  0.032755084 -0.069560798
#>  [51]  0.515621994  1.392966545  1.018563171  0.053010557  0.580274753
#>  [56] -0.594179494 -0.371855338 -1.423035200 -1.140483817  1.268504739
#>  [61] -0.280104390  0.911053990  2.004394540  0.247140421 -0.429076094
#>  [66] -2.183755122 -0.123634896  0.287383879 -0.371599269  1.626562257
#>  [71]  0.309908346 -1.564040664 -1.230623465 -1.370197482  0.106555805
#>  [76] -0.398825492 -0.353406130  0.234501126  1.769789884  0.457876384
#>  [81]  1.143758112 -1.307722621 -0.942549601 -0.094961251 -1.130046974
#>  [86] -1.456208037 -1.698717514 -1.240826691 -1.058299049 -0.754378567
#>  [91] -0.585106942  0.558142081 -0.019006583 -0.169356024  1.295330716
#>  [96] -1.426059996  1.754103727 -2.115575269  1.009702424  0.254535083
#>