Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1] -0.56350768 -0.17059271 -0.61317442 -1.26375923 -0.79911170 -0.80968026
#>   [7] -2.38017254 -0.67047725 -0.02640032 -1.30892545 -1.04407186 -0.19095545
#>  [13]  1.39845671 -0.07421041  1.10987301 -0.19652490 -0.32811827  0.26213590
#>  [19]  1.13901344  0.18547936  0.03727418  0.67088243  1.24829222  0.11899014
#>  [25]  0.67020446 -2.11524830 -0.43425468 -0.53145884  0.76518233 -1.71361819
#>  [31]  0.09937337  1.10516407 -1.46339298  0.11169394 -0.10983580  0.57594173
#>  [37]  0.04007685 -0.91732048 -0.60640848 -0.91075137  2.03384758 -0.09606532
#>  [43]  0.49201977 -1.71249660  2.13797110 -0.94953193  2.20928971 -0.04727157
#>  [49] -1.74868223 -0.58154338 -1.73818549  1.84315636 -0.28994241 -3.53778808
#>  [55] -1.24032464 -0.46559113  0.24580863 -2.09383202  0.23344219 -1.26668711
#>  [61] -0.69217947 -0.11257044  1.19783730 -0.92349100 -0.13287939 -0.86020794
#>  [67]  0.66409375  0.44206416  0.29592919 -0.95405311  0.65131681  1.09065617
#>  [73]  0.82140902  0.48712925  1.09211791  1.14910131 -0.40939738 -0.75941799
#>  [79] -0.45107566 -0.51716539  0.20191329 -1.34561019 -0.48321552 -1.35663668
#>  [85] -0.11139875  0.17086062 -0.70888441 -0.76107547 -0.40255427 -0.12930674
#>  [91]  1.22653820  0.95031039 -0.18852614  0.87896250 -0.71828262 -0.35759570
#>  [97] -0.36857902 -0.22029473  1.16595649  0.84369912
#> 
#> [[2]]
#>   [1]  0.382630231 -1.096738501  1.557084229  1.093587890 -0.834254751
#>   [6]  0.407518389 -0.528476685  1.126920842  1.226376301 -2.364532011
#>  [11] -1.727777994 -2.035793238 -0.002172788  1.593927057 -1.513138765
#>  [16]  0.466139333 -1.142325851 -0.653925372 -1.224295045 -0.122074921
#>  [21]  2.382334193 -1.302067330  1.080914624  0.901309138 -0.343119936
#>  [26] -0.878392281 -0.292415389  1.710173265  1.214667331  1.180556991
#>  [31]  0.226798993  0.512165113  0.260003149  0.831846240  2.362295282
#>  [36]  1.136679098 -0.487784771  1.590751518 -1.637652631  0.491213045
#>  [41] -0.386078207 -0.030990324 -2.014471603  0.522310461 -1.570722758
#>  [46] -0.908171707 -1.586798924 -0.631838988 -0.867056912  0.641544751
#>  [51]  0.646819887 -0.328541877  1.832941196  2.400497589  0.990630121
#>  [56] -0.160185366 -0.069068459  0.446672265  0.641144445 -0.654649170
#>  [61]  0.961186094 -1.214789936  0.272735015 -1.795336281  1.167941418
#>  [66] -0.838224986  0.110626915 -2.198507614  1.134011191  0.777088576
#>  [71] -1.113086985  0.425501666 -1.267235275  0.851095871 -1.047328101
#>  [76]  1.016003969 -0.488470204  0.188641470  0.643194689  1.413609956
#>  [81] -1.790975273 -0.363125119  0.196564873 -0.964504988  0.908751085
#>  [86] -0.435262569  0.315286708 -0.068215887  0.780630215  1.155305565
#>  [91]  1.123298216  0.478244831 -0.681540049  1.434341256  0.679358519
#>  [96] -0.382591822 -0.296234937 -0.089442746 -0.129059943  0.145108654
#> 
#> [[3]]
#>   [1]  1.01166703 -0.12416249  0.42366651 -0.23374249 -0.51110288  0.90387042
#>   [7] -0.41522652 -0.13562725 -0.05993124 -1.82189679  0.10244591 -1.09080493
#>  [13] -0.08435551 -0.30607592 -1.92725115  0.09322267  0.43134004 -1.13670259
#>  [19]  0.23777499 -0.64992002  0.02315174 -1.70076673  1.19402221 -0.08065545
#>  [25]  0.52900630  1.65788600  0.56982552  0.09344591  2.10155847 -2.07247887
#>  [31]  0.39673954 -0.42842357 -0.90834303  0.89629864 -0.17713924  0.55798967
#>  [37] -0.40947470  1.50832712  1.23088647  0.64342957  1.20768313  2.42273857
#>  [43] -1.74015283 -2.12852166  0.46158900 -0.76046550  0.90643135  0.23511323
#>  [49] -0.27472033 -2.83459660  0.65611279  0.99862844  0.75497810  0.56142859
#>  [55]  0.55147360 -0.36379886 -0.38951177 -0.68118177  1.71071740 -1.42702141
#>  [61] -0.77128196  1.72351681 -0.67597513  0.34244993  1.74681683 -0.27535091
#>  [67]  1.65197462  1.22541031  0.44308623  0.14381859 -1.20212761 -0.91771967
#>  [73]  0.84077567 -1.73408963 -0.61757544 -1.32991058 -0.71707915  1.24262524
#>  [79]  0.07306900 -0.75324513  1.30886794 -0.79238175 -1.00115582  0.53158028
#>  [85] -0.50097592 -1.06797210  2.87065259  1.20093268  0.55185797  0.64805338
#>  [91] -0.48994469 -1.06135265 -0.05623516 -0.50436985 -1.11111518 -0.12623296
#>  [97]  0.87939822 -0.77506685 -0.54633047 -1.33599740
#> 
#> [[4]]
#>   [1] -0.70280012 -0.45779691  0.34752887 -0.63278691  1.18470206  0.28014179
#>   [7] -0.94592858 -0.56795647  0.51840139 -0.45730082 -0.01795618 -0.28838349
#>  [13] -1.72591868  1.33386071  0.58591937  0.87873075 -0.05199141  0.93670371
#>  [19] -0.95145118 -0.94739049 -1.29952160 -2.07451222 -0.46494237  0.72259087
#>  [25] -2.52183942  0.19292239 -1.24989433  1.06149572 -0.86490592 -1.23611130
#>  [31]  0.80180894  0.89931331  0.01786426  1.24206028  0.44598969  1.38183645
#>  [37]  0.17388587 -0.68409480 -0.80975171  0.44369073 -0.29791609 -0.78594580
#>  [43]  0.22893638  0.47642893 -1.73604059  0.41024116  0.41634641 -1.26681877
#>  [49] -0.42267233  0.94410245  0.11068594  1.00410873  0.89690253  0.67865818
#>  [55] -0.77001822  0.17422924  1.43362461  0.07578770 -0.08591889  0.36563378
#>  [61] -0.17338576  0.93975931 -1.07740965 -0.11418098  0.72060549  2.29398477
#>  [67] -0.21782749 -0.50630253 -0.49494419  2.56973461  0.06416915  0.65798754
#>  [73]  1.68324925 -0.43190872 -0.35189879 -1.49515882  0.59476607  0.39295201
#>  [79]  0.44737285 -0.36794765 -0.45053265 -0.97145246  0.97836667 -0.61318195
#>  [85]  1.00292851 -1.50194691 -0.78810042 -0.69108978  0.18882440  1.02876338
#>  [91]  0.87213620 -1.00274430  0.27389926  0.07792555  1.78248098 -1.00748833
#>  [97]  0.99167793 -0.63692638 -0.41569376  0.38110035
#> 
#> [[5]]
#>   [1] -2.02487134 -2.32153078  0.29306346 -1.35783862 -0.07488860  0.57426128
#>   [7] -1.29558242  0.44019527 -0.70498298 -0.40965917  0.32943214  0.86388667
#>  [13] -0.58897857 -0.31758688  1.10248970  1.01914234 -1.05223192  0.78343645
#>  [19] -0.51218310 -0.06712900 -0.48438618  2.13457858  1.02560506 -0.07079140
#>  [25]  0.25398743  0.90563074 -2.15146613 -0.06094965  0.35548380  1.09078557
#>  [31]  1.02864391 -1.37146615 -1.38254042  0.51140939 -1.60462274  0.20391701
#>  [37]  1.26071804  0.61959242  0.85322127  1.09136192  0.11156741  0.51322968
#>  [43] -0.67530244  0.51757915 -0.14257655  1.58973353  0.29113753  0.61697587
#>  [49]  0.77176611 -0.05413042  3.74272285 -0.46426954  0.95131907  0.02358690
#>  [55]  0.62350186  0.88153856  0.78990725 -1.17677931  2.47189530  0.15977129
#>  [61] -0.60361347  0.78419356  0.84919433 -0.50127518  0.65207105  0.10940538
#>  [67] -1.05444869  0.82288684  0.39143937  0.11958872 -1.10965950 -0.59000084
#>  [73]  1.25551320 -0.06558315 -1.22676058  1.39313866  0.52067589 -0.71132444
#>  [79]  0.46058273 -0.86165085  1.87228965 -0.96184130 -0.66467994  0.61572369
#>  [85] -1.22791227  1.10139637  0.27627667  0.22855939 -1.55680821 -0.14763295
#>  [91] -0.23912746 -0.51985238 -0.68107241  0.56209741 -0.68114697 -0.10533210
#>  [97] -0.08155702 -2.33194161 -1.87790966  1.90119219
#> 
#> [[6]]
#>   [1] -1.557373020 -0.588086962  0.366153777 -0.644043365 -0.275878342
#>   [6]  0.780136017 -0.714323087 -1.311593482 -1.118743593  0.145085326
#>  [11] -0.058874827  1.425077784 -0.473476894  0.666564066  1.918460555
#>  [16]  0.143975814 -0.015789243  0.695429008 -0.825154251 -1.642799053
#>  [21] -0.205039103  1.501247175 -1.737824072 -1.172937029  0.601288080
#>  [26]  0.299706746  0.301488399  0.969786414  0.077866961  1.222045833
#>  [31] -1.280118292  2.211483108 -0.115426301  0.385482779 -0.092910011
#>  [36] -1.246758514  0.578148684  0.674039214  1.038442446 -1.185287625
#>  [41] -0.374328750 -2.152870089  0.833340551 -0.751328519  0.228770407
#>  [46]  0.453974264  2.165851941 -0.917347397  0.279913372 -0.009635141
#>  [51]  0.827680835 -1.185502934  0.833617441 -0.125581205 -1.330685944
#>  [56]  0.514850990 -0.671104249  0.452359217 -1.358818303  0.513184369
#>  [61] -0.635356215 -0.200764967 -0.642324746 -0.036806465 -1.029663217
#>  [66]  0.251969992  1.502974130 -1.162176032 -1.622839739  0.396704703
#>  [71]  0.191862949 -0.259688955 -0.235021386  1.040312780 -1.393648370
#>  [76]  0.464320668  0.793083518  1.371567209 -0.382194718  0.655925393
#>  [81]  1.103485435  0.968819036  2.502987546 -0.197386577  0.466478101
#>  [86] -0.992040036  0.935401514 -1.765296451  0.464139318 -1.025906814
#>  [91]  0.913219987 -2.190731615  0.577552864  0.605053373 -0.203934806
#>  [96] -1.403545821  1.500881014 -0.156273942 -0.672713849  0.571750884
#> 
#> [[7]]
#>   [1] -0.164244500  0.348684476 -0.356913422 -0.533526212  1.824510760
#>   [6] -0.176357589  1.119573820  1.175599177 -0.531049459  0.856070747
#>  [11]  2.350114536 -0.840450881 -0.121356857  0.109361816  2.803063774
#>  [16] -0.566073510 -2.666937240  1.850049975  1.012799429  0.749740171
#>  [21]  0.022337938 -0.461250313 -1.040526652 -1.791012942 -0.065122392
#>  [26] -0.634121699 -0.487486755 -1.683216670  0.547950045 -0.053859935
#>  [31]  0.450942895  0.336408180  0.725421981 -1.348817226 -0.466537032
#>  [36]  0.837142819  0.716497717  0.105373303  2.104638950 -0.532371070
#>  [41]  0.173178889 -1.164490391 -2.305134151 -0.421269516  1.416929164
#>  [46]  0.508347189 -0.635303088  1.033381176  1.519750253  0.292657014
#>  [51]  0.395233213 -0.413303569  0.147146609  1.451913716  0.694451271
#>  [56] -0.449979447  1.205935616 -0.624355430  0.227021207 -0.938619967
#>  [61] -1.190799586  0.444318440 -0.574303114 -0.150630064 -1.373794976
#>  [66]  1.179723771 -0.357441462 -0.751150897 -2.038685019 -1.042481865
#>  [71] -1.968857713  1.749690004  0.245433077  0.577995608  0.871736221
#>  [76]  0.011823389 -0.522736234 -1.686355101  0.111683786  0.112782259
#>  [81] -0.857820645 -0.490078554 -0.569547854  0.704179397 -1.289726661
#>  [86] -0.756845917 -0.560955417  0.641543145  0.410304788 -1.086749402
#>  [91]  0.355913748  0.394038189 -0.002687561 -1.158426572  1.713057620
#>  [96] -1.391988008 -0.355260257  0.224192380  1.106509735 -0.488059357
#> 
#> [[8]]
#>   [1] -0.277610188  0.534645486  0.646133170  0.704943018  0.354313006
#>   [6] -1.369708939 -0.397585543 -1.024214144 -1.683954799  0.031487504
#>  [11]  0.225992967 -0.664713516 -1.232006233 -0.473923100  1.588926498
#>  [16] -0.148274608  0.363730069  0.733396135 -0.553890447  0.553898667
#>  [21]  1.565468091  0.837778884 -0.385516541  1.236666324 -0.345574082
#>  [26]  0.826489544 -0.874614661  0.313409318  2.013537844  0.372982984
#>  [31]  0.273472729  0.008355106 -2.318256200 -0.441123797 -0.473723253
#>  [36] -0.441017762 -0.211139504 -0.945496351 -0.546705969 -0.047398449
#>  [41] -0.149339712 -0.566768462 -0.455403930  0.193416146  0.123000504
#>  [46]  0.031903095 -0.037952359 -1.158334803  0.614707117  0.868731951
#>  [51]  0.321090247 -0.900182592  0.679364146 -0.053006504 -0.109000452
#>  [56]  0.187102830 -0.753277267 -0.014975639 -0.635721353 -0.987342405
#>  [61]  1.255896960 -0.342474822  1.808246166  0.651184850  0.453567056
#>  [66] -0.313092907  0.031581091 -1.330724012 -0.759242076 -1.662630927
#>  [71]  1.049659194 -0.109503823 -0.860995897 -2.237141760  2.276251383
#>  [76]  1.102761221  0.715196800 -1.071097818 -0.688944545  1.031288197
#>  [81]  2.685861365  0.519619029  0.683022815 -2.576613456  0.446469529
#>  [86]  0.246769031  1.325230881  0.275798203  1.687586029 -0.699515184
#>  [91]  2.052049626  1.756274750 -0.083939537 -1.165949800  1.119928475
#>  [96]  1.621929019 -0.210318581  0.263431896 -1.110236462 -0.504138418
#> 
#> [[9]]
#>   [1] -0.20839858  0.69744745  0.51142517 -1.61230735 -0.71825091  0.67556881
#>   [7]  0.05613887  0.74773355  0.17768041 -0.59712832 -0.82377884 -0.17864988
#>  [13]  0.19427664  0.86685714  0.81258657  2.57526985  2.83505639 -0.27628130
#>  [19] -0.96640495  1.20189832  0.93844880 -0.76597528  0.80788649 -0.57929664
#>  [25] -0.73136469  0.51588126 -1.05726680 -0.44240690 -0.33207315  1.06082418
#>  [31] -1.10486429 -0.57748763 -1.49156122  0.69683599 -0.78386771  1.29390806
#>  [37]  0.49697932 -1.27864490  1.00812719  2.08636045  0.63751137 -0.94870263
#>  [43] -0.17537670  0.39003093  1.40661625  0.11718478  1.33468639  0.23679799
#>  [49] -1.06075472 -1.63218662 -0.23089722 -0.77580276 -0.33093585 -0.31030666
#>  [55] -0.64804632 -0.62506520 -0.38632467 -1.94128418  0.96837459 -0.26162538
#>  [61] -0.12465696 -1.01098046 -1.24231525 -1.02638610 -1.54525276  0.61852174
#>  [67] -0.82992092 -0.71862985  1.12606807 -0.83946587 -0.19575964 -0.21536578
#>  [73]  0.73435879  0.40106081 -0.01059783  0.01438224  0.75804760  0.07242628
#>  [79]  0.58516892  1.02150361  0.09253375 -1.52610690 -1.67134984 -0.22504177
#>  [85] -1.08510598 -0.44309005 -0.35510437 -0.54506146 -2.43531857  1.28112656
#>  [91]  0.46317763 -0.12991962  0.21260789 -0.62383682  0.69197000  0.08287869
#>  [97]  0.12954485  1.23810611  1.07189192  1.99058488
#> 
#> [[10]]
#>   [1] -0.86217141  0.57906325  0.32000965  0.23333914 -0.11186025  1.35870380
#>   [7]  1.66051393  1.33766022 -1.35077263 -0.17409721  1.70136059  0.22903310
#>  [13]  2.28532242  0.71354061  1.41986194 -0.64040357  0.92005827  1.27677730
#>  [19] -1.68333480  0.81425480 -0.23578254  0.47045537  1.27542315  1.34571521
#>  [25] -0.55135719 -0.94332604 -0.60138957 -0.02758950  0.61049149 -0.93047851
#>  [31]  0.62960960 -0.14608400  0.55171130 -0.01175641  1.82136611 -1.10016698
#>  [37] -0.92045547  0.66691091 -1.11701254 -0.98404113  0.18426581  2.46791613
#>  [43]  0.70693976 -0.65355481 -0.08951840  0.87220456  0.25838222  0.65619087
#>  [49] -1.36438500  0.37352482  1.25519338 -1.21027115  1.35933158 -0.02446048
#>  [55] -0.26856401 -0.71376893  0.53361275  0.57317712 -0.22772742 -0.84135237
#>  [61]  0.19508720  0.35906282 -0.04029326 -0.70239634 -0.36251075 -0.49184130
#>  [67]  0.75368523  0.34374544  0.43858698  1.61835435  0.17756335  1.24169142
#>  [73]  2.00144467 -1.10819755 -0.71679949  0.78682305 -0.37060381  0.81498563
#>  [79] -0.66471984 -0.31113010 -0.08431489  0.14649346  0.39304062 -1.01350617
#>  [85] -0.84259139 -0.63777838 -1.46181594 -0.96783799  0.09351567 -0.33346925
#>  [91] -0.47309571 -0.88014819 -0.19060439 -1.46045370 -1.08799886 -0.68858587
#>  [97]  0.81096467 -1.90195212  0.26684297 -0.41278633
#>