Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1] -1.028284466  0.585638745 -0.419581794  0.451862985 -1.501714045
#>   [6]  0.128709120 -1.236040466 -0.502713761 -0.780810019  1.808856121
#>  [11]  1.822198715 -1.377470783  0.610594089 -1.734998375 -0.059973157
#>  [16] -0.169142917  0.808481609  0.938289384 -1.364796083 -0.458194493
#>  [21]  0.830086513 -0.591853568 -1.119097110  1.423535670 -0.459504308
#>  [26]  0.295751436  0.631155810  0.461789548  0.438512683  2.686034628
#>  [31]  0.604657129 -0.627818493  1.312439059 -0.129560444 -0.534806849
#>  [36]  0.395428483 -0.378645674 -0.330436779 -0.840992134  1.216959103
#>  [41]  0.337916392 -0.590788135 -0.248385898 -2.105509965  0.855479084
#>  [46] -0.275285317 -1.123423445 -0.821577268 -1.076536379  0.126864254
#>  [51] -1.192510502  1.291492751  0.379703099 -0.093480836 -2.092872391
#>  [56] -0.242138843 -0.426792691 -0.361489866  0.046787193 -0.019958445
#>  [61] -3.223478558  1.183543223 -0.248520401 -0.136145709  0.153206416
#>  [66] -2.254944100  2.289465508 -0.753508978  0.365550627  0.306373541
#>  [71]  0.066898341 -1.255005762  1.187093509 -0.050996778 -2.329042488
#>  [76]  1.633675069 -0.844084439 -2.058865682  0.932123983 -0.485312449
#>  [81] -0.001453235  0.172899372  0.078475513  0.156210320  0.011756785
#>  [86]  0.021707576  0.608427226  1.101497553  0.439956210 -0.195706990
#>  [91] -0.564696559 -0.631102146  0.979782450  0.215062879 -0.893032909
#>  [96]  0.308966249  0.635544689 -0.173098613  0.564675599 -1.131625115
#> 
#> [[2]]
#>   [1] -0.36853279  1.87521871 -0.84896069 -2.37652324  0.37197046  0.34143463
#>   [7] -0.66746502 -0.55138086 -1.27183771 -1.97269619 -0.99157915 -1.67728394
#>  [13]  1.40235180  0.83143891 -0.18065986 -0.28899664  0.69459253  0.58583281
#>  [19]  1.23356745 -0.19732234  1.00658278 -1.12940542 -1.25274363  0.26997258
#>  [25] -0.08874667  0.81448608  0.91503365  0.39142446 -1.55214875  0.63879997
#>  [31] -0.08481167  2.20979403 -1.16099884  0.52962125  0.31769740 -0.30466430
#>  [37] -1.54422753 -0.75914066  1.08647189 -0.12280148 -0.60232305 -0.51277329
#>  [43]  0.05855222 -0.39982153 -0.36134886  0.65340760  0.07840645  0.03549463
#>  [49]  0.09042770 -0.98497191 -0.71959249  0.42353000  0.39017292 -1.59111622
#>  [55] -1.06151544 -1.44738835  0.91746283 -0.08931946 -0.74233834  0.42786986
#>  [61]  0.18926033 -0.58194314  1.36522507  0.09022922 -0.44253552 -0.88634621
#>  [67]  1.41459739  0.84394831 -0.68197267 -0.47378145  0.26210366  0.43898037
#>  [73] -0.14105877 -0.87191982 -0.98064137  0.28148215  0.13478853  0.02952019
#>  [79] -1.40001448  0.79863941 -0.32405003 -0.09429791  2.03276542 -0.47401536
#>  [85] -1.01173091 -0.75860355 -0.53541837  1.97717100  0.85858606 -0.23474940
#>  [91] -0.89434312 -2.09458469 -0.12448943 -0.22786130  1.37477150  0.35011994
#>  [97]  1.20696966  1.52413721 -0.15145979  1.51298069
#> 
#> [[3]]
#>   [1] -1.291837588 -0.611089543 -0.151381026  1.388650155 -1.002702978
#>   [6] -1.222269873  0.650226276  0.505097247  0.045262491 -0.497723032
#>  [11] -0.541435322  2.038805460  0.769848717  1.079456377  0.347117549
#>  [16]  1.293502225  0.456225943 -0.802932022  0.303316358  0.298171737
#>  [21] -0.293542576 -0.298729994  0.012954532  0.091736909 -1.050360902
#>  [26]  0.167284118 -1.427263778 -1.054956844 -0.762106614  0.843082634
#>  [31]  0.003109390 -0.130202409  0.545596969  0.640751290  0.006524336
#>  [36]  1.057408767 -1.424126908 -0.145141004 -0.947943187 -0.906582481
#>  [41]  0.013921296  0.289570882  0.894053247  1.349781612  0.943208820
#>  [46] -0.348580621  1.150581874  1.261447019 -1.101851166 -1.812449810
#>  [51]  0.161987659 -0.037501377  0.461706457 -0.751406205  0.602832591
#>  [56]  0.208272524  0.905750851 -0.073556120  1.255431991 -0.101789078
#>  [61] -0.609604452 -1.133853090  0.734604223  0.757690012  1.236839911
#>  [66]  0.684996823  0.073356146  2.056069581  0.049045748  0.096718067
#>  [71]  1.118085189  1.081977670  0.951201522 -0.855181714  0.142130426
#>  [76] -1.637823674  0.514430727 -0.631916768  1.422278231  0.898005396
#>  [81] -0.830303430 -1.309099656  1.221192312 -0.535907353  0.880973861
#>  [86] -2.331646020 -0.645199508  1.979188752 -0.554503999  0.217884215
#>  [91] -0.158515984 -0.710830541 -2.762551884 -0.378698986  0.189807282
#>  [96] -0.569739378 -0.328109082  2.952304311  0.512916249  0.206349759
#> 
#> [[4]]
#>   [1]  1.06033283 -1.33487204  1.01734918 -0.19836326  0.91336726 -0.77798291
#>   [7] -2.05104688  0.20947234 -1.29872752  1.10083486  0.20015703  2.09875375
#>  [13] -0.47472553 -2.54228403 -1.84556635 -0.24595927  1.32576114  2.04830655
#>  [19]  0.22701162  1.25178705  0.30717309 -0.02672896  1.27110671 -0.63890525
#>  [25] -0.06209467 -2.15942515  0.17692565  0.65735010  0.47815934 -0.94462424
#>  [31] -0.23921322 -0.89529822  0.62555351  0.78182947  0.91514698  0.99705384
#>  [37]  1.51091692  1.27578081  0.71995487 -0.74152143 -0.04414173 -1.51212404
#>  [43]  0.45803014 -0.32594877 -0.80396913 -0.47648992  0.15944446 -0.22587908
#>  [49]  0.61514368  0.89136458 -0.31305383 -0.58457208  1.66746505 -1.22890971
#>  [55] -1.00855095 -1.91367157 -0.06943616 -0.84898687 -0.22496534  0.80595054
#>  [61] -0.17253713 -1.34829162 -0.03980239 -0.36029766  0.71139221  0.02637298
#>  [67]  0.68337682  0.62675449 -0.94539795  1.03214863  1.05058744 -0.65975324
#>  [73] -0.41588824 -1.53635700 -0.11670931 -1.34868825  0.70168145  2.46294844
#>  [79] -0.04089963 -0.87985512 -0.81834354  0.26751889 -1.16815037  0.79600682
#>  [85] -2.31555516  0.22311138 -1.20734134 -0.32168142 -2.58362148  0.69909389
#>  [91]  0.05342671  1.18922989  0.13284760  1.04762322 -0.76738916 -1.02618800
#>  [97]  0.07637152 -0.70445101 -0.14335604 -0.20675307
#> 
#> [[5]]
#>   [1] -1.151604257  0.708170227 -0.280258694  1.305381976  0.553293451
#>   [6] -0.402800473 -1.593384292 -1.087980376  0.793092851  0.658346461
#>  [11] -1.177552958  2.092958951 -0.649591804 -0.092422386 -1.036378022
#>  [16]  0.209197383 -0.216506508  0.270036902  1.005291088 -0.218395462
#>  [21] -0.560302647  1.649822544 -0.481231101  1.609550912 -1.152614040
#>  [26] -0.423526648 -0.065251004 -0.489636816  0.588095561  0.506898384
#>  [31] -0.769600892  2.828815061 -1.707784213 -1.741539522 -1.730657643
#>  [36] -1.697836318 -0.774043200  2.496604545  0.133006871 -0.114941126
#>  [41]  0.638842345  1.272333944 -0.415089687 -0.188338348 -0.870728440
#>  [46] -1.212273505  0.401225665 -0.436729066  1.893789266 -1.342122992
#>  [51]  0.045683693 -0.206498357  0.569122777 -0.447253129 -0.711949426
#>  [56] -0.125387524 -1.718721375 -0.049256623 -0.236613054 -0.002087148
#>  [61]  0.252346815 -0.574867450  1.864767566 -0.307197569 -0.083404927
#>  [66]  0.897132680 -0.557139652 -1.183745940 -0.311333877 -0.502869407
#>  [71]  0.190956038  0.587212104 -0.706725656 -0.269776847 -0.134582121
#>  [76] -1.891408864 -1.227314047 -0.644067672  1.062077848  0.439403422
#>  [81] -1.855006275  0.244969363  0.436222004  1.234471618  1.325763950
#>  [86]  0.953563363 -0.774062944  0.014370832  2.001183697  0.080253042
#>  [91]  0.170950972  1.041249707 -0.871969119 -0.162598644 -1.266589892
#>  [96] -0.756374559 -1.438065641 -1.142793522  0.250070161 -1.649097274
#> 
#> [[6]]
#>   [1] -1.106270954  1.550994737 -0.742084765 -1.452580207  0.551243684
#>   [6] -0.220171295 -1.360179806 -0.089236185 -0.612485009  1.849970321
#>  [11] -0.227150595 -0.260140264  0.677197721 -0.791530665  0.940657994
#>  [16]  0.265856824  0.942532207  0.357894152  0.558988950 -0.534903678
#>  [21]  1.066792457  0.155130142  0.385553024  1.429620190  0.520040980
#>  [26] -0.055599218  1.251725974 -1.801735195  1.114271250 -1.804429143
#>  [31] -1.072896649  0.348548503  0.412109168 -0.802465109  0.022391946
#>  [36]  1.510681254  0.671029690 -0.897623645  0.874427955  3.415137253
#>  [41] -1.553457374 -0.463648870 -0.674508293 -0.572157761  0.131567536
#>  [46] -0.316783173  2.156731122 -0.392391342  0.341761864  0.161721662
#>  [51] -0.839515283 -0.096126556  0.178777608  1.129365002  0.005176122
#>  [56]  0.758600119 -0.439821935  1.627260873 -0.914921512 -0.703090556
#>  [61] -0.808551929  1.178523448 -1.423701966  0.761659091 -0.307983636
#>  [66]  0.202675892 -0.504530617  0.667480744  0.485692275  0.554905531
#>  [71] -0.924343809 -0.197192577 -0.253005079 -0.600180759 -0.297736686
#>  [76]  1.805412288 -0.365264182 -1.117765555  1.489916432 -0.346021893
#>  [81]  1.404644115  0.876269187  1.107180230  1.130149299  1.989066600
#>  [86]  0.296405468 -2.291062808 -0.249545061 -0.506553921  0.526416827
#>  [91]  0.607699134 -0.315356187 -0.538862580 -1.196290626  0.038601906
#>  [96]  0.022524674 -3.300595930  0.594161372 -0.817362276  0.369803202
#> 
#> [[7]]
#>   [1]  0.329260564 -0.086425975  0.457800764 -0.257553209 -0.030335680
#>   [6]  1.208372477 -0.000394582  0.341308679  0.536008065  1.037970751
#>  [11]  0.735002616  0.932772131  0.298978850  0.189748758 -1.069314397
#>  [16] -1.074426744 -0.366738539 -1.149489263 -0.664000886  0.036981657
#>  [21]  1.161472128  0.161355931  0.019498976 -0.459310406  0.118153603
#>  [26] -0.132559416 -0.547626166  0.362396202  0.683306343  2.338814560
#>  [31] -1.037832125  0.307506051  0.200948543 -0.175864316  0.166015663
#>  [36]  1.896478300  0.138082785  0.973845114 -1.348731860  0.631337764
#>  [41]  0.991163004  0.110168231 -0.650517519 -0.234114354  0.047962388
#>  [46] -0.513913617  0.732184440  1.001846347 -0.570062964  1.591703036
#>  [51]  0.919350111  0.464807534  0.467070565 -2.624037756 -1.181440476
#>  [56]  0.892901930  0.310042010 -1.948297974  0.509679135 -0.273259828
#>  [61]  1.687014154 -0.169163550 -1.900960274  0.254548498 -1.191409620
#>  [66]  0.575185546  0.221210219  1.121958403  0.652381125 -0.306134293
#>  [71]  0.024776572 -0.144102729  0.091792751  0.196191457 -0.293483413
#>  [76] -1.279306366  0.269541680  0.655778602  0.302800763  1.479309772
#>  [81] -0.090568333  1.555980894  0.035682800  1.436662497 -1.526601015
#>  [86]  0.922508135  0.060221889  1.627071096 -0.458625858 -0.915297062
#>  [91]  1.429917025 -1.388166137  0.074898496  1.183778126 -1.345644630
#>  [96]  1.498316601 -0.808775474 -0.166497634  1.258962546 -0.377640728
#> 
#> [[8]]
#>   [1]  1.024282758  1.736064927 -0.266311795  0.818869616  0.136016267
#>   [6] -0.071083626 -0.935982044  0.499335085 -1.191369394 -0.045683264
#>  [11]  0.224180046 -0.320645444  0.653975824 -0.360011830 -0.851473870
#>  [16] -2.366038483  0.437102602 -0.891341294 -0.940129225 -0.125606534
#>  [21]  0.175482266  0.031893526 -0.600894308 -1.008570028  1.847231022
#>  [26] -0.643932985 -0.315892469  0.037619246 -0.907201400  0.993375287
#>  [31] -1.488989289 -0.864183559 -0.625757693  1.073109097  1.162512196
#>  [36]  0.475066222 -1.374293527 -0.485304748 -0.886550492  0.467453809
#>  [41]  0.621157324 -1.204991809  0.324479785  0.241342739  0.755903597
#>  [46] -0.679376550  1.630164785  1.103879349  1.291638591 -0.072713912
#>  [51] -1.632654568  0.491613266 -0.447971203  1.184126007  0.757020807
#>  [56]  0.224160094 -1.511549155 -1.792737685  0.827528015 -0.235798704
#>  [61] -1.380156752 -0.535104836  1.342699976 -1.165791399  0.607159360
#>  [66]  1.383951906 -1.832985938  0.034163558 -0.040395202 -1.359452385
#>  [71]  1.371185481  0.248941669 -0.839869136  0.011572830  0.038342845
#>  [76]  2.005307266  0.334102570 -0.430108376 -0.639962217  1.417569646
#>  [81]  0.224910382 -1.894821719 -0.464632740 -1.482164434  2.192990087
#>  [86] -0.664408809  0.119799310 -0.420161218  1.928715169 -1.418325170
#>  [91]  1.259401699  1.556357015 -1.046981618 -0.218401461 -1.135473601
#>  [96] -0.790455615  0.008433244  0.942745921  0.163692198  1.015252380
#> 
#> [[9]]
#>   [1] -0.2639531414  0.0110390695  1.0181619887 -0.9693731838  1.2626733034
#>   [6]  0.3670758626  0.5379812685  0.7777938694 -1.1188001118 -0.4992250178
#>  [11] -1.9861983996  0.3740509750 -0.7614395517  0.2716926956  1.0423702048
#>  [16] -0.2534359406 -0.2567079493 -1.7841150868  0.6919899454  0.8671382088
#>  [21] -0.8564276374  0.6586286832  2.1948205165 -0.8789277584 -1.4573305150
#>  [26] -0.6634357977  0.1378635354  0.0168684594  0.6648409531  0.1750162881
#>  [31] -0.3838949416 -0.7579765207 -0.3588123929 -0.0004245418 -0.8528589839
#>  [36]  0.4052598753 -0.6149850157  0.6123743802  0.1369482059 -0.4579617911
#>  [41]  0.1318715008  1.0417891262  0.9962807166  0.9086176021  0.7698180002
#>  [46] -1.0754252996 -0.1873658441 -0.1066683264 -0.4571508583  2.0752628846
#>  [51]  0.0765057667  1.8368897736 -0.5046697373  0.3959083291  1.7569706380
#>  [56]  1.6071764074 -0.4286863535 -2.5651147075 -1.0250660764 -0.2530494486
#>  [61]  1.6289202151 -0.8625706022 -0.3126037792  1.6364249599  0.3545939029
#>  [66]  1.2185878165  0.0082964182  1.3461557052  0.8317378833  0.1140011648
#>  [71]  0.0917588236 -1.3034739587 -0.5676386475  1.5412859842  0.7458229162
#>  [76]  1.4506950960  2.3902879091 -0.5849876453  0.1371259997  0.9124747475
#>  [81]  1.0531109505  0.7188955001  0.2058060185  1.1783403507  0.2737009836
#>  [86]  0.0494615581  1.0153831550  0.9576409275  1.2527093371 -3.0218297922
#>  [91] -1.9740747140  0.2169817088 -0.2394868419 -0.4702759493  0.5933217787
#>  [96] -1.5550529720  0.8690517012  0.6635587420  0.6366256747 -1.6085329844
#> 
#> [[10]]
#>   [1] -1.12371872 -1.68842559  2.53662565 -0.90909483  0.02716763 -0.28806907
#>   [7] -1.86824377  0.50458264  1.52091832  0.25611168  0.40672747  0.23752843
#>  [13] -0.17515174  0.66877230 -0.35533342 -0.12721510 -0.46208012  0.73000116
#>  [19] -0.84363613 -0.72207462  1.57136279  0.23369637 -1.25670190 -0.03989669
#>  [25]  0.98062322  1.09342961 -0.95811634 -0.56376629  1.17111949  0.37154217
#>  [31] -0.48667779 -0.10750916 -0.55352755  0.25857917 -0.29670241 -0.67739711
#>  [37] -1.30126954  0.71109901 -2.17923123  0.68836999 -0.04818220 -0.41176129
#>  [43] -0.56557775  1.26885973  0.27180349  0.49220628 -1.26938386  0.25737926
#>  [49] -0.20740774  0.59778796  1.12099763  0.09502236 -0.30087125  0.82729275
#>  [55]  0.16647305 -0.83729464 -0.25539558 -1.04578679 -0.28590702  1.75535463
#>  [61]  0.44139438  1.05704421  0.61975289  0.38325292  0.86274656  1.21854864
#>  [67]  1.78432638 -0.87997060  0.28251729  0.26923720 -1.07472742  1.29298350
#>  [73]  2.08164001 -0.19314709  0.13466293 -0.76367151  0.95863578  0.88981530
#>  [79]  1.87378949 -0.27252976 -0.48934160 -0.16301565  0.60900952  1.50936517
#>  [85]  2.51564645 -0.01132201  1.24475629  1.16392437  0.51834207 -1.27146247
#>  [91]  0.77624938  1.22575791  0.23236600  0.82073441 -1.07618474  0.81230210
#>  [97] -1.06779841 -0.45796917 -0.98946622  0.28574812
#>