Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

Usage

samples_to_list(object, par_name = "y_pred", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object. Default to "y_pred" (variable containing predictions).

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1] -0.53765056 -0.45242607 -0.43929097 -0.61622311  0.44163455  0.48259745
#>   [7]  0.54214438 -2.29078465  0.31035185  1.40407898  1.37711764  1.06039565
#>  [13]  0.63217289  1.08492804  1.35645944  0.36242404  2.16934446  0.13913051
#>  [19]  1.37632653 -0.49144999  1.53491576 -0.41619872 -0.52054380  0.85058387
#>  [25]  0.33449657 -0.82935177 -0.21869594 -1.54508372  0.23322978  0.03106964
#>  [31]  0.35786565  1.60862422  1.42985426 -0.94833964  1.01570554  0.03773461
#>  [37] -1.74423940 -0.97007381 -0.22902736 -0.97484568  2.69237242  1.39239386
#>  [43]  1.36007615 -0.36541588 -0.86893720 -0.50655673  1.21470730  0.50695868
#>  [49] -2.09497131  0.03407924  0.85272870  0.74322814  0.55715361 -1.24858322
#>  [55] -0.20787431  0.42396946 -0.50669744 -0.61152331  2.41165945 -0.16495814
#>  [61] -0.44040605  0.52197617 -1.91832225 -1.98264996  0.52120016  0.77778320
#>  [67] -0.81211887  0.91074832  0.94875395 -1.34804945  0.35417586  0.53026393
#>  [73] -0.31097493 -0.24415553 -0.29187819 -1.12808616 -1.12288159  2.05393864
#>  [79] -0.90951843  0.45837916 -0.04661845 -0.68095460 -1.48882940 -0.39251301
#>  [85] -1.74968202 -0.03866763 -0.79715324 -0.91592054  0.07326746  1.23817132
#>  [91]  0.71837134  0.53946650 -1.06123399 -0.54073787 -0.71520471 -0.17892009
#>  [97]  0.16497028 -0.72740817 -0.53783713 -0.59966490
#> 
#> [[2]]
#>   [1]  1.18208775  0.18921288  0.24894911  1.04664341 -1.28930094  0.37511557
#>   [7] -0.55691839  0.30242663  0.22226647 -0.96216696  0.05203237  0.90004038
#>  [13] -0.39490088  0.95414600 -0.58821435 -1.47658453 -1.37777577 -1.34567231
#>  [19] -0.73663796 -0.47011150  1.38060934  1.67501093  1.17690663 -0.14889834
#>  [25] -0.17782336  0.80312186 -0.41579535  1.21248959  1.24036003  0.68567588
#>  [31] -0.02679868  0.30958051  0.24986433 -1.35646087  0.59937977  0.00864779
#>  [37]  0.09071661 -0.65705741 -0.48178225  0.01775788 -0.85953576  1.35770498
#>  [43]  1.21506118  1.45391292 -0.08591198 -0.61756875 -0.21826025 -1.32600521
#>  [49] -2.36220890 -1.40996955  0.25439714  0.29587030  0.05678979 -0.10265486
#>  [55]  1.95192303  0.87856596 -1.30661693  0.09034427  0.90235804  0.54942132
#>  [61] -0.86983920 -0.03921465 -0.51845168 -0.91184962  0.15031364  0.44035545
#>  [67]  1.31978206  0.24656113  0.94182660 -0.34215133 -0.27614681  0.38365678
#>  [73]  1.44691022  1.73390029  0.45642938  0.70750349  2.06378922  0.02391059
#>  [79]  0.24594946  0.27205914 -0.99245858 -0.02757795  2.22284516  0.15550390
#>  [85] -0.86911632 -1.17448945 -1.75967731  0.05836159  1.16451986  0.33762789
#>  [91] -1.05451872  0.66076911 -0.82340900  0.43703366  0.37238811 -1.64674913
#>  [97] -1.92355227  0.38083303  1.37570535  0.82584873
#> 
#> [[3]]
#>   [1] -0.41611500  0.98214959  0.24218073 -0.93792685  1.50585850 -1.06585117
#>   [7]  0.16227053  0.26047039 -1.48820199  1.41518387  0.56250751  0.64205574
#>  [13] -0.89053264 -0.11653752 -0.94197475  1.11582792  0.47013513  0.86061271
#>  [19] -0.07039665 -0.61318021 -0.33671496 -0.21584018  0.62113229 -1.28402652
#>  [25] -1.30009244 -0.37676946  0.10374865 -0.70356227  1.49741394 -0.30282688
#>  [31] -1.37683133  0.88317017  0.67142028  1.23002238  1.64546788 -0.18292604
#>  [37] -1.37925781  1.25060060 -1.87623677 -0.36878757 -2.12929731  0.33764603
#>  [43] -1.57457485 -0.19631941  0.91506011  1.07863654  1.45280285 -1.65328109
#>  [49] -2.76186752  0.26436916  1.02457159 -0.65918555 -0.38630438 -0.24249895
#>  [55]  0.22059324 -1.48109003  0.72469120 -2.37751470 -0.06540921 -0.46970481
#>  [61] -0.28051629  0.57483659 -0.20567685  0.23488051  0.11062832  0.27420223
#>  [67]  1.57491963  0.56776783 -0.05961225 -0.14958771  0.26616879  0.52840324
#>  [73] -0.57938809  0.94219937 -0.70636136 -1.27540210  0.73325992 -0.42020709
#>  [79] -0.25537509 -0.36923822  1.21244833 -0.44536301  1.13866752  0.19052348
#>  [85] -1.36639361 -1.29810260 -2.13709831 -0.09620553  0.69450905 -0.47459123
#>  [91]  1.76313061  2.02584606  0.42761881 -0.56647935 -0.83383531  0.39825915
#>  [97] -1.47453793 -0.39544120 -0.13070247 -1.99949395
#> 
#> [[4]]
#>   [1] -0.99593086  0.42011400 -0.06286163 -0.08013185 -0.03228341 -0.71898093
#>   [7] -1.11656132 -0.78026990 -1.77695853 -0.42783487 -2.03102701  2.75076475
#>  [13]  1.51366972  0.03398894  0.74079278 -1.27080818 -0.16356423 -0.61094549
#>  [19]  1.25188419  0.25031947 -1.70558168 -0.85541313 -0.14490163 -0.32444696
#>  [25] -0.17256490 -1.23606292 -1.90230421 -0.09450402  0.03255579  0.46129012
#>  [31]  1.38140030 -0.41647627  0.68094267 -0.41437304 -0.51834551 -0.68401973
#>  [37] -0.88564860  0.04923707  0.18556122 -0.60865779 -0.73110285  2.71514421
#>  [43] -1.33938704 -0.64601525 -0.93245461 -0.76908693  0.37157978  0.35543278
#>  [49] -0.98399849  0.21472959 -0.08008508 -1.42236955  1.10814896  1.07847764
#>  [55] -0.44025034 -0.77816901 -1.81859001 -1.12408090  1.06052384 -1.47870016
#>  [61] -1.55156017  0.77750668  1.06844014 -0.18358770  1.55824293 -0.21324238
#>  [67]  0.93053526  0.41081180 -1.27984430 -0.78236663 -2.27608345 -0.12639591
#>  [73]  1.44814671 -1.44275737  1.46718718 -0.74329990 -0.30422384  0.33765806
#>  [79] -0.60750209 -0.29556027 -0.13453714  0.81478437 -0.27292173  2.15948580
#>  [85]  1.09173757  0.74338485 -1.20785935  0.32781805 -0.53416511  1.28394672
#>  [91]  0.02893134 -0.39567707 -0.69486833 -1.49208070  1.44425724 -0.34701739
#>  [97] -0.04025917  1.24663211 -1.34630152 -0.57390128
#> 
#> [[5]]
#>   [1] -0.70595925 -1.85740454 -0.24697048 -0.33556093 -0.24937687  0.45952256
#>   [7] -0.46007548  0.05893280 -0.74259108 -2.16357400 -0.23289649  0.87490683
#>  [13] -0.97349403 -0.51274686 -0.92348799 -1.49442573 -0.16593036  1.58223832
#>  [19]  1.01460227 -1.14647834 -0.90984911  3.10918268 -1.06955277 -0.70010952
#>  [25] -0.19844447  0.74362012 -0.05015977  1.31314355  0.08601352 -0.62985137
#>  [31]  0.65507033  0.01939355  0.69626996  2.00886989 -0.05697828  0.24041550
#>  [37] -0.02189017 -0.83080816  1.40247432 -0.72386911  1.33058037 -0.81133302
#>  [43]  1.79503611  0.68662633  0.08937800  0.32454774  0.07131812  0.23304838
#>  [49]  1.98637658 -0.95167079 -0.55864232  0.25581454 -0.19807617  0.42878914
#>  [55]  0.07077011 -1.53665380  0.45398437  0.87439857 -1.29773757 -1.07896524
#>  [61] -1.13484994  0.02261431 -0.02146122 -0.27877244 -0.02420641  0.82502030
#>  [67] -0.75018278  0.53621047 -1.57271748 -0.98642742  1.98351232 -1.85137844
#>  [73] -0.90978818 -1.95144002 -0.80027646 -1.86936160 -0.75057148 -0.59093839
#>  [79] -0.74209927  0.69351208 -0.05946397 -1.86389510 -1.27450892 -1.78177081
#>  [85] -0.50859339 -1.73687829  0.04048130 -0.12417946 -0.61275343  0.16066248
#>  [91] -0.66235963 -0.33485166  0.62301159  1.02819429 -1.13457824  0.91691105
#>  [97]  1.21387479 -0.68593246 -1.09456819  0.36487410
#> 
#> [[6]]
#>   [1] -1.00801708  0.55499644  0.43606592  0.10537330 -0.25373785 -0.97649123
#>   [7]  0.39421777  1.01490303  0.63144381 -0.48339825 -1.52633394  0.61684228
#>  [13]  1.22928684  0.15745303  1.40634206  0.39680092  0.32642296 -1.45934331
#>  [19] -0.79929553  0.47001280 -1.15018052 -0.27159672  0.45742422 -0.01695794
#>  [25] -0.54159436  0.87581342  0.74083815 -0.16481239 -0.75096428 -1.25336629
#>  [31] -1.03623626 -0.02943857 -0.27494378  0.48545147  0.91698203 -0.98534392
#>  [37] -1.49118975  0.81959245  1.01340453  1.05360525 -0.07142864  0.95244461
#>  [43] -0.38403752 -0.41775999 -0.47026817 -1.69018029 -0.94549766 -0.34301196
#>  [49]  0.35846241  0.04818886  1.13017228 -0.52763231  0.47957533  1.41752930
#>  [55] -0.08009077 -0.39418700  0.52525764 -0.79593510  0.24509645 -0.77931984
#>  [61]  0.59492387  1.10889090 -0.94296903  0.70034399 -0.42467483 -1.14313760
#>  [67]  0.23344022 -0.18923658 -1.66341126  1.91569077 -0.81689444  0.38365049
#>  [73] -0.45863225 -0.71570398  0.42954854 -1.16985274  1.32208831  1.36596464
#>  [79] -0.24972371 -0.42472135  1.09843934  0.67092304  0.03095466  2.36851196
#>  [85]  2.58578811  1.21228095 -1.25119684  0.40026604  0.93470763 -0.47266985
#>  [91]  1.58301322  0.59775422  0.46608459 -0.74436754  0.95905526  0.15263708
#>  [97] -0.76936870  1.39308074  0.91866877  1.01608706
#> 
#> [[7]]
#>   [1]  0.53463544 -0.98762856  1.18779227 -0.51754221 -0.25956025 -0.32806467
#>   [7]  0.07343239 -0.24786302 -1.37386226 -0.04044582  0.42153824  0.20159751
#>  [13] -1.69719192  0.64228768 -0.99523961  0.96381390 -1.65603723  1.07086109
#>  [19] -0.10902636  1.89918639 -1.13703073 -0.27971976 -0.89412905  0.13670185
#>  [25] -0.74916542  0.51819908 -0.19233721  0.02880981  0.35859089 -0.02899503
#>  [31]  1.14704207  0.37358894  0.32333921 -0.82981932  1.39446258 -0.19154358
#>  [37]  0.27227702 -1.08165901 -2.32939936 -0.54962596 -0.07257999  1.03228840
#>  [43]  0.21513853 -0.49434012  1.51213828 -0.57946326  1.67478963 -1.00098026
#>  [49]  1.22270284  1.07718817 -0.61194546  0.50686697  0.46006837  1.48439186
#>  [55]  0.88196323 -0.53670224  1.28555371  0.58784853 -1.30848203  0.31672632
#>  [61]  1.19415830  0.91305715 -0.78675299 -0.41092970  0.47637368  0.18591587
#>  [67] -1.32471625  1.17237106  0.23149277  0.48307190 -0.53531958  1.37813582
#>  [73] -1.30267167  0.63485754  0.99965516 -0.33774895 -0.08603361 -1.71881758
#>  [79] -0.92912157  0.81367605  0.52641355  1.01195739  0.83137981  0.41513414
#>  [85] -0.66125706  0.54881129 -0.53936780 -0.17131843 -0.76673461 -0.60657485
#>  [91]  1.64731282  0.84106935  0.33817510 -0.40488308  0.90000904  1.19066666
#>  [97] -0.04753908  0.78692542  1.16668411 -0.27182387
#> 
#> [[8]]
#>   [1] -0.313704285 -0.286300150 -0.803440263 -0.094555455 -0.074046506
#>   [6]  0.714490683  0.124718366  1.296594339 -1.114492926 -0.842058813
#>  [11] -1.504294209 -0.284026455  0.042869041 -0.008866413 -2.949083784
#>  [16]  0.020358254 -0.098446980  0.589109098 -0.425751629  0.634658985
#>  [21] -0.578524902 -0.169109048 -1.919232520 -1.534266387 -1.114761221
#>  [26]  1.597811635 -0.639805078  1.566690553 -1.449155712 -0.791503959
#>  [31] -0.504480733  0.401826700  0.971396469 -0.579663020  1.604179789
#>  [36]  0.225973337 -0.514857122 -0.823788240  0.334415250 -0.093490765
#>  [41]  0.304062078 -0.476507534 -0.241311717  0.824155833 -1.555643970
#>  [46]  0.093501283 -0.366949421 -0.129408799  0.407795520 -0.583968954
#>  [51] -0.193449890 -0.269546670  0.073658232  0.357236138  0.550428418
#>  [56]  0.038401793 -1.609575292 -1.049710115  2.052033974  0.176436411
#>  [61]  1.128307358  0.435001744  0.548833492  0.647418304  0.878463454
#>  [66]  0.350797869  0.049879720  0.835749446 -0.281725199 -0.791679461
#>  [71]  0.001653727 -1.187530126  0.362417465 -0.549435319  0.692949838
#>  [76] -0.060843916 -1.193540930 -0.119908059 -0.708107950 -1.616663973
#>  [81]  0.495847824  1.299750970 -1.615985506 -1.250367945  1.582132306
#>  [86]  1.252529118 -0.209395701 -0.639033687  0.398902937  1.341973500
#>  [91] -0.037607422  0.543670114  1.740218394 -0.324455939 -0.447511333
#>  [96]  0.416082526 -1.295865368  0.682633090  0.496913333  1.439660656
#> 
#> [[9]]
#>   [1]  1.207358558  0.238514694 -0.298837878  1.384621050 -0.700069022
#>   [6]  1.860931509  1.803924959 -1.356043492 -0.314751072 -1.019265181
#>  [11] -0.612249656 -0.289068749  1.506983292  0.367722694 -0.020034941
#>  [16] -0.979921959 -1.400912333  1.445014910 -0.423481621 -0.733749597
#>  [21] -0.708484753  1.475092459  0.845004185  1.293994431  0.298161142
#>  [26] -0.405682490 -0.138807578 -0.222592731  1.747150257 -0.081704668
#>  [31]  0.124141586  0.392236141  0.378419970  0.166175673  1.153387369
#>  [36]  0.014748302  0.021440845 -0.826352170  0.047399133 -0.293107372
#>  [41]  1.207632181  0.032001384  0.180397550  0.077160580  0.558395338
#>  [46] -0.166854367 -0.227980619  0.242666056 -0.928482969  1.515406678
#>  [51] -0.811598824  1.311247408  0.815566796  0.120032668  0.881789848
#>  [56] -1.798743117 -0.957051078 -0.474058606 -1.860461283  0.253690073
#>  [61] -0.214401713 -2.370181451  0.003258078  0.669745568  0.226820843
#>  [66] -0.591868492 -0.405414494  1.414224665  0.241625825 -1.436087478
#>  [71] -0.314552745 -0.480542811  0.836254248 -0.052349287  0.951957625
#>  [76]  1.494512798 -0.344214729  0.645775407 -1.534500551  0.599672514
#>  [81] -0.504358892 -0.406909625 -0.460935206  2.214908426  1.486967820
#>  [86] -0.355098858  0.876203737 -0.214696646  2.020798363 -1.838661562
#>  [91]  0.172937164 -0.763245040 -1.118492914  1.197977461  0.981193912
#>  [96]  1.191749205  0.632833568  0.532825947 -0.820121900 -0.159325676
#> 
#> [[10]]
#>   [1] -0.017532122 -0.390664205 -0.402775110  1.276706963  0.821075754
#>   [6]  0.419785146  1.711661220 -2.215632715 -0.184878197  0.671168243
#>  [11] -0.795140242 -1.566257835  1.213732488  1.503810074 -1.169425021
#>  [16]  0.613377167  1.483931321 -0.443773736  0.268589741 -1.228216829
#>  [21]  0.478528762  1.669859573 -0.036455545 -0.440936271  0.734408890
#>  [26] -0.575309458  0.074897554  0.059301839  0.848264703  2.575396587
#>  [31]  0.094259964 -0.519949532  0.774428621  1.228289259  1.880606885
#>  [36] -0.125241723 -0.759094766 -0.509303603 -0.156039669  0.132582442
#>  [41] -0.974084022 -1.250949663 -0.046336042  1.137387771  1.264449891
#>  [46] -0.548742076  0.682391430 -0.929773363 -0.975148693 -0.027117079
#>  [51]  0.446671886 -0.329112249  0.113049080  1.051229395  0.326977400
#>  [56]  0.787037444  0.441559619 -1.019304514 -0.158698326 -1.637701292
#>  [61]  2.839266231  0.961819160  0.508432001  0.259523192 -0.773404126
#>  [66] -1.805527531  0.178383232 -0.525777309  0.400186316 -0.122587586
#>  [71] -0.071345032  2.095734365  1.322082615  1.200558995 -0.781874565
#>  [76] -0.441476451 -0.246028129  0.365792190 -0.034132882  0.296227003
#>  [81] -0.221059772  0.187706159 -0.438062172 -0.088733579  0.634250029
#>  [86]  0.865538237 -0.430593090  0.013030894  0.051249629  0.538638152
#>  [91]  0.519672613 -2.422911275 -0.806290253 -0.484679992 -0.653077284
#>  [96] -0.293764247 -0.276989004  0.136469589  1.067207722  0.001828103
#>