Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

Usage

samples_to_list(object, par_name = "y_pred", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object. Default to "y_pred" (variable containing predictions).

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  0.315228771  0.944166154  1.139540798 -0.836124718 -0.262411266
#>   [6]  0.645703690  0.513581018 -1.961957375  0.227033618  0.316375431
#>  [11]  1.679346462  0.980934162 -0.560971228  1.428510543 -0.663895064
#>  [16] -1.559786710  0.714138378  0.972410637 -0.233311372  2.266931538
#>  [21]  0.039296310 -0.386974958  0.146212658  1.893323150  1.804805295
#>  [26]  1.465056301 -0.117179724 -0.852638411 -0.108600290 -0.183792328
#>  [31]  0.171128929 -0.432364373 -0.901102249 -0.159301712  1.430939397
#>  [36]  2.080903848 -0.262684804 -0.452457407 -1.112865067 -0.766929064
#>  [41] -0.932396154  0.956224466 -1.338656645 -1.217422724 -1.212494707
#>  [46] -0.520061600 -1.450425084 -0.544089378  0.821122226 -1.598764785
#>  [51] -0.760712488  1.449566458  0.843052106 -0.377265586 -0.026149931
#>  [56] -0.394416452 -0.410390356 -0.179789989 -1.523016375 -1.283277864
#>  [61]  1.139495478 -1.043688064  0.420505535 -0.551804502 -1.688714832
#>  [66]  0.031494427  1.654879106 -0.276985680 -0.003177564 -0.412171793
#>  [71] -0.414727473  0.196983589 -1.009630398  1.485607143 -0.196494431
#>  [76] -0.148299504  1.188479006  0.030813528  1.141726277  0.834609310
#>  [81]  0.108477294  1.332733944 -1.631050324  0.719487289  1.761499112
#>  [86] -0.068333382 -0.494980892  0.041183510 -0.091492685 -0.023167780
#>  [91] -1.127498384  0.251224005 -0.379160037  0.666331485 -0.607975905
#>  [96]  0.228286229 -0.616874475 -0.542744370 -0.363702206  1.102508946
#> 
#> [[2]]
#>   [1] -0.039550077  1.351509398  1.176995470 -1.201527125  0.680036955
#>   [6]  0.540814753 -0.977657437  0.949824484  0.360951869 -0.333838356
#>  [11] -0.439420163 -1.410304869 -0.340709522  0.260437483  0.614490013
#>  [16] -0.457932218  1.329087302  0.446886525 -1.072045604 -0.034343071
#>  [21] -0.241657289  0.183135917 -0.214920035 -0.190121704 -0.685007420
#>  [26]  0.560394287  1.169945574 -0.508640051  0.652142007 -0.825427504
#>  [31] -0.392359757  0.125332970  0.843173308  1.249359805  1.178368425
#>  [36] -1.266206367 -1.007481562  0.073619028  0.206269683  0.854827168
#>  [41]  0.829776071  0.491983110  0.093405339 -0.279843773 -0.045366962
#>  [46] -0.707843696 -0.910081205  0.823288274 -0.635596259  1.212332679
#>  [51] -3.221166479  0.445253292  2.613619971 -0.787630178  0.211565893
#>  [56]  0.358624672  0.398264563 -0.449674673 -1.121634560 -1.655751269
#>  [61] -0.451972611 -0.674954809  0.005056652  0.037021158  2.059256655
#>  [66] -0.945545945 -0.382556656  0.523126301  0.701087422 -0.151073733
#>  [71] -1.975936297  0.812575236  0.651848084  0.737158789 -3.192460772
#>  [76] -2.363320286 -0.121775310 -1.106333661 -0.371233937  0.608816405
#>  [81]  0.109671990 -0.700748632  0.800433895  0.144859615 -0.138672388
#>  [86]  1.514694340 -0.074895372 -1.530814279  0.731185367  0.409365751
#>  [91]  0.504576560 -2.167476535 -0.117305214  1.495240852 -0.116103691
#>  [96] -1.291853177 -1.252506305  0.469696285  0.331163910 -0.635965096
#> 
#> [[3]]
#>   [1] -1.316552390  0.735539612  1.111417810  0.822375962 -0.539601077
#>   [6]  0.087539078 -0.224815963 -0.495094723 -0.662312425  0.068006066
#>  [11]  0.326369420 -0.304228826  0.463274962 -0.066786402  1.633575090
#>  [16] -1.276221721  0.495380111  0.396771746  0.235524066  1.014462462
#>  [21]  0.389405968 -0.432293138 -1.313386030  0.096298712 -1.472689299
#>  [26] -1.564414612  0.176479351  0.268307515  1.616211964  0.210838302
#>  [31] -2.028224981 -0.701815614 -1.113580463  1.698033902  0.817219968
#>  [36] -0.197450050 -2.475846069  0.791864235  0.516308559 -2.898374759
#>  [41]  0.186778539  0.825898542 -0.775598852 -0.831054809 -0.123288818
#>  [46] -0.053499209  1.084146809  0.807468174  1.106918878  1.143910039
#>  [51]  0.856855585 -0.605758287 -1.431117992  0.723329084  1.011692449
#>  [56]  1.282537646  1.567071471  1.042549140  0.464071578  0.302095087
#>  [61] -0.182472252  2.419602095 -1.410164438 -0.773049310  0.361210533
#>  [66] -1.101340985 -1.045746053  0.560077013 -0.290000039  2.008584528
#>  [71] -1.712163976 -0.610621575 -1.575436315 -2.102451574 -3.020208307
#>  [76] -1.165803487  0.235463433 -1.225406269  1.132256184  1.312807063
#>  [81]  0.229867611  0.742097370 -0.027523794  0.307521781  1.506709779
#>  [86]  0.726349706  0.054180614  0.836634025 -0.850107609  0.035573909
#>  [91]  0.173690720 -1.101397910  0.290551504 -1.344659202 -1.972909303
#>  [96] -1.937880151 -0.008535972 -0.171270720  0.507151300  1.575079139
#> 
#> [[4]]
#>   [1] -0.44663100 -0.30446993 -0.34860422 -1.23449112  0.42855976 -2.43135519
#>   [7] -0.95055764 -0.38812614  1.53750671 -0.52053670 -0.09927157 -0.26503843
#>  [13] -1.08570427 -0.94614697 -0.22169768 -0.39707606  2.32478539  1.07728627
#>  [19]  0.45728470 -0.36410457 -1.27071143 -0.25574805 -1.36461473  1.55125467
#>  [25]  1.70581809  1.28347735 -1.26779851 -1.21835747  0.40724857 -0.05288455
#>  [31] -0.17168919 -1.24977981  0.57820513 -0.22364744  1.47557766 -1.14278378
#>  [37]  0.22562923  0.42702182 -0.75309397  0.12797455 -0.85421060  0.47241297
#>  [43] -0.25819077  0.05505802 -0.79975744 -0.96513975 -0.37773187  2.14511476
#>  [49]  1.61738947 -0.80469755 -0.54046169  0.84461246  0.13014707 -0.11326285
#>  [55] -1.53687995  0.94500988  1.46080498  0.45824851 -1.33572908 -1.44084734
#>  [61] -0.23213441 -0.31581366  0.73839405 -0.15575868  0.32003655  0.37578396
#>  [67]  2.41700639  0.07968341  0.62371629  0.41687044 -0.12268948  1.83195041
#>  [73] -0.80017237 -1.82114020  0.68149146 -0.04136585  1.85074733 -0.19332862
#>  [79] -1.76962056 -0.76545258  0.82831234  1.58631890  0.86148105 -1.05282184
#>  [85]  0.21752628  0.52369058  0.23700238 -0.64510811 -0.68493125 -0.10036201
#>  [91]  1.34662728 -0.02937818 -1.47631333  1.41555192  0.82365268  0.54134196
#>  [97]  0.56565272 -1.28059248  0.68867847 -0.78429562
#> 
#> [[5]]
#>   [1] -0.34805557 -1.87043563 -0.98999531 -1.91099384 -0.89477960 -0.52928035
#>   [7] -0.02494027  0.81200113 -1.86465349 -0.66210902  1.11081973  0.56667752
#>  [13] -0.42951701  1.03949536 -0.40377428 -0.52639139  0.36287530 -0.37051210
#>  [19]  1.63754132  0.39981667  0.32846545  0.06622608  0.55670528  0.65764048
#>  [25] -0.64257624  0.44468173  0.88600566 -0.44305436  1.12009260  0.88822713
#>  [31]  0.44666753  1.35831605  0.05379208  1.02947998 -0.52989459  0.21618694
#>  [37] -1.26484579  0.47755889 -0.38460703 -1.68326794 -0.80758422 -1.06785226
#>  [43]  0.08932394 -0.06285882 -1.36055489 -0.30157250 -0.56787021 -0.32187052
#>  [49]  0.69852319 -0.31158868  0.83734596  0.39571086  0.68611788 -1.42890184
#>  [55]  1.46759627  0.57087134 -0.02890340 -0.90417300 -1.20987566  0.99335256
#>  [61]  0.83136355  0.01502505  0.23610104 -1.07228534 -0.95658428 -2.02431031
#>  [67]  0.47297275 -1.29143657 -0.04266269  0.14876157  0.81792601  0.59295292
#>  [73]  0.01853084  0.57300991 -0.10921259  1.11110543 -0.55463792 -1.86711588
#>  [79] -0.34412172  0.40215550  2.43210065 -1.14869874  0.51782988 -0.43901735
#>  [85]  0.72556104 -1.31411078  0.57201267 -0.11230363 -0.30888251 -0.48959613
#>  [91] -0.14126775  0.72024676  0.57639760  0.94717519 -0.53520882 -1.84123327
#>  [97] -0.04556789  0.98140915 -0.19015280  0.54530041
#> 
#> [[6]]
#>   [1] -0.003634807 -3.288604578 -0.471717527 -0.707346056 -0.461622196
#>   [6]  0.179280862 -0.080221878  1.822415383 -0.876358772  0.923403737
#>  [11]  0.721844001  0.111734216 -1.389617828  0.804174606  0.818431389
#>  [16] -1.044673678  0.411488727 -0.041376457  0.208884396  0.569135994
#>  [21] -0.663281727 -1.100158915 -1.060382139 -0.491970846  0.044903407
#>  [26] -0.546348414 -0.096606705 -0.407346981  0.362591131 -0.851439453
#>  [31]  1.072185959  0.384844062  1.209396649  0.210113230 -1.572253833
#>  [36] -0.779930449  0.862803654 -1.313959877  0.225197145 -1.269885881
#>  [41]  0.826020326 -0.181536096 -0.628783903  1.201685395 -1.541822502
#>  [46] -0.356713056 -0.397197159 -0.064191140 -1.384326652 -0.851902946
#>  [51] -0.144789264 -0.046032436  0.291163743 -0.035423175  0.039875427
#>  [56]  1.404472803 -0.479030328  0.291345024  0.022162149 -0.809279590
#>  [61]  1.180940333 -1.120739941  0.807894627  0.392228027  2.987862141
#>  [66]  0.046680258  1.010567846  0.044377518 -1.018979967 -1.313703131
#>  [71] -0.475660395  0.250127724 -0.658128379  0.669110501  0.823195068
#>  [76] -0.343091287  1.801449292  0.813793500 -0.957490874 -0.144536227
#>  [81] -0.773946800 -1.470208298  1.128046645  0.210304928 -0.482113755
#>  [86]  2.363430552  0.137063756 -1.265451452  1.470403654 -1.085291319
#>  [91] -0.012836647 -1.551586878 -2.508764233  0.144083346 -0.813662674
#>  [96] -0.320389089  1.175339118 -0.385324231 -0.243306057 -0.120700217
#> 
#> [[7]]
#>   [1] -0.64625925 -2.08009176  0.36892142  2.16354573  1.31811363  0.83719259
#>   [7] -0.08970492  1.02461854  1.18180588  1.33569275  1.34425570 -0.11515923
#>  [13] -0.61903495 -1.12885237  1.10751007 -0.50798060  1.15114773  1.28975928
#>  [19]  0.16754019  0.20146405  0.34744039  0.03184021  3.67598415 -0.46838501
#>  [25] -0.14999780  1.67750926  0.39146167  0.98838521  1.24373892  0.58207746
#>  [31] -0.64803344 -0.23502884 -2.69613688 -0.76650033 -0.20408710  0.16678338
#>  [37]  2.76837551  0.54243823  2.36575302 -0.57126327 -0.64266375 -0.17524767
#>  [43]  0.46603837  0.81626694  0.23030334  1.51425654  0.38337287 -0.88263529
#>  [49]  0.49512158 -1.65892476 -1.12542614  1.11032673 -0.45880094 -0.72424205
#>  [55]  0.14541627  0.83159013  1.01798340 -0.49171151 -0.01751654  0.24180312
#>  [61] -1.18415697  0.06771741  2.46766995 -0.69446111 -0.29293561 -0.20057870
#>  [67]  0.17221767 -0.19986451 -0.96624673  0.23516543  0.01358102  0.48228577
#>  [73] -2.17778796 -0.08993221  0.62420718  1.03530932  0.46351284 -0.36624791
#>  [79]  0.58171312 -0.25513021  1.90293454 -0.91963106  0.14121240  1.74222118
#>  [85]  0.96542935  0.23037238  1.76578925 -1.56641653 -0.52079990  0.56019621
#>  [91] -1.39866728  0.22161558  0.81887512 -0.44922272 -0.51903773  0.04289485
#>  [97]  0.87201633  1.14610431 -0.72693644  0.45505199
#> 
#> [[8]]
#>   [1] -0.84624048  0.58556677  0.96563067 -0.82937744 -0.70863310 -0.62519786
#>   [7]  0.92005904  0.72095526  1.19500664  0.44539828 -0.98508105 -0.93156945
#>  [13] -0.43788880 -1.70475522 -0.06650246  0.49837798  1.14703652  0.71942705
#>  [19]  1.11490653  1.13563490  0.37607883  0.93027701  1.25367045 -0.60465173
#>  [25] -0.76997576  0.02795256  0.26708909 -0.79322078  1.81491014  1.03113852
#>  [31]  0.15359302  0.59187870 -0.15417591  1.08670698 -0.50668355 -0.45188446
#>  [37] -0.74018143 -0.08852819 -0.43455661 -2.50953520 -0.03883684  1.29197427
#>  [43]  1.25092697 -0.24110566  0.42631414  0.52582448 -0.38100348  0.92709193
#>  [49] -0.56916098 -1.69655699 -1.16570723 -0.31180079  0.35862907  0.35542552
#>  [55]  0.34361215 -1.04625208  1.29058496 -0.14274547 -0.35081513 -0.27016285
#>  [61] -0.30394862  0.78419270  0.87472284  0.74662840 -1.33483250 -0.36183810
#>  [67]  1.75058881 -0.25992006 -0.03828673 -0.31355572 -2.34986485  0.33660779
#>  [73]  0.52030852  0.15077807 -0.99118814 -0.89250376  0.21637065  0.62124587
#>  [79] -0.95849785  1.00568321  2.41988978 -0.38139337  0.91547640  0.65756775
#>  [85] -2.32472699  1.25644248 -0.33835664  0.82336668 -0.30790728 -0.54210284
#>  [91] -0.02058160  1.82213194 -0.03367723 -0.68154280 -0.25943253 -0.36324793
#>  [97]  0.78506151  0.65930480  0.07880938  1.67757359
#> 
#> [[9]]
#>   [1] -0.24679096  0.43647730  0.60537724  1.08719676 -0.16627664  1.66040686
#>   [7] -0.74091330  1.82941144 -1.72945248  0.51517428 -0.57475576 -1.62168819
#>  [13] -0.01105101 -1.46730065 -1.13837833 -1.33224546  0.25889104 -0.19287420
#>  [19]  0.33297955 -1.53287026  0.41543565 -1.38800935 -0.73456393  1.30379478
#>  [25] -0.35925396  0.26594087  0.53307494  1.71143632  1.34719859 -1.48032947
#>  [31]  1.69652608  0.57158390 -0.77224401  0.25320620  0.22979493 -1.90048322
#>  [37]  1.07790860 -1.31189985 -0.12108235 -0.32042724  0.65520942  1.06018770
#>  [43]  0.56871282  2.50590489 -0.99408402 -1.44425818  0.70401035  1.04500407
#>  [49] -0.85430987  2.40666097 -0.83419546 -1.06570211 -0.83910835 -1.25616445
#>  [55] -0.29696525 -1.42136977 -0.36189636 -1.95179224 -1.06331853  1.22792727
#>  [61] -1.06269804 -0.22606414  1.61297971  1.62309769  0.28906943 -0.68369171
#>  [67] -0.74258617 -0.78078544 -0.42103494 -0.74517779  0.17865239  0.81529501
#>  [73]  0.04785899 -0.26297123  0.63479757  0.30476884  0.28408255 -0.59981501
#>  [79] -1.04813882  0.22901673 -0.13908344  1.01146236 -1.30372945 -0.03001321
#>  [85]  0.23877214 -2.49176184  0.32088110 -0.47320548  0.84281724  0.06780483
#>  [91]  0.20241621  0.34777372 -1.42713263  0.16783811  0.81797978 -1.52162383
#>  [97] -0.47763704  0.02839282 -2.82566971  2.15157530
#> 
#> [[10]]
#>   [1] -0.352962915  0.043330152  1.034715179 -2.745068009  0.188292411
#>   [6] -1.097767512  0.775929139  0.698588422 -1.018313096  1.390981564
#>  [11]  1.689637981 -0.420269926  0.942134769  0.126921425 -0.201456092
#>  [16]  0.980253942  0.714586841  1.025265537  0.921807187 -0.649383878
#>  [21]  0.318515896 -1.615870960  0.827639478  0.527849454  3.222688460
#>  [26] -0.525896019  1.577898120  0.509932324 -0.013015369  2.508464685
#>  [31] -0.047147537 -0.146962999 -0.196664397  1.210426934 -0.315860375
#>  [36] -1.035097750  0.498554386 -0.571064505 -0.360606191 -0.378165692
#>  [41] -0.425163112  0.446628530 -0.720990622  0.707803811  1.077478040
#>  [46] -0.172426154 -0.284169961  1.658992724  1.754012147  0.400631702
#>  [51]  0.081135162  0.863819447  0.038391490  0.307689975 -0.055807174
#>  [56]  1.043138666  1.751596295 -0.332809356 -2.074897443 -1.790865504
#>  [61] -0.641886538  1.372131500  0.320686808 -2.053037652 -1.534292832
#>  [66]  0.829927276  1.187918376 -0.036029469  0.695351314  0.802230898
#>  [71]  1.103290484 -0.187539620  0.499851138 -0.094691807 -0.078582721
#>  [76]  0.869063910  0.894128705 -1.831340774  2.651770740  0.015050582
#>  [81] -0.003503306  0.004914012  1.033488127 -0.748204359 -0.222167051
#>  [86]  1.188048971 -1.366080638 -1.759400782  1.297062238  0.994074770
#>  [91]  0.716327485  0.102331684  0.494139914  0.180827057 -1.412365309
#>  [96] -0.298169142 -1.518593757 -0.789598552 -0.864851897  1.002412233
#>