Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

Usage

samples_to_list(object, par_name = "y_pred", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object. Default to "y_pred" (variable containing predictions).

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1] -1.07391224  0.72632282  0.74767794  1.19446552  2.16435753  0.72656146
#>   [7] -0.65843939  1.06392165  0.56861340  0.01567327 -0.69870945 -1.61001460
#>  [13]  1.37176453 -0.56644868  1.12176136 -1.98543298 -1.55298373  0.66192288
#>  [19] -0.51318550  1.99926439 -0.95442197  1.54320117 -0.67729055 -1.61908353
#>  [25]  0.63824090 -0.94965453 -0.56227281  1.40596884 -0.27225807  0.57904272
#>  [31] -0.31176197  0.01958789  0.43343780  1.02339073  0.83274125  0.71272400
#>  [37] -1.16291478  0.03078538 -0.43900058 -0.78292244 -0.63606466  1.32657509
#>  [43]  1.12055820 -0.47356500  0.63085602  0.33004904  2.39777276 -0.38877170
#>  [49]  1.49936047 -0.20228847 -0.97211238  1.02994921  0.66820849  1.45877818
#>  [55]  0.10845491  0.12094121 -0.28337588 -0.33615202  0.56392160  0.56346627
#>  [61]  0.64927877 -0.60085364  0.74339267 -0.79760362  0.61135726  0.47397756
#>  [67]  1.16223723  1.62764924 -1.18152256  0.12397404 -1.52252180 -1.73896649
#>  [73]  1.06493921  0.80153047  1.68693363 -0.74507596  0.47542852 -0.63075867
#>  [79] -0.58560203 -0.87002384  0.06874574  0.06270511 -1.83485168 -0.09313480
#>  [85]  0.39508151 -0.83218773  0.09602872 -0.53847223 -0.01864987  1.23642276
#>  [91]  0.65164824  0.12918809  0.16917948 -0.06363146 -1.93982500 -1.73708200
#>  [97] -0.18965761  0.07802207 -0.52794124 -0.59561911
#> 
#> [[2]]
#>   [1]  2.506987239 -1.647395007 -1.177922471 -1.222182385 -1.879650014
#>   [6]  1.523403001 -1.194218105 -0.954351855  1.118387491 -0.582734717
#>  [11]  0.694856411 -0.346850312  1.588429662 -0.428521064 -0.428642947
#>  [16]  1.304954026  0.404664736  0.771389516  0.629679242 -0.452346776
#>  [21] -2.153798318 -0.566007467  0.682099685  1.079579192 -0.377190931
#>  [26]  0.154953571 -0.836822545  1.409999834  0.165546349  1.718106176
#>  [31]  0.162822076  2.999104589  0.512360996 -0.586697965 -0.599755821
#>  [36]  0.021920406  1.190560981 -0.544090010 -0.075081708  0.269755752
#>  [41]  0.692768493 -1.163129325 -0.888713425 -0.993519864 -0.989850403
#>  [46] -1.460952079  0.330132269 -0.510533176 -0.536561640  0.318898664
#>  [51] -0.372149800 -0.936992050  1.842076821  1.372335338 -2.572490237
#>  [56] -0.981732241  0.137466561  1.955653759  0.788804295  0.063552739
#>  [61] -0.426500636  0.449623830 -1.266948913 -0.297072897  0.603137093
#>  [66] -0.755083077 -0.069021496 -0.063220458 -0.307269556 -1.590949362
#>  [71] -1.204069758  2.521365588  1.436886730 -0.854227130 -0.003649967
#>  [76]  1.656731902  0.367203237  0.671449026  0.740100362 -1.058944077
#>  [81] -1.270330487  1.329339826  1.286746428  1.545438343  0.172838358
#>  [86] -0.154494700 -0.105630727 -0.089512236 -0.922736030 -0.658713992
#>  [91] -0.469081769  0.390248887 -1.018834175  1.352895156  1.349076078
#>  [96] -0.019791577 -0.895928473 -1.033038373  0.089957079  0.036091068
#> 
#> [[3]]
#>   [1] -0.27333063 -0.27610447  0.18991337  0.69232761  0.43768752  0.07043479
#>   [7]  1.51828710  0.58025998  0.57542987  0.20753799  0.91337722  1.18906591
#>  [13] -1.48646441  2.06007149  1.05598007 -0.34840271 -1.09307938  0.26620845
#>  [19]  0.46773948  0.57543507  0.13561647 -1.63161558 -0.36105309 -1.17173598
#>  [25] -1.95068936  0.32423290  0.44049357 -1.74020148 -0.19989745 -0.56592972
#>  [31]  2.50131290  0.51318834 -0.88533814 -1.55074799  0.11501835  0.91339401
#>  [37]  0.81218844 -0.24113456  2.58215531 -1.05981037  0.27823667 -1.03409592
#>  [43] -1.02303434  0.81795413 -1.18243371  1.92115700  1.06144950 -0.77568864
#>  [49] -0.19520052 -0.09991337 -0.08640749  1.87256915 -1.28017893 -0.27510891
#>  [55] -0.60518549  0.22690006 -1.45862992 -0.36862325  1.25202007  0.09559597
#>  [61]  1.09640425  0.72921374 -0.43446720  0.39409827 -1.07957988 -0.11039700
#>  [67] -1.89029315 -0.29818666  1.40722008 -0.15629478 -0.34758989 -0.21654242
#>  [73] -0.91322151  0.86148808  1.16059398  0.47851827 -0.25330602 -1.69698706
#>  [79] -0.38398773  0.27004575 -0.32300583 -0.59545618 -1.78829947  0.73508886
#>  [85]  1.56578477  0.40152084 -0.42141925 -0.05312468  1.59539764 -0.36953180
#>  [91]  2.09824521  0.83943674  0.25865698 -0.55593380  0.44432728 -0.67238382
#>  [97]  1.13246107  1.03519947 -1.60999550 -0.74538402
#> 
#> [[4]]
#>   [1]  1.38117218  0.08480017  1.02495197  0.79998181  0.12694249 -0.20208260
#>   [7]  0.19504115 -1.12852873 -0.28229261 -1.32444163  0.38073380 -1.41433978
#>  [13] -0.37081917  0.44635146  2.34240634  3.14601903 -1.51466613  0.35127043
#>  [19]  0.81290784  0.24135639 -1.57915805 -0.09381920  0.49755338  1.60772737
#>  [25]  0.25748405  0.51370360 -0.46921101 -1.21419821 -0.24177060 -0.72374315
#>  [31]  0.81763965 -1.91226583 -0.65056177 -0.36704828  1.66540368  1.86259969
#>  [37] -0.81254295 -0.72288892  0.11938490  0.66022930 -2.15834125 -0.72735195
#>  [43] -0.30679676 -1.75491250 -1.50746963 -0.76797275  0.26927811  0.37005110
#>  [49]  0.03459042  0.45427214  0.54415282  1.10780400  0.83915275  0.75960184
#>  [55]  0.55386281 -0.17574712  0.28534555 -0.09611311  0.73787578  0.44504715
#>  [61]  0.24933902  0.64817137 -0.34996096 -0.11529444  0.07137286 -0.20143293
#>  [67]  0.20499469  0.77144922  0.69381645 -0.97289070 -0.09994862  0.92285625
#>  [73]  0.01597948 -0.03685962  0.91841862  1.53992121 -0.69633198 -0.77076071
#>  [79] -0.71735535 -0.28713298 -0.23850492  0.47382679 -1.02353573 -0.34594382
#>  [85] -1.04285546  0.24332697  0.59810555 -0.07118692 -0.18308858  0.04236427
#>  [91] -0.51250874 -1.12034113 -0.16022012 -1.00812359 -0.86498954  0.19904051
#>  [97] -1.53974192 -1.67255254 -0.75804122  0.65011652
#> 
#> [[5]]
#>   [1] -1.75966603  0.07882397 -1.10105163 -1.27909072  0.39073412 -0.88450735
#>   [7]  1.00431723 -0.30263874 -0.50826313  1.85929708  0.84699818  0.30994603
#>  [13] -0.34958027 -0.51451004  0.08675165 -0.76455763  3.36591997  1.68534308
#>  [19] -0.78634098  0.40504928 -0.75971846 -0.56523706  0.40542358  1.88047632
#>  [25]  2.17898004  0.86937071  0.52914556  0.13629995 -0.26140447  0.83938914
#>  [31] -0.10468352  0.59749570 -1.17194879  1.35113467  0.52922462  0.15640525
#>  [37]  0.53094965 -0.20346857 -1.02325728  1.16659849  0.14693246  0.16096852
#>  [43] -0.36226492 -0.57058770  1.76368378  0.01360511 -0.42002316 -0.70072892
#>  [49] -1.81284386 -0.59811799 -0.67904259  1.82873294 -0.34973123  0.30031843
#>  [55] -1.30434119 -0.49620393  0.81133927  0.69334036 -0.10587426  0.04562836
#>  [61]  1.51158223 -0.70615569 -0.61038609 -0.42065325 -1.40768900 -0.67794653
#>  [67] -0.75321776  0.95167141  0.79038908 -1.13005624  0.19055314 -0.41412401
#>  [73] -0.84282575 -0.91669488 -3.07007062  0.49923643  2.21107564  0.26782552
#>  [79]  0.91667954  1.20625486 -0.60592057 -0.17590427 -0.43581946 -2.38003089
#>  [85] -1.11205889  0.32381936  0.07037668  0.71927397  1.67841318  0.32820851
#>  [91] -0.79368466 -1.39225800 -0.62438613  0.01434880  1.26868223  0.75304364
#>  [97]  0.17392451  0.10659862  0.37685745 -0.12036793
#> 
#> [[6]]
#>   [1] -0.6541048451  0.3955800205  0.3617158978  0.5384894928  0.1164112977
#>   [6]  1.3643751153  0.3214319655  1.2842951723 -1.6594483397  0.5669599473
#>  [11]  0.4545154227 -2.2161811341 -0.3963183616 -0.0587097434 -1.9801769785
#>  [16] -0.1693452512 -0.1414526342  1.9466547827 -0.0470986978 -0.0158973933
#>  [21]  0.8757972837  1.9007289221 -0.5347377782  0.9447868482  0.8568677405
#>  [26]  1.4067495707  0.1626234737 -0.3675951386  1.4467281962  0.5173810416
#>  [31] -0.0731040205  1.0968591226 -0.2207495592 -0.3278973922 -0.6357938845
#>  [36] -0.0004533774 -1.1559842311 -0.1130253516  0.4874163368  0.4852402940
#>  [41] -0.7362533559 -0.6854619808  0.4626427301  0.8878179691 -0.8469149204
#>  [46] -0.3620600741 -0.2826892996  1.1278121981 -1.3263101936  0.9976071979
#>  [51]  0.1685144439 -0.4908947884  0.1334723737 -0.6858790153  0.5590303647
#>  [56] -1.9580992624  0.8230390539  0.7152416381 -0.1049600471 -1.1709293222
#>  [61] -0.7881311202 -0.9327265847  0.1033752496 -0.1318753356 -0.4543954267
#>  [66]  0.3563005109 -0.8182882425  0.0200729699 -0.0128313839  0.1829914889
#>  [71] -0.9122030321  0.2849892263 -3.0773678761 -0.7290004536 -0.6766247362
#>  [76] -0.5043606202  0.1623989761  0.9260112846 -0.4153914571  2.4220905353
#>  [81] -1.1602028146 -1.5150782655 -0.6422341703  0.1250580843  0.1435633531
#>  [86] -0.3171942575 -0.5564301474  0.6195650189  1.2670788478  1.9995548447
#>  [91]  0.3478174742 -0.7482812210  0.9270541840  0.5065587896 -1.0280324999
#>  [96] -0.1911524656  1.5535183388  0.8924878422  0.7724111665  1.1226768163
#> 
#> [[7]]
#>   [1]  0.46306696 -0.72621906  0.15726585 -2.04131904 -1.42173749  0.43762384
#>   [7]  0.29063317 -0.03877927  1.29883393 -0.66976593 -0.20159304  1.09879888
#>  [13] -0.59396060  0.66466656  0.60139529 -0.52586748 -0.77581797 -0.60984083
#>  [19] -0.59911324 -0.33059514  1.21228834  1.80374424 -0.15263121  0.09339159
#>  [25] -0.97493594  0.42406985 -0.47761491 -1.05151542 -1.15484134 -2.21723095
#>  [31] -0.96820756 -1.37074140  0.77050846  0.55294862 -1.90574301 -0.25992591
#>  [37] -0.02680732  1.64111137  1.55690595 -1.15220462 -0.07654427 -1.15096619
#>  [43] -2.51273455  0.48852547  0.15468978  0.15294453  1.39240545  0.11140400
#>  [49] -0.47968380 -0.10552448 -0.29867080 -1.03739254 -1.10088586  0.08439542
#>  [55]  0.29960431  0.33995851 -1.31199143  0.22024774  0.19740998  0.83457778
#>  [61]  0.10952215  1.14765464  1.07924170  0.82005022  0.03034144 -0.61773291
#>  [67] -1.02502373 -0.08256784  0.50311417 -1.87459168  0.13255422 -1.04742998
#>  [73]  0.23767202 -1.23263962 -1.54032820  0.13777740  0.49461683  0.60598061
#>  [79]  0.45919335  0.10480618 -0.56800327  0.24498556 -0.06324096 -0.84987748
#>  [85]  0.87971436 -0.87165745  0.10068379  0.45026544  2.67245066  1.52154755
#>  [91]  0.81057673 -0.35930161 -0.79117568  0.95054096  0.60119579  0.93258562
#>  [97] -0.05717022  1.22469134 -1.86295507  1.49782715
#> 
#> [[8]]
#>   [1]  0.4975698565  0.5334456751  0.0089164666  0.2462387598  0.3038403435
#>   [6]  0.5918999914  0.2162272265  0.3658653617  0.3280311217 -0.8823203887
#>  [11]  1.1202358311 -0.0960379203  0.6026261103  0.7323090146  0.7149697484
#>  [16]  0.9133219486  2.3628276020 -1.3486650649 -0.8067928312  0.8552450870
#>  [21]  1.3177161603  1.7201849198 -0.3249018182  1.2168591163 -0.9826852240
#>  [26] -0.5854324195 -0.3451931548  2.0316574769  1.0944218552  0.8561512362
#>  [31]  1.5350577562 -0.3631117216 -0.8858677819 -0.9514075716  0.1910337237
#>  [36] -1.2758892701 -0.0802450401 -0.6082524935  0.7781137092 -0.9567991876
#>  [41] -0.5272814110 -0.3895486107 -0.0585738849  0.3311986641  0.0871214673
#>  [46] -1.3228289973 -0.9929137925  0.2061843611  0.1904828296  2.6181403735
#>  [51] -0.0614589356  1.3801829543  1.9398433344  0.1823286907 -0.5377418548
#>  [56]  0.9787552170 -1.0648672228  0.4838030339  0.0041447855  1.0088828132
#>  [61] -0.3618222361  0.4503501498 -2.9704783676  0.0220646481 -1.2089369739
#>  [66]  0.9566145197 -0.0002211627  0.5542231960  0.0576845046  1.4089985870
#>  [71]  0.1773888172 -0.7629126842  0.9498912132  0.5184861622 -1.1239603117
#>  [76]  0.5996287022 -0.9148617602  0.2086806516  1.3323407213 -1.1329294379
#>  [81]  0.3299285596 -0.2309252185  0.0516148640 -0.3448431872 -0.4236668395
#>  [86] -0.3013603962 -1.9278410429 -1.2233527371  1.0063018539 -1.1944233809
#>  [91] -0.2834274059  2.9995385625  0.1838915141  0.6906847069  0.1906545228
#>  [96]  0.6973646238 -0.3104020996  0.5706042504 -0.0158074286  0.0669393887
#> 
#> [[9]]
#>   [1] -0.011393431  1.085964504  0.647786816 -0.697923607 -1.145966679
#>   [6] -1.096190505  0.849482176  0.591876243  1.696491224  1.079850435
#>  [11] -1.456196898 -0.544149580  0.215143547 -0.265649805  1.026950214
#>  [16] -0.536510252  0.108638923  0.181164538 -0.934009560  1.162802731
#>  [21]  0.182762602  1.374353898 -0.899930741  1.067887962  1.502434664
#>  [26] -1.025219167 -0.016903991 -0.162805592 -1.144795403  0.174430941
#>  [31]  0.836515461 -1.029743888 -0.481918564 -0.088370464 -0.001663947
#>  [36] -0.658961519  0.134125403 -1.288302734 -2.279853910 -1.038353578
#>  [41]  0.062771633  0.657545496  1.241477009 -1.357141658 -0.116807130
#>  [46] -0.773332611 -0.534075873  1.465531903 -0.496957313 -0.717285138
#>  [51] -1.431907373  0.234131278  1.218626997  0.246245727 -1.737313320
#>  [56] -0.798997412  0.270981160 -1.238282144  0.143075117 -1.196849026
#>  [61]  0.732773175 -0.998893319 -0.576948122 -0.437421287 -1.029728011
#>  [66]  0.132524720  0.114103371 -0.191216974 -0.808268305  0.318803076
#>  [71]  0.104728596 -0.601243571  0.101335090  0.507381779  0.242606772
#>  [76] -0.487963029  0.124559429  0.460319328 -1.010870695  1.667732620
#>  [81] -0.505955521  0.031858886 -0.393496794  0.654789266  0.415608131
#>  [86] -0.222601745  0.432614896 -0.861580410 -1.644709868  1.027923328
#>  [91] -1.599709075 -0.463796202 -1.624124003  1.578944434 -0.929852514
#>  [96]  0.936513654  0.625519845  0.178203728  0.379096574  0.279865066
#> 
#> [[10]]
#>   [1]  1.464699708  0.708408938 -1.208473636  0.913674986  0.358173162
#>   [6] -0.308519905  0.903039881  0.799871595 -0.120098335 -0.421495049
#>  [11]  1.204873112 -0.463028778 -0.283315332 -2.746608926 -1.013876285
#>  [16] -2.681693283 -1.775096279 -1.142192151 -1.057871407 -0.445886314
#>  [21]  1.166305351 -0.201691502 -0.272217266 -0.315499550  0.420722071
#>  [26] -1.695917141 -1.101697607  0.389035364 -0.790149576  0.148200695
#>  [31]  0.399947312 -0.966988930  2.297243593  0.551542941 -1.040081895
#>  [36]  0.080633101 -0.922457665  0.588676941 -0.751850307 -0.064899969
#>  [41]  1.761739940  0.919662520 -0.539578234 -1.651499895 -1.286774187
#>  [46]  0.454170963  0.709103364  1.853295831  1.413147495  1.496911039
#>  [51] -0.868241745 -0.695662781  0.711028182  2.832952348  0.173825802
#>  [56] -0.557890983 -0.185233786 -0.101896090  0.029275633  1.358675647
#>  [61]  0.093679404 -0.131178069 -0.242318097  0.596690432 -0.616032661
#>  [66]  0.104939461  0.772059381  0.005913627 -0.773001566 -0.482461149
#>  [71] -0.013914788 -0.969230489 -0.732023663 -0.052835127  1.579402994
#>  [76] -0.592288201  0.403231656 -1.510727540  1.331485688  0.434781148
#>  [81]  0.602543291  0.928377660  0.981365173 -1.395874929 -0.858162831
#>  [86] -0.460890870  1.363103647  0.663433936  0.981712101 -1.240611812
#>  [91]  0.151930039  0.028145867  0.620198095  0.413493736  0.399275140
#>  [96] -1.153075874 -0.628354038  0.925914552 -1.454884414 -1.457751344
#>