Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object
Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name
Name of variable to extract when object is a stanfit object.

n_samples
How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))

#> [[1]]
#>   [1]  0.840801808  0.791534124 -0.168848948  0.612722104 -0.771158924
#>   [6]  0.888628993  0.013214477  0.225339515 -0.729915210 -1.222487070
#>  [11]  0.406805171 -0.751012223 -0.162116540  0.352010126 -0.289058300
#>  [16]  0.104662227  0.720186531 -0.611046082 -1.106914072  0.534803326
#>  [21]  0.736067968 -1.222501574  1.021415310  0.465165158  0.790472705
#>  [26] -0.130264801 -0.930285334 -0.364851004  0.153872493  0.413154818
#>  [31]  2.480823360 -2.179956742  0.420874578 -0.357528325 -0.646861514
#>  [36] -0.050141801  0.416942847 -0.632587542  1.150146673 -0.235475907
#>  [41] -1.643107386 -1.503382146 -2.050584847 -0.753198229 -0.134141958
#>  [46]  1.005782847  2.167186798  2.322556540 -1.020423391  0.048814436
#>  [51] -0.771888628 -0.785235068 -0.726603031  0.681878032 -0.229843287
#>  [56] -1.510601724 -0.583727687 -2.022918454  0.403504676  0.550015549
#>  [61]  0.028357122  0.893165020 -0.376555496  0.605884808 -0.004874726
#>  [66] -0.520796373 -0.639018598 -0.635894137  0.106586975  1.176914248
#>  [71]  0.447391153  2.272954766  0.136058206 -1.999039133 -0.420500870
#>  [76] -0.378407395  1.220774789 -1.541030292 -0.310310122 -0.020108184
#>  [81] -2.390200336  0.889865359 -1.482813325  0.445750348  1.369775856
#>  [86] -0.020110027 -0.109217587  0.264661745  0.303848264 -0.183388483
#>  [91]  0.559649672 -0.186553842 -0.812275372 -1.640581672  0.507922478
#>  [96]  1.754336961  0.592400202  1.016713288  0.121620586 -1.078067265
#> 
#> [[2]]
#>   [1] -1.14356572 -0.52964368 -0.68127316 -0.20244756  1.68449572 -1.03377324
#>   [7] -0.15597667 -0.04640064 -0.95362873  0.41626080  0.11402961  0.06391875
#>  [13] -0.91933224  0.90133529 -0.79772830  0.66822120  0.15521430  0.12868809
#>  [19] -1.53306545  0.20236067 -0.71753865  0.36169476  1.39900429  0.37269896
#>  [25] -1.56564429 -0.05169454  0.51408210  0.54989952  0.86781691  0.68436008
#>  [31] -0.16267998 -1.78436472 -1.03714557  0.83014772  0.60734694 -0.12218636
#>  [37]  0.93312514 -0.96127668  0.25508171 -0.54540157  0.93036074 -0.53765056
#>  [43] -0.45242607 -0.43929097 -0.61622311  0.44163455  0.48259745  0.54214438
#>  [49] -2.29078465  0.31035185  1.40407898  1.37711764  1.06039565  0.63217289
#>  [55]  1.08492804  1.35645944  0.36242404  2.16934446  0.13913051  1.37632653
#>  [61] -0.49144999  1.53491576 -0.41619872 -0.52054380  0.85058387  0.33449657
#>  [67] -0.82935177 -0.21869594 -1.54508372  0.23322978  0.03106964  0.35786565
#>  [73]  1.60862422  1.42985426 -0.94833964  1.01570554  0.03773461 -1.74423940
#>  [79] -0.97007381 -0.22902736 -0.97484568  2.69237242  1.39239386  1.36007615
#>  [85] -0.36541588 -0.86893720 -0.50655673  1.21470730  0.50695868 -2.09497131
#>  [91]  0.03407924  0.85272870  0.74322814  0.55715361 -1.24858322 -0.20787431
#>  [97]  0.42396946 -0.50669744 -0.61152331  2.41165945
#> 
#> [[3]]
#>   [1] -0.16495814 -0.44040605  0.52197617 -1.91832225 -1.98264996  0.52120016
#>   [7]  0.77778320 -0.81211887  0.91074832  0.94875395 -1.34804945  0.35417586
#>  [13]  0.53026393 -0.31097493 -0.24415553 -0.29187819 -1.12808616 -1.12288159
#>  [19]  2.05393864 -0.90951843  0.45837916 -0.04661845 -0.68095460 -1.48882940
#>  [25] -0.39251301 -1.74968202 -0.03866763 -0.79715324 -0.91592054  0.07326746
#>  [31]  1.23817132  0.71837134  0.53946650 -1.06123399 -0.54073787 -0.71520471
#>  [37] -0.17892009  0.16497028 -0.72740817 -0.53783713 -0.59966490  1.18208775
#>  [43]  0.18921288  0.24894911  1.04664341 -1.28930094  0.37511557 -0.55691839
#>  [49]  0.30242663  0.22226647 -0.96216696  0.05203237  0.90004038 -0.39490088
#>  [55]  0.95414600 -0.58821435 -1.47658453 -1.37777577 -1.34567231 -0.73663796
#>  [61] -0.47011150  1.38060934  1.67501093  1.17690663 -0.14889834 -0.17782336
#>  [67]  0.80312186 -0.41579535  1.21248959  1.24036003  0.68567588 -0.02679868
#>  [73]  0.30958051  0.24986433 -1.35646087  0.59937977  0.00864779  0.09071661
#>  [79] -0.65705741 -0.48178225  0.01775788 -0.85953576  1.35770498  1.21506118
#>  [85]  1.45391292 -0.08591198 -0.61756875 -0.21826025 -1.32600521 -2.36220890
#>  [91] -1.40996955  0.25439714  0.29587030  0.05678979 -0.10265486  1.95192303
#>  [97]  0.87856596 -1.30661693  0.09034427  0.90235804
#> 
#> [[4]]
#>   [1]  0.54942132 -0.86983920 -0.03921465 -0.51845168 -0.91184962  0.15031364
#>   [7]  0.44035545  1.31978206  0.24656113  0.94182660 -0.34215133 -0.27614681
#>  [13]  0.38365678  1.44691022  1.73390029  0.45642938  0.70750349  2.06378922
#>  [19]  0.02391059  0.24594946  0.27205914 -0.99245858 -0.02757795  2.22284516
#>  [25]  0.15550390 -0.86911632 -1.17448945 -1.75967731  0.05836159  1.16451986
#>  [31]  0.33762789 -1.05451872  0.66076911 -0.82340900  0.43703366  0.37238811
#>  [37] -1.64674913 -1.92355227  0.38083303  1.37570535  0.82584873 -0.41611500
#>  [43]  0.98214959  0.24218073 -0.93792685  1.50585850 -1.06585117  0.16227053
#>  [49]  0.26047039 -1.48820199  1.41518387  0.56250751  0.64205574 -0.89053264
#>  [55] -0.11653752 -0.94197475  1.11582792  0.47013513  0.86061271 -0.07039665
#>  [61] -0.61318021 -0.33671496 -0.21584018  0.62113229 -1.28402652 -1.30009244
#>  [67] -0.37676946  0.10374865 -0.70356227  1.49741394 -0.30282688 -1.37683133
#>  [73]  0.88317017  0.67142028  1.23002238  1.64546788 -0.18292604 -1.37925781
#>  [79]  1.25060060 -1.87623677 -0.36878757 -2.12929731  0.33764603 -1.57457485
#>  [85] -0.19631941  0.91506011  1.07863654  1.45280285 -1.65328109 -2.76186752
#>  [91]  0.26436916  1.02457159 -0.65918555 -0.38630438 -0.24249895  0.22059324
#>  [97] -1.48109003  0.72469120 -2.37751470 -0.06540921
#> 
#> [[5]]
#>   [1] -0.46970481 -0.28051629  0.57483659 -0.20567685  0.23488051  0.11062832
#>   [7]  0.27420223  1.57491963  0.56776783 -0.05961225 -0.14958771  0.26616879
#>  [13]  0.52840324 -0.57938809  0.94219937 -0.70636136 -1.27540210  0.73325992
#>  [19] -0.42020709 -0.25537509 -0.36923822  1.21244833 -0.44536301  1.13866752
#>  [25]  0.19052348 -1.36639361 -1.29810260 -2.13709831 -0.09620553  0.69450905
#>  [31] -0.47459123  1.76313061  2.02584606  0.42761881 -0.56647935 -0.83383531
#>  [37]  0.39825915 -1.47453793 -0.39544120 -0.13070247 -1.99949395 -0.99593086
#>  [43]  0.42011400 -0.06286163 -0.08013185 -0.03228341 -0.71898093 -1.11656132
#>  [49] -0.78026990 -1.77695853 -0.42783487 -2.03102701  2.75076475  1.51366972
#>  [55]  0.03398894  0.74079278 -1.27080818 -0.16356423 -0.61094549  1.25188419
#>  [61]  0.25031947 -1.70558168 -0.85541313 -0.14490163 -0.32444696 -0.17256490
#>  [67] -1.23606292 -1.90230421 -0.09450402  0.03255579  0.46129012  1.38140030
#>  [73] -0.41647627  0.68094267 -0.41437304 -0.51834551 -0.68401973 -0.88564860
#>  [79]  0.04923707  0.18556122 -0.60865779 -0.73110285  2.71514421 -1.33938704
#>  [85] -0.64601525 -0.93245461 -0.76908693  0.37157978  0.35543278 -0.98399849
#>  [91]  0.21472959 -0.08008508 -1.42236955  1.10814896  1.07847764 -0.44025034
#>  [97] -0.77816901 -1.81859001 -1.12408090  1.06052384
#> 
#> [[6]]
#>   [1] -1.47870016 -1.55156017  0.77750668  1.06844014 -0.18358770  1.55824293
#>   [7] -0.21324238  0.93053526  0.41081180 -1.27984430 -0.78236663 -2.27608345
#>  [13] -0.12639591  1.44814671 -1.44275737  1.46718718 -0.74329990 -0.30422384
#>  [19]  0.33765806 -0.60750209 -0.29556027 -0.13453714  0.81478437 -0.27292173
#>  [25]  2.15948580  1.09173757  0.74338485 -1.20785935  0.32781805 -0.53416511
#>  [31]  1.28394672  0.02893134 -0.39567707 -0.69486833 -1.49208070  1.44425724
#>  [37] -0.34701739 -0.04025917  1.24663211 -1.34630152 -0.57390128 -0.70595925
#>  [43] -1.85740454 -0.24697048 -0.33556093 -0.24937687  0.45952256 -0.46007548
#>  [49]  0.05893280 -0.74259108 -2.16357400 -0.23289649  0.87490683 -0.97349403
#>  [55] -0.51274686 -0.92348799 -1.49442573 -0.16593036  1.58223832  1.01460227
#>  [61] -1.14647834 -0.90984911  3.10918268 -1.06955277 -0.70010952 -0.19844447
#>  [67]  0.74362012 -0.05015977  1.31314355  0.08601352 -0.62985137  0.65507033
#>  [73]  0.01939355  0.69626996  2.00886989 -0.05697828  0.24041550 -0.02189017
#>  [79] -0.83080816  1.40247432 -0.72386911  1.33058037 -0.81133302  1.79503611
#>  [85]  0.68662633  0.08937800  0.32454774  0.07131812  0.23304838  1.98637658
#>  [91] -0.95167079 -0.55864232  0.25581454 -0.19807617  0.42878914  0.07077011
#>  [97] -1.53665380  0.45398437  0.87439857 -1.29773757
#> 
#> [[7]]
#>   [1] -1.07896524 -1.13484994  0.02261431 -0.02146122 -0.27877244 -0.02420641
#>   [7]  0.82502030 -0.75018278  0.53621047 -1.57271748 -0.98642742  1.98351232
#>  [13] -1.85137844 -0.90978818 -1.95144002 -0.80027646 -1.86936160 -0.75057148
#>  [19] -0.59093839 -0.74209927  0.69351208 -0.05946397 -1.86389510 -1.27450892
#>  [25] -1.78177081 -0.50859339 -1.73687829  0.04048130 -0.12417946 -0.61275343
#>  [31]  0.16066248 -0.66235963 -0.33485166  0.62301159  1.02819429 -1.13457824
#>  [37]  0.91691105  1.21387479 -0.68593246 -1.09456819  0.36487410 -1.00801708
#>  [43]  0.55499644  0.43606592  0.10537330 -0.25373785 -0.97649123  0.39421777
#>  [49]  1.01490303  0.63144381 -0.48339825 -1.52633394  0.61684228  1.22928684
#>  [55]  0.15745303  1.40634206  0.39680092  0.32642296 -1.45934331 -0.79929553
#>  [61]  0.47001280 -1.15018052 -0.27159672  0.45742422 -0.01695794 -0.54159436
#>  [67]  0.87581342  0.74083815 -0.16481239 -0.75096428 -1.25336629 -1.03623626
#>  [73] -0.02943857 -0.27494378  0.48545147  0.91698203 -0.98534392 -1.49118975
#>  [79]  0.81959245  1.01340453  1.05360525 -0.07142864  0.95244461 -0.38403752
#>  [85] -0.41775999 -0.47026817 -1.69018029 -0.94549766 -0.34301196  0.35846241
#>  [91]  0.04818886  1.13017228 -0.52763231  0.47957533  1.41752930 -0.08009077
#>  [97] -0.39418700  0.52525764 -0.79593510  0.24509645
#> 
#> [[8]]
#>   [1] -0.77931984  0.59492387  1.10889090 -0.94296903  0.70034399 -0.42467483
#>   [7] -1.14313760  0.23344022 -0.18923658 -1.66341126  1.91569077 -0.81689444
#>  [13]  0.38365049 -0.45863225 -0.71570398  0.42954854 -1.16985274  1.32208831
#>  [19]  1.36596464 -0.24972371 -0.42472135  1.09843934  0.67092304  0.03095466
#>  [25]  2.36851196  2.58578811  1.21228095 -1.25119684  0.40026604  0.93470763
#>  [31] -0.47266985  1.58301322  0.59775422  0.46608459 -0.74436754  0.95905526
#>  [37]  0.15263708 -0.76936870  1.39308074  0.91866877  1.01608706  0.53463544
#>  [43] -0.98762856  1.18779227 -0.51754221 -0.25956025 -0.32806467  0.07343239
#>  [49] -0.24786302 -1.37386226 -0.04044582  0.42153824  0.20159751 -1.69719192
#>  [55]  0.64228768 -0.99523961  0.96381390 -1.65603723  1.07086109 -0.10902636
#>  [61]  1.89918639 -1.13703073 -0.27971976 -0.89412905  0.13670185 -0.74916542
#>  [67]  0.51819908 -0.19233721  0.02880981  0.35859089 -0.02899503  1.14704207
#>  [73]  0.37358894  0.32333921 -0.82981932  1.39446258 -0.19154358  0.27227702
#>  [79] -1.08165901 -2.32939936 -0.54962596 -0.07257999  1.03228840  0.21513853
#>  [85] -0.49434012  1.51213828 -0.57946326  1.67478963 -1.00098026  1.22270284
#>  [91]  1.07718817 -0.61194546  0.50686697  0.46006837  1.48439186  0.88196323
#>  [97] -0.53670224  1.28555371  0.58784853 -1.30848203
#> 
#> [[9]]
#>   [1]  0.316726325  1.194158303  0.913057148 -0.786752993 -0.410929700
#>   [6]  0.476373682  0.185915868 -1.324716253  1.172371055  0.231492772
#>  [11]  0.483071904 -0.535319580  1.378135819 -1.302671671  0.634857535
#>  [16]  0.999655161 -0.337748953 -0.086033607 -1.718817584 -0.929121572
#>  [21]  0.813676046  0.526413546  1.011957390  0.831379809  0.415134145
#>  [26] -0.661257061  0.548811286 -0.539367800 -0.171318426 -0.766734605
#>  [31] -0.606574846  1.647312821  0.841069348  0.338175097 -0.404883076
#>  [36]  0.900009039  1.190666655 -0.047539083  0.786925425  1.166684112
#>  [41] -0.271823871 -0.313704285 -0.286300150 -0.803440263 -0.094555455
#>  [46] -0.074046506  0.714490683  0.124718366  1.296594339 -1.114492926
#>  [51] -0.842058813 -1.504294209 -0.284026455  0.042869041 -0.008866413
#>  [56] -2.949083784  0.020358254 -0.098446980  0.589109098 -0.425751629
#>  [61]  0.634658985 -0.578524902 -0.169109048 -1.919232520 -1.534266387
#>  [66] -1.114761221  1.597811635 -0.639805078  1.566690553 -1.449155712
#>  [71] -0.791503959 -0.504480733  0.401826700  0.971396469 -0.579663020
#>  [76]  1.604179789  0.225973337 -0.514857122 -0.823788240  0.334415250
#>  [81] -0.093490765  0.304062078 -0.476507534 -0.241311717  0.824155833
#>  [86] -1.555643970  0.093501283 -0.366949421 -0.129408799  0.407795520
#>  [91] -0.583968954 -0.193449890 -0.269546670  0.073658232  0.357236138
#>  [96]  0.550428418  0.038401793 -1.609575292 -1.049710115  2.052033974
#> 
#> [[10]]
#>   [1]  0.176436411  1.128307358  0.435001744  0.548833492  0.647418304
#>   [6]  0.878463454  0.350797869  0.049879720  0.835749446 -0.281725199
#>  [11] -0.791679461  0.001653727 -1.187530126  0.362417465 -0.549435319
#>  [16]  0.692949838 -0.060843916 -1.193540930 -0.119908059 -0.708107950
#>  [21] -1.616663973  0.495847824  1.299750970 -1.615985506 -1.250367945
#>  [26]  1.582132306  1.252529118 -0.209395701 -0.639033687  0.398902937
#>  [31]  1.341973500 -0.037607422  0.543670114  1.740218394 -0.324455939
#>  [36] -0.447511333  0.416082526 -1.295865368  0.682633090  0.496913333
#>  [41]  1.439660656  1.207358558  0.238514694 -0.298837878  1.384621050
#>  [46] -0.700069022  1.860931509  1.803924959 -1.356043492 -0.314751072
#>  [51] -1.019265181 -0.612249656 -0.289068749  1.506983292  0.367722694
#>  [56] -0.020034941 -0.979921959 -1.400912333  1.445014910 -0.423481621
#>  [61] -0.733749597 -0.708484753  1.475092459  0.845004185  1.293994431
#>  [66]  0.298161142 -0.405682490 -0.138807578 -0.222592731  1.747150257
#>  [71] -0.081704668  0.124141586  0.392236141  0.378419970  0.166175673
#>  [76]  1.153387369  0.014748302  0.021440845 -0.826352170  0.047399133
#>  [81] -0.293107372  1.207632181  0.032001384  0.180397550  0.077160580
#>  [86]  0.558395338 -0.166854367 -0.227980619  0.242666056 -0.928482969
#>  [91]  1.515406678 -0.811598824  1.311247408  0.815566796  0.120032668
#>  [96]  0.881789848 -1.798743117 -0.957051078 -0.474058606 -1.860461283
#>