Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  0.93297738  2.54902257 -0.73124221 -0.38803238  1.69249697 -0.20276797
#>   [7] -0.83106098 -1.23489076 -0.82251538  0.24937661 -0.17736017 -0.24904001
#>  [13]  0.43602077  2.00438453  0.59870040 -0.74768164  1.54480916 -1.82589778
#>  [19]  0.70690513 -0.84340598 -0.94965983 -0.32191970 -0.29607190  0.68202827
#>  [25] -0.33966791 -0.98203515  1.77331590 -1.66207180  1.50846448  0.48350615
#>  [31]  0.73474990 -0.12704309  0.90684556  2.19754798  1.38329199 -1.41087017
#>  [37]  0.64208996  1.55336179  0.06791282 -1.57666159  0.05387579 -0.51513071
#>  [43]  0.11842988  0.87117661  0.44958246 -0.89956560  0.55547746  1.72346425
#>  [49] -1.09226207 -0.50560631  0.59586066  0.79250414  0.85633308 -1.02003105
#>  [55]  0.79713139 -0.95583745  0.18581351  3.00760514  1.98104694  0.91046310
#>  [61]  0.41569659 -1.38788956 -0.07995132 -1.33494834  0.78197551 -0.29909502
#>  [67] -0.38764223  0.75319604  0.20769409  0.14033739 -1.48963846 -0.01745870
#>  [73] -0.01000903 -0.78933929 -0.01612022  0.01581879 -0.15096918  0.19055739
#>  [79]  1.04569595 -1.23232090  0.52030885 -0.09540745  0.83860482 -0.34096240
#>  [85] -0.35461398  1.04316791  0.04515063 -0.75950095  0.56608316  1.24141726
#>  [91]  1.12240935  0.19420408  1.52343825 -0.90491899 -0.46838031 -0.65551125
#>  [97] -0.91747441  1.70190477 -0.35727731  0.69649150
#> 
#> [[2]]
#>   [1]  0.42126006 -0.99907836  0.77614578 -0.62790611 -0.56053841  0.06381193
#>   [7] -0.30338379  0.52681396  0.47883928  0.73539365  1.63141928  0.25887708
#>  [13] -0.63906561  0.33825300 -0.37593834  0.91317861 -1.37562964  1.15473927
#>  [19] -0.44124079 -0.26951063  1.39527028  0.21783090  0.28851149 -1.76614746
#>  [25] -1.79876413  0.06817029 -1.27664386 -0.90549299 -1.09647995  1.17907995
#>  [31] -0.58633942  0.08901683 -1.41167876 -0.59032087  1.06267803 -0.22469578
#>  [37] -0.38299157  1.33164912  1.26018766 -0.11943738 -0.36959411 -0.52558109
#>  [43] -0.70126791 -0.70091358 -0.28970002 -1.25685040  1.15237186 -0.67956040
#>  [49]  0.81276088 -0.91804752 -0.54045538  0.45119894  1.40556640 -1.13270008
#>  [55] -0.77280214  0.35849364  0.35805349  0.68736022 -0.96920151 -0.70468438
#>  [61]  0.22093924  0.42955480  0.15565639 -0.78498852  0.84579649  1.54229972
#>  [67] -0.79822440 -1.78581718 -0.24930279 -0.68532443 -1.14140488 -1.19889925
#>  [73]  1.14644157 -0.90396935 -0.41892724 -0.18144512  0.61823159 -0.05258628
#>  [79] -0.33917258 -0.67222432 -0.67610171  1.26942573 -0.27828026  0.38467592
#>  [85]  1.36284716  0.80502101 -0.59223664  0.54586788  0.92023360 -1.60719705
#>  [91] -0.26813138  0.79386687  1.24516627 -0.47587323 -0.32154686  1.78200828
#>  [97]  0.66096576  0.62800804  0.66145601  2.70813725
#> 
#> [[3]]
#>   [1]  0.075584189  0.679880020  1.127880784 -0.464788885  1.170544581
#>   [6] -1.914903140 -0.848548696  1.100682281  0.431504460 -0.234362506
#>  [11] -0.862150112 -0.252356515 -1.292242177 -0.446287207 -0.443394619
#>  [16]  0.783118528 -1.777104008 -0.512237268 -1.079325291 -0.100249627
#>  [21]  2.496355329  0.255068581  0.323489010  1.622325202  1.389036242
#>  [26] -0.011460522 -0.707714185  0.013289695 -0.671252041 -1.015526525
#>  [31] -0.448078810 -0.137602022 -0.230268443 -1.727925255 -1.289682037
#>  [36]  1.690670608 -0.973612537 -1.026069093  0.320480368 -0.096250105
#>  [41] -0.309125303 -0.600495692  1.463503186  0.627348670  0.109351942
#>  [46]  0.820216550  0.519526389 -0.763615735 -0.869901421 -1.032465764
#>  [51]  0.742050369  1.068559866 -2.675016366 -0.755678934  0.849356891
#>  [56]  2.244069363  1.052942531  0.853540098 -0.628998002 -1.345533343
#>  [61] -0.903854495 -0.047728550  0.447946453  2.759473570 -0.164849286
#>  [66]  0.001298251  1.341488290  0.157771831  0.791941880 -0.778295295
#>  [71]  0.383313869  0.318727404  0.416175311  1.138074752 -1.505607563
#>  [76]  0.814917408  0.471233222 -0.876562992  0.310039558  0.287758657
#>  [81]  0.088928535  1.778845082 -2.038927027  0.081795721  0.015714929
#>  [86]  0.412347586 -1.529711712 -0.507627389  0.756518586 -0.446221098
#>  [91] -1.111814333  0.593029166 -0.132424102  2.196558620 -0.730155574
#>  [96]  0.269237287 -0.662921706  1.178645196 -0.248844466  0.312886743
#> 
#> [[4]]
#>   [1] -1.13169625  0.27983432  0.21523178  0.86351955  1.73912912  0.43903048
#>   [7] -0.97751568 -0.29029813  0.64091244 -0.19645103 -0.46757026  0.74669215
#>  [13]  0.80597712  0.02520511  1.22453846  0.34561081 -2.31501894  0.47232672
#>  [19]  0.13034097  0.50951412 -1.72430166 -1.06583657 -0.06955257  1.87074826
#>  [25] -1.67102190  0.40496825 -0.79209151 -0.17575807 -0.77672225  0.93186232
#>  [31]  0.49047742 -0.60621414 -0.94078871 -0.69139111 -0.85221310 -1.95647116
#>  [37]  1.34233490  1.35116883 -0.91306111  0.70256861 -0.06311069 -0.51144375
#>  [43]  0.84673384 -1.16690789 -0.14956758  1.34056183  0.25854281  1.86443829
#>  [49] -0.89888047  0.45120020  0.87040260 -0.66555066 -0.70383943  0.32764928
#>  [55] -0.27676750 -0.36552524 -1.04474530 -0.21502804 -1.15654324 -0.01988041
#>  [61] -1.41824453 -0.30054967 -1.12296489  1.09917193  0.66810951 -0.92771117
#>  [67] -0.60662153  0.77249874 -0.89846628 -0.15088115  0.39981089 -0.24845957
#>  [73]  0.71899935 -0.28406021  0.39687143 -0.31704237  0.07493508  0.45812239
#>  [79]  1.00188937 -1.46087581 -0.46869592 -1.02346187 -0.69117723 -0.23928863
#>  [85]  1.11829097  0.87624041  0.19527208 -1.10057548  0.68694984 -0.20197443
#>  [91] -1.18518068 -0.09119031  0.17694977 -0.33170831 -0.49364175 -0.81663426
#>  [97] -0.02214870 -0.20212609  1.15416362 -0.53431948
#> 
#> [[5]]
#>   [1]  0.349176071  0.559897884  0.241188306  0.397797091 -1.033183749
#>   [6]  1.063136676 -0.621130125  0.165790537 -0.332804488 -2.205040982
#>  [11] -1.685354662 -1.497432909 -0.541293271 -0.799924543 -1.922983431
#>  [16]  0.284378848 -0.762051691 -1.502337370 -0.564059777  0.163354747
#>  [21] -2.084662777 -0.040834659 -0.863213224  0.780580112  0.843745072
#>  [26]  0.158871485  0.058106139  0.257660760  1.509804867  1.077366646
#>  [31]  1.462720175  0.726134581 -0.002898603  0.474571721  0.511094001
#>  [36] -1.597724379 -1.360617238 -0.147311354  0.830987838 -0.938147618
#>  [41] -0.991037830  1.624787586  0.213007667 -1.965009899 -1.848794280
#>  [46]  0.976306702 -1.172471433  0.763686725  1.828207650 -0.015694810
#>  [51]  0.140726552  0.408983021 -0.700422885 -0.107488604 -1.338371350
#>  [56]  0.447197981  0.484746851  0.816089002  0.867163317 -1.065764033
#>  [61]  0.870400877 -0.467748596 -0.326233411  0.045379566 -0.530276790
#>  [66] -0.718725978 -0.410162103  1.261201204  0.591638691 -0.280983039
#>  [71]  0.893048030  1.122440468  0.705197600 -0.501651983  1.537933095
#>  [76]  0.838401853  0.410563598  2.278517948  0.131350032 -1.089618085
#>  [81]  1.875127421  1.662751834  1.328738123 -1.128492076 -0.437029227
#>  [86]  1.602912657  0.893896135 -1.301520658  0.783193453  0.451918342
#>  [91]  0.462403587  1.197721781 -1.353533014  1.080611861  1.157107029
#>  [96] -0.542181304 -1.250385183 -0.588349155  1.637235530  0.288334341
#> 
#> [[6]]
#>   [1]  1.276893903 -0.164380698 -0.152411589  1.127386105  0.215506877
#>   [6]  1.304181720  0.624570020  1.416932749  2.836022506  0.032467662
#>  [11]  0.276444571 -0.946234284 -1.199369217 -0.639856335 -2.137097357
#>  [16] -0.487405370 -1.678365451 -0.576084693 -0.219628747 -0.451803361
#>  [21]  0.335135528  0.070307820 -0.731144581  0.588944479 -1.068344183
#>  [26]  1.192901626  0.294442292  0.832705457  1.480067616 -1.164840598
#>  [31] -1.942664066 -1.331623420 -0.023094167 -0.320215065 -1.485878453
#>  [36] -0.187380891  0.064750142 -1.175676111  1.250813510 -1.661689258
#>  [41]  1.424170431 -0.506171323  0.184555686  1.042669544 -0.170964163
#>  [46]  1.427671106 -1.035446877  0.456652425  1.833143900 -0.112108127
#>  [51] -2.777199943 -0.944288045  0.375453237 -0.082548483  0.210151374
#>  [56] -0.680922137 -1.482078810 -0.438070996 -0.233200312 -0.614708032
#>  [61]  0.411464007 -0.869280714 -0.139463431 -0.130541094  0.275026054
#>  [66] -0.398161658 -2.601414625  1.268709925  0.904441925  0.867445649
#>  [71]  0.072596513  3.290470549 -0.513660620  0.587708694 -0.264553143
#>  [76]  0.004029164 -0.166023379 -0.719093465 -0.822963775  0.927093746
#>  [81] -0.860314302 -1.005565806  1.293789260 -0.197649597 -0.082301258
#>  [86]  1.168591172 -0.084158061  2.344280857  1.428951545  0.305028915
#>  [91]  0.706225957 -1.441780302  2.043208486 -2.518444433  0.605289007
#>  [96] -0.716211216 -0.571818945 -0.818848009 -0.363196319 -0.908227823
#> 
#> [[7]]
#>   [1] -0.55942835  0.65737468  1.26687004  0.24154903  0.78691396 -0.69602524
#>   [7]  1.14977553 -0.84936998 -0.74464751 -0.18718991 -0.29447345 -0.80320479
#>  [13]  0.73288781 -1.04289343  0.26947571 -0.80732939  0.16011333 -0.51567007
#>  [19] -1.10837443  1.16582171 -1.05380569  0.17110120 -1.25755719  1.11684118
#>  [25] -1.03577765 -0.77031293  0.30746714  1.09496387 -0.58449286  1.16061218
#>  [31] -0.32191316 -0.19744557 -0.99925007 -0.08096678 -0.89882774 -0.04680552
#>  [37]  0.94212006  1.44452031 -1.24322524  0.31545153  0.53554992  0.23860966
#>  [43]  0.85249392 -0.88872007  1.09672394 -1.66283777 -0.16813073  0.49699539
#>  [49]  0.96794026 -0.38720738 -1.06149426  0.84166733 -0.55170870  0.41665353
#>  [55] -0.39735228 -0.36597587 -0.65839514 -1.46948677  0.34761855 -0.97766122
#>  [61] -0.16795764  0.72656204  1.22574119  0.69278771 -0.69953821 -0.17699830
#>  [67] -0.24208857 -3.09639940 -0.65346559  1.99487386  0.68890642  0.88477554
#>  [73]  0.58782268  2.02740900 -0.14313074  1.60494825  0.67612429 -0.41656547
#>  [79]  0.23654169  1.37625105 -0.11694461  1.20321936  0.70932837  1.79873404
#>  [85] -0.37096709  0.04521814 -1.17498737  1.13980614  0.95744615  0.27164131
#>  [91]  1.21190309  1.17107058 -0.51714337  0.29782466  0.21116751  0.19816170
#>  [97] -0.19782234  0.11177164 -0.80495740 -1.59401284
#> 
#> [[8]]
#>   [1] -0.40352574  0.38466161 -0.50468038  0.60681213  0.08132276  0.42370276
#>   [7] -2.28518767  1.28778497  0.35135831  0.07001312  1.31437294 -2.34663200
#>  [13] -1.55706254 -1.12331254  2.15068762  0.72366804  0.55498685 -1.24364793
#>  [19] -1.35351739 -0.86094052 -1.15054862  0.55465992  0.74250189 -0.31423963
#>  [25] -0.16536801 -0.33003291 -0.50795044  0.13429650  1.75381170 -1.21809885
#>  [31] -1.11524568 -0.66013000 -0.79419359 -0.37798157 -1.40773694  0.40947962
#>  [37] -2.10430284 -0.87116184  1.64027826 -2.28064191 -0.16869382  0.00889695
#>  [43] -0.81634741  0.09933712 -0.51311513  0.46671229 -1.67810246  1.27544082
#>  [49]  0.80423474 -0.51383436 -0.78720877  1.33280919  1.49502884 -0.16313235
#>  [55] -1.00115144  0.43420780  0.12875659 -0.16901613 -1.12312214  0.16573769
#>  [61]  0.13544754 -1.09424587 -1.02063362  1.07976929 -1.99012584  1.39129276
#>  [67]  2.06620033  1.86082248  0.33380107 -0.74556859 -1.04532980  0.88788696
#>  [73]  0.54696728  0.53319714  1.30833234 -0.22214628 -1.85584076 -1.81802380
#>  [79]  0.47614217 -1.09194078 -2.11128208 -0.88395471  1.02927332  1.70457805
#>  [85] -0.22836089 -0.46309547  0.18479830 -1.04090933 -0.24562612 -1.22082764
#>  [91] -0.72077946 -1.55772894 -0.49486626 -0.02819557 -0.41083819  0.85066870
#>  [97] -1.19501466  0.69011466  0.40800688 -0.47992303
#> 
#> [[9]]
#>   [1]  1.8633411157  0.9097199779 -2.0414032836  0.2078106633 -0.4445964697
#>   [6]  0.6339017424 -0.2076697565 -1.3723369076 -0.7224447149  0.4750549386
#>  [11]  0.8974536854  1.0056099278 -0.2489033410  1.9875024936 -1.4085643999
#>  [16] -0.5041670754 -0.2678480142 -1.7690982237 -0.0102613484 -1.0802481625
#>  [21] -0.1854568672  1.1437072764 -0.9823383759  1.1935169222  0.0400721832
#>  [26] -2.3167397960 -1.2698885365  0.7614117789  0.8149192868  0.1587592362
#>  [31]  1.1523353038 -0.4284867487 -0.0004546045 -0.1584226357 -2.1316766874
#>  [36]  0.2438155855  0.3916911007 -0.1403786161 -0.1426226217 -0.5085524108
#>  [41]  0.8286499034  0.4577761656  1.5510662425  1.5415365250 -0.7014028118
#>  [46]  0.2332240196  0.3758240037 -1.1556601473  2.0863220349  0.2128247935
#>  [51]  0.0541447789 -1.4079837253 -0.0920057500  0.5145700537  0.1628930598
#>  [56]  3.1703326237 -1.4163914577 -0.2769521014 -0.4348527831  0.5989365316
#>  [61]  0.0128126617 -0.0574731261  1.1929610043  0.2661120130 -1.1631773423
#>  [66]  0.7837440539  0.0192317511  0.2883839320  0.0154098796  0.8432516460
#>  [71]  0.8687501655  0.9295638847 -0.0167584941  0.4091405612 -0.3260569191
#>  [76] -1.1163030474  0.3975777579  1.2542697456  1.0343622832 -1.0396765373
#>  [81]  0.0529357187  0.3764271766 -0.0568863052 -1.0884840762  2.0171918382
#>  [86] -1.0586842673 -0.9965884300  0.5138288868 -1.1285767142 -1.0249075508
#>  [91]  0.4936851094  0.4034999856  1.5051733611 -1.1813714175  1.7773422119
#>  [96] -2.2200412695 -1.1652686239  0.2801027660 -1.1131400991 -0.2044339901
#> 
#> [[10]]
#>   [1]  0.750802509  1.521069660 -0.919600775  0.577718083  0.516284228
#>   [6] -1.063740357  0.579562584  1.489275068  0.348851102 -0.887318218
#>  [11] -0.222563395  0.480152889  0.179219897  1.226057924  0.051210427
#>  [16]  1.006051651  0.049103371  0.265250890 -0.890589167  0.148409086
#>  [21]  0.773025044 -0.837655874  1.272357736 -1.267799000 -0.153712045
#>  [26] -1.389230815  1.460966214 -0.506198253 -0.585759090 -0.263854861
#>  [31] -0.323365311 -1.401186030 -0.556212807 -0.330069478 -0.690050641
#>  [36]  0.086497313 -0.597637494  1.207664061 -0.494062514  0.763225969
#>  [41]  1.133243547  0.119994934 -0.342727200  0.811389480 -0.568902311
#>  [46]  1.649224832 -0.533867636  1.849034508  0.063719780 -0.362576100
#>  [51] -0.129210907  0.004375417  1.749490328  0.100286360  1.216161973
#>  [56]  0.651566358 -1.210769168 -0.914359622 -0.221070248 -0.063256624
#>  [61]  1.402983540 -0.268044569 -0.729164174  1.319297393 -1.255146426
#>  [66] -0.241016304 -0.416423024  1.176148895  1.732936211  0.812830938
#>  [71] -1.316037233  0.431260345  1.865735752 -0.764959554 -0.374158924
#>  [76] -0.404012495 -0.286794306  1.593151789 -0.741994234  0.609892135
#>  [81]  1.410763218 -0.316253145  0.014210040 -0.161386973  2.157195688
#>  [86] -0.309638111  0.051024224 -0.477205640 -0.596695912 -0.720996890
#>  [91]  0.400338725 -0.235111435  0.337622640  1.746006716 -1.332644513
#>  [96] -1.300279902  0.668514820 -1.254217975 -0.023092469  0.774691554
#>