Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  0.485138014  0.237866095  1.307371557  1.200021420 -0.290995613
#>   [6]  0.224794361  0.401952386 -0.591396046  0.331402021 -0.574786313
#>  [11] -1.344325892  1.043639551  0.164086986 -0.104812599 -1.435956277
#>  [16] -0.882949046 -0.136177409  1.806862113  0.598096979 -0.646981827
#>  [21]  0.313484370  2.405795238  0.664825420 -0.353881267 -0.519646470
#>  [26]  0.361060734  1.552528519 -0.337030021  1.887089639 -1.468584511
#>  [31] -0.239340885 -0.021897660  0.608280033 -0.776320116  0.445508324
#>  [36] -1.893347290  0.724938749  0.075016757  2.000354249 -1.458561543
#>  [41] -0.702318169  2.399513809  1.500758844 -0.568087310 -0.339405287
#>  [46] -0.216315173 -0.519922698 -0.891193918  1.257527118 -0.693033339
#>  [51]  0.178203665 -0.136688018  0.477015533 -1.740120419 -0.247099084
#>  [56] -1.335323524  0.931644261 -0.622573066 -1.429578910  0.166272229
#>  [61] -1.198922907  1.407463910 -0.079590078 -0.304647645  1.430559314
#>  [66] -0.451516472 -1.163884453 -2.943838773 -1.299400950 -0.854390496
#>  [71] -0.315183275 -0.257396889  0.395209269  0.633123395 -0.157495360
#>  [76]  1.230856483 -0.656730069  1.206874399  2.641140672  0.005287525
#>  [81] -1.316199246 -1.249683074  1.985401526  1.969996530 -2.097523860
#>  [86]  1.542006007  0.007265891 -0.776491800 -0.278553156  0.366752144
#>  [91] -0.077228838 -0.378258786 -1.598815131  0.277391767 -0.912718056
#>  [96] -1.091076781 -0.903036313 -1.898475281 -0.640039969 -2.751874170
#> 
#> [[2]]
#>   [1] -0.459446624  0.324128691 -0.340276659  0.976832897 -2.013752795
#>   [6]  0.437546045  0.039460886  0.692380982  1.008712999 -0.377808357
#>  [11] -0.274056715 -0.485781211  0.976327426  0.443353584  0.175227509
#>  [16] -1.762746339 -0.652535069  0.483614477 -1.315490679  0.944334480
#>  [21]  0.877500365  0.764522624  1.100550939 -1.294809568  0.910110173
#>  [26] -0.883547921  2.426519301 -0.912973153 -0.351787355  0.089447659
#>  [31]  0.609383215 -0.293370374 -0.633992333  1.721956204  0.232745535
#>  [36] -0.714390981  1.972280925  0.633180087 -0.268462704  1.412354715
#>  [41] -0.001981771 -0.687621401 -0.158991893  0.547844744  0.135447729
#>  [46]  0.593486387 -0.334324594  0.123059122 -0.748853062  1.588214758
#>  [51]  0.667297400  2.056710107 -0.221293807  0.047867365 -0.354508380
#>  [56]  0.725527233  0.419191074  1.393577195  0.049917212  3.357044838
#>  [61]  0.051283097 -0.146718966 -0.173727137 -0.042394321  0.281309409
#>  [66] -0.278187452  0.988976984 -0.217121979 -0.812759370  1.571190361
#>  [71] -0.209704922  0.522588534 -0.326842045 -0.185174539 -0.873973882
#>  [76] -0.330715257 -0.226140264  0.364922676 -0.793250341 -0.297584671
#>  [81]  0.183436928 -2.043468669 -0.335636188 -0.207255744  1.703270207
#>  [86]  0.543589568  0.134126951 -1.154658630  0.201300619 -0.939199027
#>  [91] -0.010846439 -0.816302408  1.154868860  0.464507508  0.403975784
#>  [96] -0.931504604  0.097741485  1.563778878  1.279152318  1.124818636
#> 
#> [[3]]
#>   [1]  0.47866613 -0.35024332  0.37990986  0.03173632 -0.74625199  1.11017999
#>   [7]  0.50622648 -0.71900671 -0.34722678 -1.30146116 -0.83384170 -1.85085097
#>  [13] -0.64011647 -0.09901543 -0.06877289 -0.64078532  1.35113168  0.22352599
#>  [19] -0.70002799 -0.42014593  0.38375682  0.79195270  0.84130117 -0.56408222
#>  [25] -1.62661192 -1.14508458 -0.59510289 -1.45503324 -0.01250168 -0.47086938
#>  [31] -0.27620887 -0.44820067 -0.19538842  0.88195870 -0.85327185 -2.65515754
#>  [37] -0.50808531 -0.78920367 -0.35174029 -0.25047895 -3.06345358  1.08312571
#>  [43]  0.96527491 -0.38098396 -0.05311327  0.93482009 -1.52193528 -0.42646411
#>  [49]  0.62336539  1.66542576  0.46824407  0.19065817 -0.41698275 -0.36346912
#>  [55] -1.15151241 -0.36742152  0.27290706  0.27694812 -0.55687829 -0.48210085
#>  [61] -1.87744666 -2.25743025 -0.50849821  0.41110112 -0.65052819  0.89616456
#>  [67]  0.21972077  0.68346750 -0.64002703 -0.60106268 -0.88795188  0.87759178
#>  [73]  0.27662583 -0.78148040 -0.36524373 -0.05975164  0.17156013  1.11761294
#>  [79]  0.44417873 -0.58852500  0.09896127 -0.18219615  0.41438529  0.17796980
#>  [85]  0.15977700 -0.50872146  0.42436125  1.56072875 -0.08417436 -0.21255559
#>  [91]  0.29931444 -0.24640670  0.24087967 -0.63355037 -0.33804280 -0.37707124
#>  [97]  0.87536933 -1.14604506  0.65790626  0.29286283
#> 
#> [[4]]
#>   [1] -0.505487252 -0.633825052  1.016691807 -0.241994687 -1.336481859
#>   [6] -0.404864037 -1.318967390  1.420760435  0.019360680 -0.148460918
#>  [11]  0.507779393  0.772986835 -2.490367939  0.118573677 -2.814066502
#>  [16]  1.342288278  1.313848519 -0.424875022 -0.186548425 -0.385038168
#>  [21]  0.598544873  0.390428009 -0.514568537  1.180114129 -0.577125862
#>  [26] -0.365008673  0.025810327  1.435358200  1.133020868 -0.129164557
#>  [31] -1.613679265  1.403793054 -0.441155629  0.436249230 -0.497736966
#>  [36] -0.226342917  0.155051339 -0.394992979 -0.468214549  0.133062191
#>  [41] -0.510948523 -0.017416281  0.575007169  0.291781708 -0.851308132
#>  [46]  1.809734203 -1.258620100  0.263811237  0.592579545  0.676117469
#>  [51] -0.913266113  0.474332996 -0.546363972 -0.512196644  0.172478486
#>  [56] -0.443790821  0.140594794 -0.210540943  0.846903860 -1.297666986
#>  [61]  0.088570255  0.937297517  0.745776621  0.355917159 -0.938753422
#>  [66]  0.564100361  0.278723882  1.065726915 -1.286853831 -0.563949706
#>  [71]  1.150128736  0.339520674  0.506362669  1.744368864 -0.902125083
#>  [76] -0.169778681  0.998783650 -1.305345991  0.001670463 -0.615919642
#>  [81]  0.898174022 -0.578405596 -1.166021166 -0.064660246  0.208718912
#>  [86] -0.396260766  1.251232967  0.964386463 -1.305937860 -0.122995735
#>  [91]  0.553792061 -0.425273217 -0.639822979 -0.152957160  0.379360019
#>  [96]  0.472128259  0.762293871  0.316646506  2.211294676 -0.316060167
#> 
#> [[5]]
#>   [1]  0.09057180 -0.46824884 -1.09036203  0.54478022 -0.48387951  0.08122852
#>   [7] -0.34153961 -1.99569669  1.47543657  0.03215331 -0.71649190 -0.83420218
#>  [13]  1.62540507  0.54219989 -0.84404329  0.12851783 -1.03656683 -0.15384387
#>  [19]  0.38086723 -1.41737161  0.48308091  0.32356393  1.14112121 -0.90191579
#>  [25] -0.35684061  0.59726744 -0.13894970 -0.50651119 -1.21376724  0.54852792
#>  [31] -2.32292246  0.54811178  1.21114135  0.13145080  0.07090045  1.36588979
#>  [37] -0.15597619 -0.38192398 -1.45309829  0.52073993 -1.09144532 -1.19488675
#>  [43]  0.30665164  0.45703526  0.52345699 -0.15523407 -0.93776614 -0.01083002
#>  [49] -1.48533661 -0.69374492  2.54667908  1.21325656 -0.56648640  0.62626972
#>  [55] -0.48675735  1.33092194 -0.09825294  0.77532065 -1.00541779  1.63131209
#>  [61] -0.38064505 -0.51361576  1.63682834  0.06234024 -0.28267811 -0.41302953
#>  [67]  0.16077104  0.03814784 -0.03207538  0.80708903 -0.19881476  0.82633651
#>  [73] -0.87464183  0.53540536 -1.38678845 -0.89483815 -0.43179437  0.70769572
#>  [79]  0.66054040  0.60979165  0.71579456  0.46240983 -0.38781359 -0.66292687
#>  [85]  1.14759639  1.27712997 -1.47925260 -1.66133623  0.60855833  0.18732854
#>  [91] -0.02833203  1.34728229 -0.66386197 -0.17510257  0.57481004 -0.29205061
#>  [97]  2.34716509 -1.22574023  1.62867079 -1.10077724
#> 
#> [[6]]
#>   [1]  0.64303974 -0.75436282 -0.10447806  0.27495459  1.06153646 -0.45488334
#>   [7]  2.47253071 -2.41195575  1.54811846 -0.21157880  1.59423688  0.62416628
#>  [13] -0.62216857  1.11128923 -0.49837564  0.85292961  0.41038363  0.13123508
#>  [19] -1.00865251  0.36497493  0.11149009  0.21749620  1.66174343  0.04618288
#>  [25]  1.53137909 -0.15761341  1.55479028 -1.17533815 -0.41412007 -0.64244084
#>  [31] -0.49756379  0.17240869 -0.54578648  2.36185159  0.83163484  0.91395511
#>  [37] -0.51472687 -0.48790740 -0.81895606 -0.52741731  0.06770856 -0.59898196
#>  [43]  0.09307723 -0.15894235  1.63193753  0.74584353  0.09002120  0.68473553
#>  [49]  0.73559087 -1.93270433 -1.12940101  0.21592985  0.39394282 -0.54024585
#>  [55]  1.27206616  0.09405075  1.16473315  0.37534210  0.13039241  1.24566739
#>  [61] -0.69513050 -0.40887236  0.23416763 -0.49852516  1.91253768 -0.15529203
#>  [67] -1.26070773  0.87385264 -0.78908321 -0.13851643 -1.65639618 -1.32587325
#>  [73] -0.66705614 -1.11746624 -0.26416918 -0.70523436  0.01309828 -1.24601969
#>  [79]  0.62864323 -1.38785210 -0.85401728 -0.08889215  0.98995257  0.33860593
#>  [85]  0.25493112 -0.24241125 -1.43968052  1.31486512 -1.21880292 -1.82893445
#>  [91]  1.51151550  0.36992562 -0.38065247  0.09387208 -0.37707955  1.90405916
#>  [97]  1.43511558  1.16940770  1.81111923  0.49570153
#> 
#> [[7]]
#>   [1] -0.389957321  0.628840075 -0.390104239  1.179787535 -1.586226158
#>   [6] -0.311353675 -1.188135964 -0.293997586 -0.428066587 -0.157275343
#>  [11]  0.282179455  0.340361105  0.352260082  0.868400243  1.205571000
#>  [16]  0.171309812  0.818465606 -0.114759191 -0.239613140  0.209041526
#>  [21]  2.061860870 -1.705381240 -1.384835773 -1.578298739  1.313838104
#>  [26]  0.432758017  0.878838723 -0.150410442 -0.860990325  0.344229052
#>  [31] -0.615557948 -0.755484040  0.724140930  0.759546249 -0.895315945
#>  [36] -0.034349238  0.063562265 -2.239134075  1.325966742 -0.367600559
#>  [41] -1.660837307  0.523549234  0.399802289 -0.001340836  0.062516820
#>  [46] -0.912919634  0.136229017 -1.074694354  0.809376344  0.956680265
#>  [51]  0.718499082 -0.659461349 -1.245074933 -2.615559494  0.101572041
#>  [56] -0.830264374  0.733261792 -0.847780276 -0.483201695  0.226645197
#>  [61] -1.372592436 -0.254456721  0.250590206 -1.291285170 -0.005181780
#>  [66]  1.912745428 -0.245255686 -0.430600500  0.988866966  0.236077952
#>  [71] -0.285534660  0.716756235 -0.859718137 -1.285001219  1.044656866
#>  [76]  0.304041209  0.186448455  0.500569956 -2.224216149  2.856015415
#>  [81] -0.571369839  0.616873897 -0.402151615 -0.069422539 -2.588073167
#>  [86]  0.894512945  0.572580781  1.615620258 -0.903355574  0.429489980
#>  [91]  0.051027956 -0.861951818  0.452869947  0.721707549  0.046374233
#>  [96] -0.512309869  0.502994996 -0.760871972 -0.908568739 -0.817593190
#> 
#> [[8]]
#>   [1]  1.95723472  0.34186386 -0.07499524 -0.86984485  0.24454034 -0.54741267
#>   [7]  0.21062148  0.36240131  0.08987109  1.50985629  0.80622205  0.05876796
#>  [13] -1.22734903  1.07387822 -1.16589018 -0.29089456  0.72294794 -0.71326608
#>  [19] -0.34817531  0.27240847  0.76235224 -0.46806861 -0.19037040  0.25714516
#>  [25]  0.72450602 -0.70936089 -0.49897196  0.93868931  0.95155611 -0.13511871
#>  [31]  0.50040180 -0.07108151 -0.08862171 -0.21391492 -0.31721666 -0.71535717
#>  [37] -0.62840540 -0.57201530  0.32907877 -1.37025259  0.25586724 -0.85695127
#>  [43] -0.58573843  1.82779355 -0.70558406  0.85597431  1.06961055  0.60656164
#>  [49] -0.93284904 -0.44303189 -1.84944273 -0.25414191  0.30416454  0.51744156
#>  [55] -1.77742568  0.57662044 -2.44110844 -1.70923880 -1.18882580  1.81003539
#>  [61] -0.66512199  0.30359672  1.64296491 -0.83014334 -0.59186642  0.45667091
#>  [67]  0.84951774 -0.40887053  0.81379766 -0.41015504 -0.48039759 -0.55656289
#>  [73] -0.34786451  2.08115435 -0.59898423  1.38402642  1.89895050 -0.86447415
#>  [79]  0.09797187  0.07034918 -0.16359562  1.93014606 -1.11172542 -0.31845429
#>  [85]  0.74934521  0.43291697 -0.43334507  0.98273351 -0.92375531  1.49611888
#>  [91] -1.10643374  1.40521862  1.80489837  1.70872150 -1.13892800  1.62954296
#>  [97]  1.27137807 -0.17049101 -1.68221073 -0.95929454
#> 
#> [[9]]
#>   [1]  0.731905472  0.627010460 -0.544943110 -0.767828332  0.690518387
#>   [6] -0.124707689  0.161732570  0.654634844 -0.332301529  0.468187243
#>  [11] -0.021165252 -0.741342227 -0.105011719 -0.326403731  0.108763046
#>  [16]  0.192208331  0.762424073  1.133361831  1.980785128 -1.641787945
#>  [21]  0.142737595  1.358748384  2.060206984 -0.863402096 -0.259934027
#>  [26] -1.507418931  1.744833705 -0.217207965  0.475721295  0.448781993
#>  [31]  0.767693860  0.075374733 -1.294735487  0.408578470  0.118562587
#>  [36] -1.854284965 -1.737637458  0.669065997 -0.595643611  0.327815351
#>  [41] -2.315471389 -0.153005851  0.553564570 -2.086783696 -2.736772302
#>  [46] -0.630452213 -1.978840655 -0.625546649 -0.124832511  1.051239687
#>  [51]  1.466775328 -0.306614732  2.192129707  0.096505024  0.520272204
#>  [56]  0.608222125 -1.794811666 -1.134931097 -0.542543579  1.641786549
#>  [61] -0.652394651  1.207252801 -1.523036201 -0.739968022  0.145080312
#>  [66]  0.414152465  1.388689959  0.982320671 -1.106664882  0.163844080
#>  [71]  0.634724028  0.799897876 -1.075782189  1.101991289  0.674838782
#>  [76] -0.648583314 -0.622836285  0.333174664 -0.548675844  1.014467126
#>  [81]  3.532623652  0.916702948  1.472748453 -0.654638009 -0.747134040
#>  [86]  0.424800456  1.000690992 -0.008490132  0.280967605  0.035538020
#>  [91]  0.414844870  0.141031006 -2.613896820  1.332134557  1.166098857
#>  [96]  1.352103574  1.445956904 -1.459525307 -0.853872146 -0.417920561
#> 
#> [[10]]
#>   [1]  7.119500e-01 -5.384515e-01 -1.225418e+00 -7.263703e-01 -1.799739e+00
#>   [6] -6.007222e-01 -4.097406e-01 -8.697026e-01  8.931720e-01  9.994651e-02
#>  [11] -1.030433e+00 -2.005332e+00  5.413882e-01  1.676064e-01 -1.317060e-01
#>  [16] -1.632644e+00  2.361381e+00 -1.162165e+00 -1.341624e+00  4.189786e-02
#>  [21] -8.882076e-02  9.538080e-02  2.303019e-01 -3.053491e-01  1.807520e+00
#>  [26]  1.500346e+00  7.397009e-01  3.718740e-01  1.234429e+00 -5.321783e-01
#>  [31]  7.737359e-01 -6.134440e-05 -3.230249e-01  2.313831e+00  1.218280e+00
#>  [36] -6.939957e-01  1.038128e+00  3.371936e-01 -3.880261e-01  2.097110e+00
#>  [41]  9.855338e-01 -9.103535e-01  6.429093e-01  8.050003e-01 -1.081518e+00
#>  [46]  7.577603e-01  3.539210e-02 -5.251593e-01 -6.021752e-01  1.449184e+00
#>  [51] -1.580085e-01 -1.018353e+00  6.426220e-01 -1.752687e+00  5.566698e-01
#>  [56] -1.157541e+00 -3.239635e-01  3.526309e-01 -1.095484e+00  4.770006e-01
#>  [61]  6.050185e-01  4.185304e-02  2.496467e-01  1.502104e+00  1.305351e+00
#>  [66] -8.568439e-01  1.715253e+00  4.743208e-01 -1.776683e-01 -7.742901e-02
#>  [71] -4.573285e-01  1.104540e+00  1.933463e+00 -6.354822e-05  9.579511e-02
#>  [76] -9.369508e-01  2.715327e-01 -1.199541e+00 -6.044138e-01 -1.063897e+00
#>  [81] -7.358611e-02  3.009020e-01 -2.323927e-01 -1.111295e+00  1.042693e+00
#>  [86]  2.511882e-01  8.887527e-01 -9.521221e-01  1.824860e-01 -1.637586e+00
#>  [91] -7.951438e-02  7.927211e-01  1.055496e+00 -7.492103e-01  3.323605e-01
#>  [96] -7.442764e-01  1.139054e+00  1.113961e+00  1.928216e-01 -3.637049e-01
#>