Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object

Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name

Name of variable to extract when object is a stanfit object.

n_samples

How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))
#> [[1]]
#>   [1]  0.57483406  0.41755445 -1.05791696 -0.73487808  1.09639104  0.12612628
#>   [7]  0.61568659  0.97879611 -1.13658357 -0.09566206  1.15419660  0.41361054
#>  [13]  1.44044335  1.32821230 -2.29675965  0.49034719 -0.29819054  0.13828357
#>  [19] -1.49090688 -0.09764307 -0.59231104 -0.52300771 -0.05147419  1.24666581
#>  [25]  1.89926946 -0.99994041  0.36424430 -0.83711261 -1.03243957  0.16129069
#>  [31] -0.92519057  0.01001403  1.26599304  0.59054799 -0.21378772  0.75258847
#>  [37]  0.97482065 -0.71055565  0.44638714  0.89581496  0.77372440 -0.10347194
#>  [43] -0.09586341 -1.46057923 -1.06852044 -0.50046537 -1.19946821  1.05902253
#>  [49] -0.41643900 -0.05942823  0.52650559  0.50133748  0.35981689 -0.27674618
#>  [55] -1.26699418  0.66140479 -1.17801451 -1.42363120  0.15126797 -0.26741729
#>  [61]  1.73282520  0.53581030  0.56457197  0.52833090 -0.56825700  1.43095686
#>  [67] -0.55369108  0.82433351  0.84354694 -0.59372956 -0.71477326 -0.27855874
#>  [73]  0.25766603 -2.40080818  0.16804988 -1.08015411 -0.02602141 -0.88995619
#>  [79]  0.12683191  1.11495971 -1.07049549 -2.30339135  1.65182845 -0.47764335
#>  [85]  0.42330173  0.90803225  0.89533589  0.74359884 -0.06802675 -0.45997723
#>  [91] -0.23871054 -0.39786080  0.85294166 -1.24482648  0.45925514  0.06326784
#>  [97]  1.65312768  1.18040198  0.03106259  1.36416363
#> 
#> [[2]]
#>   [1] -2.080597757 -1.488729634 -2.728339657 -0.217097110 -0.795327383
#>   [6]  0.160283340  0.448092882 -1.521885090  0.569771193 -0.895016826
#>  [11]  0.861999478 -1.603219352 -0.322833441 -2.946449687  0.153883276
#>  [16] -0.044755083  1.549704043  0.580171284  1.468612002  3.113210114
#>  [21] -1.803001726 -0.136187920  1.470319015  0.721506055  0.320377463
#>  [26] -0.404363227 -1.236301974 -0.099903819 -0.655713624  1.490589994
#>  [31]  0.342433443 -0.778620255  0.190900064 -0.040996891 -0.020479857
#>  [36]  1.122766055 -0.850483515 -1.748008935  0.942497563  0.358710861
#>  [41]  0.758940993  0.877882513  0.161064651  0.587031648  0.273009742
#>  [46] -2.301910815 -0.101591124  0.524764594  0.726388425  0.292857969
#>  [51] -1.003181302  0.003850233 -0.257914402 -0.100788415 -0.869495954
#>  [56] -0.474125134  0.831734280  0.160241004  0.594571883  0.391119497
#>  [61] -1.831405393 -1.313709304 -1.395092556 -0.776371153 -0.962881315
#>  [66] -0.236185242  1.350894293  0.698396457 -0.342984726  0.855744892
#>  [71]  0.516357985 -1.089530606  1.800813313  1.849093455 -0.213490957
#>  [76]  1.412493319  0.981903335 -3.388648145 -0.261162885 -0.382086008
#>  [81] -0.360236510  0.876173078 -0.705931835 -0.273815043  0.749259524
#>  [86] -0.494828672 -0.604085837  0.500639326 -1.027569868  0.997875195
#>  [91]  0.039826489  0.727378595 -1.684535355  1.427396444  1.348576022
#>  [96]  1.421412251 -3.781066678  1.049503154  1.636398465  0.497381112
#> 
#> [[3]]
#>   [1] -0.318988726 -0.328234108  0.931400860  0.975179462 -0.604503086
#>   [6] -1.207006486  0.145563251 -1.476393664  0.399059492  0.022123195
#>  [11] -0.398126968 -1.089155564 -0.654665583  0.776599876 -1.089559254
#>  [16] -0.462901074 -2.361899677 -0.052040781 -1.041844419 -0.764573786
#>  [21]  1.020805298  0.660960560 -0.627676524  0.079323407  2.391741113
#>  [26] -0.006564852 -0.392092042  0.829085935 -0.784526936  1.307920798
#>  [31]  0.036995830  1.452995818 -0.163421560  0.505520615  0.096919797
#>  [36]  0.177479033  0.401515795 -0.538393319  0.112744864 -0.242886229
#>  [41] -0.161872162  0.299031707 -0.035968643  0.337958013 -0.129434734
#>  [46] -0.134159078  2.065204723  0.038861735 -0.333620914  0.465280325
#>  [51]  1.006569248 -0.932395955 -0.640758948  1.726321647  0.110541904
#>  [56] -0.659402493  0.846062096 -0.395979933  0.808263941 -2.146703446
#>  [61] -0.434716843  1.150655203 -0.762631404  1.600694456  0.349310691
#>  [66] -2.016859008 -1.107541366 -0.613897467 -1.602059273  0.507109355
#>  [71]  0.901707626 -0.594420856 -0.935702006  0.345373961  0.811491905
#>  [76] -1.247801947 -1.057576504 -0.309587486  1.004566112 -0.672763391
#>  [81] -0.277296968 -1.052151983  1.289207062  0.846774898 -1.811643144
#>  [86]  1.261068283 -0.407353080 -0.819192823  0.782119043 -0.105123388
#>  [91] -2.001678146 -0.979400254 -0.502156220 -0.117606278  0.173574680
#>  [96] -0.482447477 -0.498629676  0.300159992  0.506008059 -1.213278315
#> 
#> [[4]]
#>   [1] -0.987176456  1.593893059  0.530284510  0.738886924 -0.294283594
#>   [6]  0.407473368  0.134282427  0.312052425 -0.251993609  0.471189874
#>  [11]  1.269993467  1.284526710  0.601619582 -0.357150560  0.498319942
#>  [16] -0.400269551 -1.244673670 -0.735417071  0.984332427 -0.623145057
#>  [21] -0.341832993 -0.290464066 -1.064146657 -0.912643617 -0.047461667
#>  [26]  1.212722405 -0.892994009 -0.370463704 -0.619466147 -1.618847818
#>  [31]  0.002168322 -0.802302791  0.600082786 -1.760393308  1.252427730
#>  [36]  0.232961309  1.613351775  0.548967834  1.417701944 -0.079313313
#>  [41] -1.015119620 -0.790414069 -0.598543226 -1.945209235 -1.666300513
#>  [46]  0.192921837  0.224887774  0.033987377 -0.797256354 -0.204125668
#>  [51]  1.572753357 -0.427221124  0.669894298 -1.396930084 -0.079714193
#>  [56] -0.341815792 -0.321877902  0.311404306  0.406338974 -1.379994054
#>  [61]  0.915717661  1.153549074 -0.168714347  1.181746978  1.419771327
#>  [66] -0.851413424 -0.601355817  0.039454994 -1.986737955  0.152288371
#>  [71]  0.906735044  0.152245110  0.413260097  0.111005470  0.730938707
#>  [76] -1.417552223  1.791681318  1.737891284 -1.133410255  1.658712318
#>  [81] -0.207406845  0.390752157  0.054599695 -1.104390921 -0.857979553
#>  [86] -0.568631418 -0.292641305 -0.690921000 -0.061446678  0.284258297
#>  [91]  2.092415482  0.427306162 -0.987260263  0.102290310  0.439495850
#>  [96]  0.450870931 -0.306482743 -0.307849535 -0.268215921 -1.049247465
#> 
#> [[5]]
#>   [1]  1.338783849 -0.090864236 -0.516683036  0.301795756  0.737865665
#>   [6] -0.270007264  1.231910313  1.251412011 -0.072056533 -0.094376196
#>  [11]  1.483363519 -0.026768224  0.277322104 -0.009083241  1.549557454
#>  [16] -1.656598692  2.027278205  0.158285383 -0.582325671  0.689589075
#>  [21]  1.928233234  0.532820750 -0.015528717  0.087705543  0.909403206
#>  [26] -0.121608652 -0.212722429  0.291745836 -0.026472256 -0.962521369
#>  [31]  0.434669227  0.158117111 -1.396546843  1.260790419  1.007008813
#>  [36] -0.673878734  0.313939883 -0.581783570 -0.079226702 -0.911432482
#>  [41] -0.835942879  0.617607905  0.782105970  0.422847251 -0.481833950
#>  [46]  0.035056474 -0.303142748 -0.049032729 -1.424453727  1.512828934
#>  [51]  1.498587210  0.813815854  2.727843418 -0.946156829  0.286239898
#>  [56]  1.378093510 -1.470602061 -0.416968999  1.601067208  0.569953710
#>  [61]  0.400746124  0.984644919  0.270942150  0.301159658 -0.532869233
#>  [66]  1.212070512 -0.061193908  0.615651729  0.711178936  1.612479350
#>  [71]  1.240316283 -1.624897195  1.708249659  0.671611916 -1.572168465
#>  [76] -0.131802061 -0.607117546 -1.712846872  0.208088691  0.266927102
#>  [81] -1.141757770 -1.422840909  2.133852618  1.131149971 -0.362894518
#>  [86]  1.905737829 -0.275903782  0.699003412 -0.903562313  0.854312626
#>  [91]  1.647361869 -0.227195390  2.222021131  1.049424052  0.132465760
#>  [96]  0.552960436 -0.278622777 -1.162258628 -1.530650666 -0.676732408
#> 
#> [[6]]
#>   [1] -0.888653611  0.707321098  0.404173854 -0.823149360  1.044888892
#>   [6]  1.147822481  0.978505729 -1.239447176 -0.930140729 -1.041257730
#>  [11] -1.215168460 -1.585717189  1.423449351 -0.956355332 -1.026513125
#>  [16] -1.918450271  1.034733345 -2.082718020 -0.671541974 -0.794311362
#>  [21]  1.279871919  1.179120092 -0.680758820 -0.795628363 -0.851130741
#>  [26] -0.814464994  0.696648537 -1.814802855 -0.634718511 -0.278632138
#>  [31] -0.325019738 -0.494149697 -0.002100727 -0.315818061  2.245201514
#>  [36]  1.881575194 -1.489738107  1.531265015  0.772854991  0.724761944
#>  [41]  0.513969884  0.003942222 -0.578007266  0.777191992 -0.361146265
#>  [46] -1.048801320  1.769070575 -0.230120927  0.637208154 -0.967709416
#>  [51]  0.230639613  1.474183464  0.901773971  0.796800924  0.175369860
#>  [56] -0.853597831 -0.960698018  0.499161725 -1.310210328 -1.564261704
#>  [61]  0.349256200 -0.546610032  0.332803408  0.244209773 -1.258720312
#>  [66]  1.041565820 -1.152787599 -1.264891059 -0.869541532  0.561905830
#>  [71] -0.907984311  0.769574603  0.451469669 -1.744561598 -1.549528372
#>  [76]  0.159890382 -0.113306576  0.100048602  0.674521653 -1.069357685
#>  [81]  1.758641599  0.461645866  0.204190514  1.639283935  1.213423806
#>  [86]  1.218579261  0.863110754  1.543004609  1.224181076 -0.313584433
#>  [91] -1.591548228 -0.076244205 -0.255886399  0.115142167 -1.150880983
#>  [96] -1.594690726 -0.407275127  0.210156516 -0.463722346 -0.559519429
#> 
#> [[7]]
#>   [1] -0.4039482221 -0.5715042851 -1.7839867923  1.3894479499  0.4261276275
#>   [6] -1.1028934171  0.2398097031 -2.1261573749 -1.0565023719 -0.0614353915
#>  [11] -1.8334435388  0.1687770620  1.3931318383 -2.2670040946 -0.9000924701
#>  [16] -0.4641769952  0.2240786649 -0.4176135183 -0.2332136808  0.5897951021
#>  [21] -1.1579057516 -0.1566283870 -0.5221527933 -0.0145825156  0.1514574968
#>  [26] -0.6892664031  2.2527765845  0.5207312846  0.1994804502 -1.3512970548
#>  [31] -0.5083728285  0.4330818802  0.5470924541 -0.4189084830  1.6698693230
#>  [36]  0.3359011426 -0.2690674344 -0.5060088206  0.4817001248 -0.1666733724
#>  [41] -1.4545965570 -1.2316757868  0.4038904554 -0.7527445376  1.0667638819
#>  [46] -0.5777490834  1.9609941945  0.8452672718  0.8654646304 -0.0004504431
#>  [51]  0.7614823199 -0.2475232103 -0.0598485206  1.0912856123 -0.4107607486
#>  [56] -0.4623472069 -0.7510450534 -0.0270815454  0.0189878705 -0.4419031535
#>  [61]  0.5097306543  0.1576669176  0.2776530574  1.1541630004  0.4467210013
#>  [66] -0.4415285091 -1.8748190504 -0.2198307969 -0.6045791876 -0.3466069234
#>  [71] -0.1541250371 -0.0740820574 -0.7394184395 -1.7204342524  0.8786209545
#>  [76] -0.1841288692 -1.6365911229  0.1135966681  1.0404408753 -0.4265990082
#>  [81]  0.1743253528  0.7736938524  1.3183646545  0.4681243243 -0.1627555728
#>  [86] -0.9938002931  0.7358727835 -1.5490098263  1.5618622446  2.2494119889
#>  [91] -0.2603178741  1.5736751449  0.3685775439  2.3149791662 -0.0144445624
#>  [96] -1.0145490178 -0.3595348132  0.3009951149 -0.3871191073  0.7932447358
#> 
#> [[8]]
#>   [1]  0.61374299  1.74172399  0.60165426  0.32050693  0.16379280 -1.07265203
#>   [7]  0.21256739  0.05658113 -0.88313852 -1.58175839 -0.97032148  0.91224933
#>  [13] -1.22326387 -0.85521006 -0.07052058 -0.24611400  0.59457023 -0.38606849
#>  [19] -1.04846827  2.17381578  1.66018210 -0.21770282 -0.02837013  0.48265594
#>  [25] -1.18357326  0.93143695  0.97976126 -0.51802983 -0.42722566  1.81115101
#>  [31]  0.11008916 -0.50758649  0.79184388 -0.61795889 -0.76757635 -1.20760274
#>  [37]  0.27817430  0.43524874  0.93863777  0.53696635 -0.71217880  0.46371856
#>  [43] -2.59327560  0.11033063 -0.62947874  1.14273232  1.44703798 -0.09904297
#>  [49]  0.31609228  1.73071372  0.73040627  1.37806791  1.37258046 -0.48104539
#>  [55] -0.81625697 -0.62587182 -0.02152121  0.91854283  0.06923353 -0.68575186
#>  [61]  1.41983309  0.94320220 -1.08921945 -1.31774695  0.33150709 -0.46975135
#>  [67] -0.17784927  0.31579367 -1.66720795  0.95039695  0.76195891 -1.15470215
#>  [73] -0.89130275 -0.33356178  0.14487897 -0.71040528  0.26485774 -1.04680411
#>  [79]  0.04987776 -0.92039069 -0.05137750 -0.12969349 -0.97489165  0.59885888
#>  [85] -0.47794633  0.26362691  0.63808790 -0.74112954 -0.03776536 -0.22861897
#>  [91]  0.60538122  0.71431819 -1.20647412  1.15713377 -0.13775326 -0.44912327
#>  [97] -0.57453750 -0.81525526  0.88500422 -0.16251514
#> 
#> [[9]]
#>   [1] -1.56157012 -1.32968399 -0.27975154 -2.63751818  0.05424088 -0.83459010
#>   [7] -0.62702813  0.87618426 -1.26916946 -1.45248808 -0.82198136  0.52220027
#>  [13]  1.59975916 -0.89079429  1.16314754 -0.17321167 -0.79752304  0.28631695
#>  [19] -0.67600410  0.26992151  0.01872893  1.62295947  0.04320126  1.17450410
#>  [25]  0.10285053  0.74451030  1.49795271 -1.05634598  0.75144107  0.47289861
#>  [31]  1.43894785 -0.42245339 -0.66599865 -1.27230132  0.47406676 -2.39675220
#>  [37]  0.45837384 -0.27637942  1.02128533 -0.72751352 -0.13979223 -0.34426704
#>  [43] -0.48721761  1.59042329 -0.79531409  1.18155851 -1.32689715 -0.61317218
#>  [49]  1.37108287  1.17850083  0.69771034 -0.97032089  1.00471983 -0.96356517
#>  [55] -0.83176192  0.63196206  0.04089220  0.85277487 -0.24025282  1.73437814
#>  [61]  0.30951888  0.89769888 -0.68028312 -0.88129965  0.73287224 -0.08372474
#>  [67]  0.58638426 -0.39547077 -1.97441200 -0.94137756 -0.64736565 -0.15115298
#>  [73] -0.48997409 -0.75099002  0.35929863 -0.31600886 -0.29726848 -0.05381180
#>  [79] -1.31368176 -1.22608213  0.30493214 -0.76975748  2.10226973 -0.92541334
#>  [85]  0.31755846  0.49632108  1.28083447 -1.17881309 -0.87409129 -0.69768074
#>  [91]  2.63071322  1.19293220 -0.74608790  0.61262253 -0.77486006 -1.31024640
#>  [97] -0.35133200  0.34929488 -1.49499910  1.01650352
#> 
#> [[10]]
#>   [1] -0.344896392  0.491004715 -2.238009407  1.505573833 -0.044979939
#>   [6]  0.293067403  0.013328364 -0.065777726 -1.832914088  1.809579191
#>  [11]  1.519497873  0.082879000 -0.068860965 -0.531249970  2.263271723
#>  [16]  0.469847080 -2.016337593 -1.526256254  1.065136561  0.011355148
#>  [21]  0.898457231  1.638929741  0.408220076  1.021374191 -2.021668712
#>  [26]  4.013985486  0.659478648  0.702234277 -0.950046039  0.504773521
#>  [31] -0.007792198 -1.024507195  0.500931593  1.669525285 -2.356794652
#>  [36]  0.796919184 -0.176812603  0.752720441 -1.136498066 -0.055606683
#>  [41]  0.001591786 -0.799919772  0.615858079  0.606681215 -0.633526545
#>  [46] -0.517332943 -0.534123365 -0.986210811 -0.671440133 -0.103258211
#>  [51]  2.176423583  0.987556987 -1.523123246  0.793745858  0.308987599
#>  [56]  0.191602639 -0.447754133 -0.425404656 -0.899382398  0.796028224
#>  [61] -0.006983876  1.081306244  0.032677200  0.161043104  0.232876268
#>  [66]  1.771714791 -0.238664291  0.924627041 -0.740544138 -1.640483904
#>  [71]  0.309110693  0.566256615 -1.148494297 -0.335155136 -1.661265641
#>  [76]  0.329350894 -1.447347387 -2.095130823  1.246199413 -0.626423589
#>  [81]  0.344213580  2.005461390  1.580290944 -0.097440601  0.205465277
#>  [86]  2.789949656  2.095717679 -0.577448266 -0.222025919  0.982974494
#>  [91]  0.645755825  1.197934302 -0.436111075 -0.444328720  1.075155702
#>  [96] -1.269459945 -0.199293534  1.994378114  0.160780600  0.845034083
#>