Process samples to a list that can be included to a dataframe

Process samples to a list that can be included to a dataframe

samples_to_list(object, par_name = "", n_samples = NULL)

Arguments

object
Stanfit object containing variable par_name or matrix with rows representing samples, columns representing variables.

par_name
Name of variable to extract when object is a stanfit object.

n_samples
How many samples to return. Default (=NULL) to all samples.

Value

List of vector of samples

Examples

samples_to_list(matrix(rnorm(1e3), nrow = 1e2))

#> [[1]]
#>   [1] -0.94630671 -0.07517996  0.18792650  1.24910764 -0.54872690 -1.71033113
#>   [7] -0.70098598  1.89782704  1.11426860  0.84696175  0.39087260  1.05365534
#>  [13] -0.55548426 -0.62941268  0.30336454 -0.03803220  1.76272950  0.61462929
#>  [19] -1.66058679 -1.04845809 -0.92936450  1.54100663  0.17928638 -1.34494038
#>  [25] -0.74890428 -0.98942495 -0.84796718 -0.91029604  0.40996180 -2.82005204
#>  [31]  0.32521078 -0.82326653 -0.09100570  0.49405516  1.04161559  0.09354927
#>  [37] -0.52194444 -0.50994176  0.57820522 -2.06716149  0.53088578 -1.79744671
#>  [43] -0.32449648 -0.02025971  0.49643123 -0.31644161  1.66884891 -0.14304240
#>  [49]  1.59574669 -0.11143778 -1.08400007 -0.23867534 -1.95882984  0.61909327
#>  [55]  0.33412612 -1.04205906  1.03925451 -0.66561101  0.80347337 -1.51731972
#>  [61] -0.83830985  0.39554618  0.24454313  2.61729695 -0.73787645  0.58422841
#>  [67] -1.23887256 -0.89009968 -1.51002537 -0.58518240 -1.55367584 -0.57609515
#>  [73]  1.22697496  0.46124464 -1.07496099 -0.24960903 -0.48845389  0.46626270
#>  [79]  0.22335490 -0.56576449  0.12761785  0.36803004  0.60893511  0.91460695
#>  [85]  0.49576725  0.37704158 -1.14818247  0.94169946  0.99879796  0.73649980
#>  [91] -1.52710698 -1.35143765 -0.74458649  0.15558838 -0.74563933 -0.01795545
#>  [97] -0.02165123 -1.95809100  1.70021562  0.55418426
#> 
#> [[2]]
#>   [1]  0.976761067 -0.301992872 -0.398904868  0.410035337 -0.075537200
#>   [6]  1.296061452  0.866002423 -0.649405332 -0.991951077 -0.078909951
#>  [11]  0.598054622  0.302688695  0.731362882 -0.861133085 -0.678175782
#>  [16] -0.234640819  0.676094735 -0.066852462 -1.316840474 -0.290607212
#>  [21]  0.048993175  0.953771109  1.471338123  0.782765578 -0.362841824
#>  [26]  0.592050971  1.129756397 -0.372378665 -0.377801927  1.329455226
#>  [31] -0.771358832  0.518812002  0.106576734  0.581726895 -1.666407282
#>  [36]  0.821326926 -0.809716528  2.059445870 -1.029270464 -0.130562707
#>  [41]  1.872110296 -0.730017510 -1.279434102 -0.868359781  1.192816352
#>  [46] -1.444344362  0.255458729 -0.079880854  0.783372145  1.107952079
#>  [51] -0.452654070 -1.401415632 -0.509990723  1.562905383 -0.347334211
#>  [56]  0.683714301  0.825479660  0.311709224  1.763198672  0.261834809
#>  [61]  0.138689204 -0.345026689  0.737805137  0.330902565  0.106067409
#>  [66] -0.878674610  1.463621162 -1.121818156  1.205909866  0.590022411
#>  [71] -0.535741379  1.187797648  0.551681421 -2.138907220 -0.792653848
#>  [76] -0.695657374  1.094752616 -0.100362614  1.374677921  0.359881635
#>  [81] -0.911636110  0.372154103  0.824123998 -0.277860876  0.290705252
#>  [86] -1.462234408 -0.936234046 -0.005393725  1.146549255 -0.300266206
#>  [91] -1.672400721  0.342697385  0.004421065 -1.329155221 -0.755702197
#>  [96] -0.401535183 -0.142446497 -0.614792116 -1.294828736 -1.256173491
#> 
#> [[3]]
#>   [1]  0.48084755 -0.25453096 -0.69820946 -1.47512092 -0.81827145  2.08320146
#>   [7]  0.26754333  0.71818536  0.70988511 -2.38630624  1.40426569 -0.20961072
#>  [13] -0.99024937  1.96196662  0.59404591 -0.28169565 -0.56878596 -0.06509551
#>  [19] -0.96074408  1.16969956  0.01865945  1.85665562 -1.11318445 -0.25276798
#>  [25] -1.39607224 -0.61915155  0.26331401 -1.40401480  0.82375153 -0.08501601
#>  [31]  0.64938082  1.01145141  1.62993341  0.85528502 -0.30187573 -0.39575701
#>  [37]  0.25298543 -0.08484486  0.64585268 -1.38928361  1.48373145 -0.06463873
#>  [43] -0.80526765 -1.93488547  0.08484413  0.51687622  1.41375161  0.39713983
#>  [49]  0.37010424  1.73923589 -0.68153988 -1.11248045 -0.50197427 -0.45545620
#>  [55] -0.95503454 -0.34784849  0.22522393 -0.59512055  1.36536633  1.74617818
#>  [61] -0.17434665  1.00339677 -0.48560012 -0.52588290 -0.13652250  1.08675927
#>  [67] -0.34185392  1.77385232 -0.17168849 -0.87486044 -0.89005250 -1.11555464
#>  [73] -0.40113332  0.23996381  0.65854431  0.85166108 -0.62901011 -0.79454969
#>  [79] -0.40371944  1.30141076  1.20973243  2.28758844  0.70336470 -0.60386586
#>  [85]  1.02624779  0.50741909 -0.87608115 -0.71209045 -0.60493192  0.33180791
#>  [91] -0.38419582  1.86682384  0.05620488 -2.37549193 -0.05972797  0.42387511
#>  [97]  1.41780153  1.51697497  0.95264669 -0.50883594
#> 
#> [[4]]
#>   [1] -0.93151956 -0.89403525  0.70194970 -0.29554239  0.35112089  1.03655147
#>   [7]  0.28401640 -0.49001972  0.51862789  1.84030336  1.01635979  0.51043708
#>  [13] -1.98510890  2.17144117 -0.62395461  0.93994245 -0.42718539  1.61985553
#>  [19] -0.28783710 -1.09983951  0.30297204 -0.74738746 -0.62417622  0.40981039
#>  [25]  0.26364199 -0.32743474 -0.26446903  0.33028409 -0.27123434 -0.55850463
#>  [31] -0.48729364  0.91727594  0.46176051  2.40768979  0.68675647 -1.25265217
#>  [37]  0.12767029 -0.88804576  0.93349971  0.93675224  0.49051353 -0.07953660
#>  [43]  0.74045414  0.24437098 -0.06103008 -1.55866809 -0.18457723 -0.88789202
#>  [49]  0.62763447  0.78208595 -1.70301997  1.30401652 -0.94560273 -0.29895192
#>  [55] -0.11270688 -0.50888259 -0.64262955  1.05189685 -0.98010827  0.57785397
#>  [61]  0.27472747  0.28393422 -0.79814933 -0.10147950  0.73936886 -1.94578575
#>  [67] -0.44878154  0.19015021 -1.48560480 -1.53219433 -0.55113202 -0.16504810
#>  [73] -0.79828398  0.92665042 -2.11625115 -0.97677426  0.03431268  0.45629474
#>  [79] -0.09319871  0.06120260 -1.22382708 -0.17700586  0.15860971  2.64891611
#>  [85]  0.15454608 -0.12561159  0.69867088 -1.39446424 -0.39846920  0.46239462
#>  [91] -1.17945094  1.39089009 -0.60361621  0.70969501 -0.82249442  0.24869571
#>  [97]  0.83113548  0.96840035  0.25300953 -0.27684488
#> 
#> [[5]]
#>   [1] -1.544614428  0.794946417 -1.402200428 -1.467966833  0.421302478
#>   [6]  0.779037110 -1.345856750  1.319267714  0.360042283 -1.925003834
#>  [11] -0.336588980  1.011115197  0.374007495  0.662637365  0.536609454
#>  [16] -0.072412805  0.621193471  1.714475232 -0.590259789 -0.706321721
#>  [21] -0.489620045 -0.794092503  0.413360256  0.441552620 -0.649977395
#>  [26]  1.007883142 -0.102131690  0.321177482 -0.399044703 -1.420341143
#>  [31] -0.906551132  0.827854480 -1.311273204  1.069182031  0.121174474
#>  [36]  1.575295720  1.586886046 -1.461250052  0.975184288  1.114554944
#>  [41]  2.617780809 -1.769955151 -3.237400211  0.307004133 -0.174265752
#>  [46]  0.657495410 -0.008831523  0.421837595  0.650005772  0.390177072
#>  [51] -0.768557294  0.856942396 -0.353348987  0.406633202  0.182891462
#>  [56] -1.876715379 -0.510741518 -0.158000116  1.215151659  0.042218893
#>  [61]  0.397407043  1.290503157  0.866671362  0.616154225 -1.482557543
#>  [66]  0.136187368  1.294988672 -1.290650016 -0.977093403  1.580500626
#>  [71] -0.198718784  0.178222356  0.942686623 -0.723284390 -0.396482359
#>  [76]  0.315830272 -1.304693990 -1.809377201 -0.918340185  0.328061194
#>  [81]  0.197851881 -1.143879686 -0.064507457 -0.711701969 -1.948634720
#>  [86]  0.305012422 -1.510894658 -0.488629527  1.941073936  0.088127170
#>  [91] -0.471546085  0.356480834  2.437968092  0.842094303  0.838628315
#>  [96] -0.177798768 -0.337828787  1.883786434 -0.115525568 -1.072619499
#> 
#> [[6]]
#>   [1]  0.17367338  0.60140013  0.09680069 -0.34519317  0.36897584 -1.53920452
#>   [7] -2.56588512  0.44741348 -0.45457056 -0.06052660 -0.58955641  0.25064740
#>  [13]  0.13026258  0.20961119 -1.26331084 -0.66348099 -0.44400557 -0.58994731
#>  [19] -0.50944690 -0.31905239  0.94005678  0.54591913  0.88730254 -0.57680266
#>  [25] -0.65227600  0.58632743 -1.59076603  0.60360250  0.76580104  0.79827653
#>  [31] -0.09239831 -0.93231744  0.30235499  1.58908520  0.91385911 -0.47625433
#>  [37] -1.53308454 -1.74222897  0.18903617  1.20928247 -0.20397042 -1.55747651
#>  [43]  0.37162199  0.45953996 -0.86277594  0.67292393 -0.59329735 -0.13489906
#>  [49] -0.04633516  0.35199693  1.62280594 -0.79464507 -0.23049986 -1.44351177
#>  [55] -0.52425207  1.26628943 -0.71617199  0.14548212  0.96053098 -0.67424691
#>  [61]  1.12467803  2.47293735  1.74020335 -1.53075669  1.41702094 -0.39685031
#>  [67]  0.48999401  0.71862163 -1.78788581  1.20675202 -1.57149222  1.10356699
#>  [73]  1.72887445 -0.11600011 -0.07319157  0.30214692  0.14651204  1.50528851
#>  [79]  0.07467346  0.40495924  0.13197059  0.79605215  0.35968998  0.71575058
#>  [85]  0.11974905 -0.72601810 -0.27974141 -0.57414146  0.52048222 -1.23750182
#>  [91] -0.74541211  1.58557275 -0.35586285 -1.51305956  1.51557414  0.84428005
#>  [97]  0.33901072  0.49973128  0.87800578 -0.82969225
#> 
#> [[7]]
#>   [1]  0.43109539 -1.25098050  0.68328345 -0.79931099  0.40124381 -0.02780726
#>   [7] -1.30972238 -0.70708721 -1.46737704 -0.08421625  0.05817145 -1.41343575
#>  [13] -0.22945194  0.66127757 -0.97509056 -0.69556901 -1.28258947 -1.30746432
#>  [19] -1.50847617  0.19102599  1.61651866 -0.19085495 -0.03855950  0.56202508
#>  [25]  0.64327098  0.30449537 -0.39741918  0.37146595  0.96195119  0.71031599
#>  [31] -1.14549838 -0.98185265 -0.92117508 -0.06650559 -1.31892585 -0.58514071
#>  [37] -0.26759452  0.23861358 -0.31664116 -1.28955774  0.90584001  0.02976631
#>  [43] -0.34927184  0.58408005 -0.27064906 -0.59792633  1.11260749 -0.27365534
#>  [49]  1.20093166  0.36723092 -0.74724124 -1.41487085 -0.93108864  1.43418487
#>  [55]  1.44503438 -1.02784713  0.27790238 -0.39761916  0.25374031 -1.26047622
#>  [61] -0.11465603  1.39502319 -1.47615111 -0.08441358  0.54650824 -0.82054972
#>  [67]  0.52879615 -0.22545194  0.15845059 -1.75429575 -0.38745576  0.75426858
#>  [73]  0.63401004  0.35567270  0.86643619  0.27126674  0.63315895  0.78528378
#>  [79] -1.39431105 -0.12315526  1.65625632 -0.58826145  0.72493595 -1.04506180
#>  [85]  1.65335447 -1.09639133 -0.51611585 -0.26687012  0.63262447 -0.17403882
#>  [91] -0.39223928 -0.33492180  0.26835997  0.52749529  0.68708263  0.83221930
#>  [97]  0.45659361  0.24648807  0.32907091  0.29811925
#> 
#> [[8]]
#>   [1] -0.49550869 -0.13439721  0.00860995  0.81829294 -0.89852773 -0.72530289
#>   [7]  0.43793313 -0.83728747 -0.80291812 -0.06528758 -1.32739903 -0.69440341
#>  [13] -0.64270406  0.41587004  0.94736761  0.10276019 -0.54905999 -0.92219076
#>  [19]  0.20959217  0.11715161  1.34632425 -1.24548202  0.20879001 -0.54308209
#>  [25]  1.39308285  1.05017220 -0.10679694 -0.58353832 -0.31568020  0.47630777
#>  [31] -1.10421098  0.24695358 -0.48491616 -0.61965897 -0.18863221 -0.90766227
#>  [37]  0.34789709  0.43510151 -1.31491747 -0.39270947  0.87004058 -1.09073773
#>  [43] -0.59826094  0.88819035 -1.37891353 -1.65180584 -0.56060808 -2.17404911
#>  [49] -1.06159591  1.04091878  1.06595335  0.69454028  0.13909976 -0.20109284
#>  [55]  0.03390428  0.36396599  0.12470718 -0.56587277  0.14183925  0.35387328
#>  [61]  0.71957479  1.16804358 -0.86386058  1.09143123  0.44329681 -2.56252354
#>  [67]  2.28456615  0.06785592  0.37321284  0.12176869  0.58237050  1.08279222
#>  [73]  0.72016215  0.08659465  1.39385444 -0.29576449 -0.51116471  0.37175370
#>  [79]  3.07708815 -0.84177348 -0.75735158  0.22551531  0.82101143  0.44300601
#>  [85] -0.79665963 -0.54537668  0.54329763 -1.66100366  0.52384943  1.33077248
#>  [91]  1.61164823  0.07202230 -0.71075675 -0.87130672 -0.30749825 -1.73808897
#>  [97]  1.04042798  2.25679194  1.74195148  0.11156808
#> 
#> [[9]]
#>   [1]  0.778078319  1.293373040 -0.291159884  0.531275075  0.055874846
#>   [6]  2.098537411  0.777969786  0.017406662  0.909998561  0.054319704
#>  [11]  1.710472073  0.661729231 -0.610819108 -0.102528388 -0.769892028
#>  [16] -1.250920393  0.569002047 -0.611318615  0.397500945  0.896814349
#>  [21]  0.005403033 -0.850009853  1.241061065  0.364704370 -0.775399745
#>  [26] -0.395000655 -3.262717401 -0.915951797  1.264955073  0.096894386
#>  [31] -1.509469689  0.063721662 -2.142291259  0.349268423  1.634436182
#>  [36]  2.003644354  0.220709202 -0.765586295 -1.118195566  1.020963508
#>  [41]  0.880848356 -1.647363168  0.240867062 -2.122455318  0.113138265
#>  [46]  0.258181355 -0.579197348 -0.042879056  0.754832074 -0.167650813
#>  [51] -0.752456624 -0.592449106 -1.321808972 -0.457432639  2.902705707
#>  [56]  0.619973237 -3.518039134 -0.043689976  0.257642438 -0.226239008
#>  [61]  0.407938692  0.076403371  1.386345070  0.934222641  1.301377332
#>  [66]  0.327919606 -1.160049579  0.754031493  1.093314539  1.223281845
#>  [71] -0.416236099 -0.663272138  0.991758613 -0.189276256 -0.363977548
#>  [76]  0.175347405  1.209704754  0.325911866 -0.109459095  0.033975836
#>  [81] -1.217289646  0.511670185  1.170941387  0.344549912  0.549838755
#>  [86] -0.956127628  0.307160555  0.091512083 -0.936362357  1.905773930
#>  [91]  0.333696024 -0.446588845  0.337141604  1.008309684  2.433366614
#>  [96]  0.685266453  0.289611476 -0.135677606 -0.153306466 -1.016513973
#> 
#> [[10]]
#>   [1] -0.06133107 -0.40519219 -0.68370077  0.73367431 -1.68757778  0.33437046
#>   [7] -0.54043706  1.02046305 -1.00433366 -0.11780586  0.49529099  2.58014889
#>  [13]  1.80346755 -1.23211530 -0.06505926 -1.13000150  0.82501143  0.59267509
#>  [19] -0.66034799  0.05987437 -1.05545080  0.98828970 -1.35048647 -1.12441979
#>  [25]  0.22991078 -1.02374175 -1.08769667 -1.34363433 -0.35955168 -1.25745660
#>  [31] -1.15764579 -0.19059760 -0.50379130 -0.46928925  1.00168370  1.27346352
#>  [37]  0.61625003 -0.45335837  0.52291020  0.35340073  0.22512154  1.64184381
#>  [43] -0.85181344 -1.80005350 -0.34388198 -2.13449444 -0.13448241 -0.26611653
#>  [49]  0.32118623  0.53150247 -0.01490665  1.03045447 -1.75315770 -1.18328773
#>  [55] -1.30542375 -0.52052928 -1.03493721  0.58213947  2.48363892 -0.96524857
#>  [61]  0.03529504  0.52385676  0.10614468  0.37226054  0.84176565  1.86080870
#>  [67] -1.01037372 -0.17763203  2.02436176  0.63441872  1.15848726  1.29053599
#>  [73] -0.14449512 -0.25171529  1.24817304  0.14307464 -1.38532984  0.84460890
#>  [79]  0.15277240  0.06432563  0.86126376 -0.27330126  0.56062937  1.29076849
#>  [85]  1.52421576 -0.62036300  0.50538569 -0.78202729 -1.12394125  0.46579816
#>  [91]  0.52163859  0.51009030  0.43994067 -0.11190319 -1.39818854 -0.41152121
#>  [97] -0.14293066  0.79490596 -0.57494044 -0.29389292
#>